CFBD Conferences Endpoint Overview

Get college football conference information Pulls all college football conferences and returns as data frame You can call this function simply with:

cfbd_conferences()

cfbd_conferences()

Source

https://api.collegefootballdata.com/conferences

Value

cfbd_conferences() - A data frame with 94 rows and 5 variables:

conference_id:

Referencing conference id.

name:

Conference name.

long_name:

Long name for Conference.

abbreviation:

Conference abbreviation.

classification:

Conference classification (fbs,fcs,ii,iii)

...

Examples

# \donttest{
   cfbd_conferences()

#>    conference_id                        name
#> 1              1                         ACC
#> 2              4                      Big 12
#> 3              5                     Big Ten
#> 4              8                         SEC
#> 5              9                      Pac-12
#> 6             12              Conference USA
#> 7             15                Mid-American
#> 8             17               Mountain West
#> 9             18            FBS Independents
#> 10            20                     Big Sky
#> 11            21                        MVFC
#> 12            22                         Ivy
#> 13            24                        MEAC
#> 14            25                         NEC
#> 15            26                         OVC
#> 16            27                     Patriot
#> 17            28                     Pioneer
#> 18            29                    Southern
#> 19            30                   Southland
#> 20            31                        SWAC
#> 21            37                    Sun Belt
#> 22            40                   Big South
#> 23            48                         CAA
#> 24           100          American Southwest
#> 25           102                        CCIW
#> 26           103                  Centennial
#> 27           104                        CIAA
#> 28           105                        ECFC
#> 29           106                    Empire 8
#> 30           107                       GLIAC
#> 31           108                 Great Lakes
#> 32           109             Great Northwest
#> 33           110                  Gulf South
#> 34           111                   Heartland
#> 35           112             Independent DII
#> 36           113            Independent DIII
#> 37           114             American Rivers
#> 38           115              Liberty League
#> 39           116                   Lone Star
#> 40           117                    Michigan
#> 41           118                 Mid America
#> 42           119                Mid Atlantic
#> 43           120                     Midwest
#> 44           121                   Minnesota
#> 45           122                      NESCAC
#> 46           123          Commonwealth Coast
#> 47           124                  New Jersey
#> 48           126                 North Coast
#> 49           127                Northeast 10
#> 50           128                        NACC
#> 51           129                Northern Sun
#> 52           130                   Northwest
#> 53           131                        Ohio
#> 54           132                Old Dominion
#> 55           133 Pennsylvania State Athletic
#> 56           134                 Presidents'
#> 57           135              Rocky Mountain
#> 58           136                        SIAC
#> 59           138                    So. Cal.
#> 60           139              South Atlantic
#> 61           142               Upper Midwest
#> 62           143                   USA South
#> 63           144               Mountain East
#> 64           145                   Wisconsin
#> 65           146              Great American
#> 66           147           Southern Athletic
#> 67           151           American Athletic
#> 68           160                       MSCAC
#> 69           165      Great Midwest Athletic
#> 70           166                      NEWMAC
#> 71           175                         AWC
#> 72           201                     Western
#> 73           202             Missouri Valley
#> 74           203              Rocky Mountain
#> 75           204                   Southwest
#> 76           205                     Pacific
#> 77           206                    Southern
#> 78           207                       Big 6
#> 79           208             Missouri Valley
#> 80           209              Mountain State
#> 81           210                       Big 7
#> 82           211                     Skyline
#> 83           212                         Ivy
#> 84           213                        AAWU
#> 85           214                       Big 8
#> 86           215            Western Athletic
#> 87           216                       Pac-8
#> 88           217                        PCAA
#> 89           218                   Southland
#> 90           219                        SWAC
#> 91           220                      Pac-10
#> 92           221                    Big West
#> 93           222                    Big East
#> 94           223                      Border
#>                                               long_name abbreviation
#> 1                             Atlantic Coast Conference          ACC
#> 2                                     Big 12 Conference          B12
#> 3                                    Big Ten Conference          B1G
#> 4                               Southeastern Conference          SEC
#> 5                                     Pac-12 Conference          PAC
#> 6                                        Conference USA         CUSA
#> 7                               Mid-American Conference          MAC
#> 8                              Mountain West Conference          MWC
#> 9                                      FBS Independents          Ind
#> 10                                   Big Sky Conference         <NA>
#> 11                  Missouri Valley Football Conference         <NA>
#> 12                                           Ivy League         <NA>
#> 13                      Mid-Eastern Athletic Conference         <NA>
#> 14                                 Northeast Conference         <NA>
#> 15                               Ohio Valley Conference         <NA>
#> 16                                       Patriot League         <NA>
#> 17                              Pioneer Football League         <NA>
#> 18                                  Southern Conference         <NA>
#> 19                                 Southland Conference         <NA>
#> 20                     Southwestern Athletic Conference         <NA>
#> 21                                  Sun Belt Conference          SBC
#> 22                                 Big South Conference         <NA>
#> 23                        Colonial Athletic Association         <NA>
#> 24                                   American Southwest         <NA>
#> 25                                                 CCIW         <NA>
#> 26                                           Centennial         <NA>
#> 27                                                 CIAA         <NA>
#> 28                                                 ECFC         <NA>
#> 29                                             Empire 8         <NA>
#> 30                                                GLIAC         <NA>
#> 31                                          Great Lakes         <NA>
#> 32                                      Great Northwest         <NA>
#> 33                                           Gulf South         <NA>
#> 34                                            Heartland         <NA>
#> 35                                      Independent DII         <NA>
#> 36                                     Independent DIII         <NA>
#> 37                           American Rivers Conference         <NA>
#> 38                                       Liberty League         <NA>
#> 39                                            Lone Star         <NA>
#> 40                                             Michigan         <NA>
#> 41                                          Mid America         <NA>
#> 42                           Middle Atlantic Conference         <NA>
#> 43                                              Midwest         <NA>
#> 44                                            Minnesota         <NA>
#> 45                                               NESCAC         <NA>
#> 46                        Commonwealth Coast Conference         <NA>
#> 47                                           New Jersey         <NA>
#> 48                                          North Coast         <NA>
#> 49                                         Northeast 10         <NA>
#> 50             Northern Athletics Collegiate Conference         <NA>
#> 51                                         Northern Sun         <NA>
#> 52                                            Northwest         <NA>
#> 53                                                 Ohio         <NA>
#> 54                                         Old Dominion         <NA>
#> 55               Pennsylvania State Athletic Conference         <NA>
#> 56                                          Presidents'         <NA>
#> 57                                       Rocky Mountain         <NA>
#> 58                                                 SIAC         <NA>
#> 59                                             So. Cal.         <NA>
#> 60                                       South Atlantic         <NA>
#> 61                                        Upper Midwest         <NA>
#> 62                                            USA South         <NA>
#> 63                             Mountain East Conference         <NA>
#> 64                                            Wisconsin         <NA>
#> 65                            Great American Conference         <NA>
#> 66                        Southern Athletic Association         <NA>
#> 67                         American Athletic Conference          AAC
#> 68   Massachusetts State Collegiate Athletic Conference         <NA>
#> 69                    Great Midwest Athletic Conference         <NA>
#> 70    New England Women's and Men's Athletic Conference         <NA>
#> 71                                   ASUN-WAC Challenge         <NA>
#> 72                                   Western Conference      Western
#> 73 Missouri Valley Intercollegiate Athletic Association        MVIAA
#> 74                            Rocky Mountain Conference          RMC
#> 75                                 Southwest Conference          SWC
#> 76                             Pacific Coast Conference          PCC
#> 77                                  Southern Conference     Southern
#> 78                                     Big 6 Conference        Big 6
#> 79                           Missouri Valley Conference          MVC
#> 80                   Mountain State Athletic Conference         MSAC
#> 81                                     Big 7 Conference        Big 7
#> 82                                   Skyline Conference      Skyline
#> 83                                       The Ivy League          Ivy
#> 84         Athletic Association of Western Universities         AAWU
#> 85                                     Big 8 Conference        Big 8
#> 86                          Western Athletic Conference          WAC
#> 87                                 Pacific 8 Conference        Pac-8
#> 88                   Pacific Coast Athletic Association         PCAA
#> 89                                 Southland Conference    Southland
#> 90                        Southwest Athletic Conference         SWAC
#> 91                                           Pacific 10       Pac-10
#> 92                                  Big West Conference           BW
#> 93                                  Big East Conference           BE
#> 94          Border Intercollegiate Athletic Association         BIAA
#>    classification
#> 1             fbs
#> 2             fbs
#> 3             fbs
#> 4             fbs
#> 5             fbs
#> 6             fbs
#> 7             fbs
#> 8             fbs
#> 9             fbs
#> 10            fcs
#> 11            fcs
#> 12            fcs
#> 13            fcs
#> 14            fcs
#> 15            fcs
#> 16            fcs
#> 17            fcs
#> 18            fcs
#> 19            fcs
#> 20            fcs
#> 21            fbs
#> 22            fcs
#> 23            fcs
#> 24            iii
#> 25            iii
#> 26            iii
#> 27             ii
#> 28            iii
#> 29            iii
#> 30             ii
#> 31             ii
#> 32             ii
#> 33             ii
#> 34            iii
#> 35             ii
#> 36            iii
#> 37            iii
#> 38            iii
#> 39             ii
#> 40            iii
#> 41             ii
#> 42            iii
#> 43            iii
#> 44            iii
#> 45            iii
#> 46            iii
#> 47            iii
#> 48            iii
#> 49             ii
#> 50            iii
#> 51             ii
#> 52            iii
#> 53            iii
#> 54            iii
#> 55             ii
#> 56            iii
#> 57             ii
#> 58             ii
#> 59            iii
#> 60             ii
#> 61            iii
#> 62            iii
#> 63             ii
#> 64            iii
#> 65             ii
#> 66            iii
#> 67            fbs
#> 68            iii
#> 69             ii
#> 70            iii
#> 71            fcs
#> 72           <NA>
#> 73           <NA>
#> 74           <NA>
#> 75           <NA>
#> 76           <NA>
#> 77           <NA>
#> 78           <NA>
#> 79           <NA>
#> 80           <NA>
#> 81           <NA>
#> 82           <NA>
#> 83           <NA>
#> 84           <NA>
#> 85           <NA>
#> 86           <NA>
#> 87           <NA>
#> 88           <NA>
#> 89           <NA>
#> 90           <NA>
#> 91            fbs
#> 92           <NA>
#> 93            fbs
#> 94           <NA>

# }