CFBD Ratings and Rankings Endpoints

Postseason polls are after Week 13

At least one of year or team must be specified for the function to run

At least one of year or team must be specified for the function to run

cfbd_rankings(year, week = NULL, season_type = "regular")

cfbd_ratings_sp(year = NULL, team = NULL)

cfbd_ratings_sp_conference(year = NULL, conference = NULL)

cfbd_ratings_srs(year = NULL, team = NULL, conference = NULL)

Arguments

year
(Integer optional): Year, 4 digit format (YYYY)

week
(Integer optional): Week, values from 1-15, 1-14 for seasons pre-playoff (i.e. 2013 or earlier)

season_type
(String default regular): Season type - regular or postseason

team
(String optional): D-I Team

conference
(String optional): Conference abbreviation - SRS information by conference
Conference abbreviations P5: ACC, B12, B1G, SEC, PAC
Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC

Source

https://api.collegefootballdata.com/rankings

https://api.collegefootballdata.com/ratings/sp

https://api.collegefootballdata.com/ratings/sp/conferences

https://api.collegefootballdata.com/ratings/srs

Value

A data frame with 9 variables:

season

integer.

season_type

character.

week

integer.

poll

character.

rank

integer.

school

character.

conference

character.

first_place_votes

integer.

points

integer.

A data frame with 26 variables:

year

integer.

team

character.

conference

character.

rating

double.

ranking

integer.

second_order_wins

logical.

sos

logical.

offense_ranking

integer.

offense_rating

double.

offense_success

logical.

offense_explosiveness

logical.

offense_rushing

logical.

offense_passing

logical.

offense_standard_downs

logical.

offense_passing_downs

logical.

offense_run_rate

logical.

offense_pace

logical.

defense_ranking

integer.

defense_rating

double.

defense_success

logical.

defense_explosiveness

logical.

defense_rushing

logical.

defense_passing

logical.

defense_standard_downs

logical.

defense_passing_downs

logical.

defense_havoc_total

logical.

defense_havoc_front_seven

logical.

defense_havoc_db

logical.

special_teams_rating

double.

A data frame with 25 variables:

year

integer.

conference

character.

rating

double.

second_order_wins

logical.

sos

logical.

offense_rating

double.

offense_success

logical.

offense_explosiveness

logical.

offense_rushing

logical.

offense_passing

logical.

offense_standard_downs

logical.

offense_passing_downs

logical.

offense_run_rate

logical.

offense_pace

logical.

defense_rating

double.

defense_success

logical.

defense_explosiveness

logical.

defense_rushing

logical.

defense_passing

logical.

defense_standard_downs

logical.

defense_passing_downs

logical.

defense_havoc_total

logical.

defense_havoc_front_seven

logical.

defense_havoc_db

logical.

special_teams_rating

double.

A data frame with 6 variables:

year

integer.

team

character.

conference

character.

division

logical.

rating

double.

ranking

integer.

Examples

if (FALSE) {
  cfbd_rankings(year = 2019, week = 12)

  cfbd_rankings(year = 2018, week = 14)

  cfbd_rankings(year = 2013, season_type = 'postseason')
}
if (FALSE) {
  cfbd_ratings_sp(year = 2019)

  cfbd_ratings_sp(team = 'Texas A&M')

  cfbd_ratings_sp(year= 2019, team = "Texas")
}
if (FALSE) {
  cfbd_ratings_sp_conference(year = 2019)

  cfbd_ratings_sp_conference(year = 2012, conference = 'SEC')

  cfbd_ratings_sp_conference(year = 2016, conference = 'ACC')
}
if (FALSE) {
  cfbd_ratings_srs(year = 2019, team = "Texas")

  cfbd_ratings_srs(year = 2018, conference = 'SEC')
}