CFBD Recruiting Endpoint

Gets CFB recruiting information for a single year with filters available for team, recruit type, state and position.

If only start_year is provided, function will get CFB recruiting information based on position groups during that year for all FBS teams.

Gets CFB team recruiting ranks with filters available for year and team. At least one of year or team must be specified for the function to run

cfbd_recruiting_player(
  year = NULL,
  team = NULL,
  recruit_type = "HighSchool",
  state = NULL,
  position = NULL
)

cfbd_recruiting_position(
  start_year = NULL,
  end_year = NULL,
  team = NULL,
  conference = NULL
)

cfbd_recruiting_team(year = NULL, team = NULL)

Arguments

year
(Integer optional): Recruiting Class Year, 4 digit format (YYYY). Note: 2000 is the minimum value

team
(String optional): Team - Select a valid team, D1 football

recruit_type
(String optional): default API return is 'HighSchool', other options include 'JUCO' or 'PrepSchool' - For position group information

state
(String optional): Two letter State abbreviation

position
(String optional): Position Group - options include:
* Offense: 'PRO', 'DUAL', 'RB', 'FB', 'TE', 'OT', 'OG', 'OC', 'WR'
* Defense: 'CB', 'S', 'OLB', 'ILB', 'WDE', 'SDE', 'DT'
* Special Teams: 'K', 'P'

start_year
(Integer optional): Start Year, 4 digit format (YYYY). Note: 2000 is the minimum value

end_year
(Integer optional): End Year, 4 digit format (YYYY). Note: 2020 is the maximum value currently

conference
(String optional): Conference abbreviation - Select a valid FBS conference
Conference abbreviations P5: ACC, B12, B1G, SEC, PAC
Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC

Source

https://api.collegefootballdata.com/recruiting/players

https://api.collegefootballdata.com/recruiting/groups

https://api.collegefootballdata.com/recruiting/teams

Value

A data frame with 14 variables:

recruit_type

character.

year

integer.

ranking

integer.

name

character.

school

character.

committed_to

character.

position

character.

height

double.

weight

integer.

stars

integer.

rating

double.

city

character.

state_province

character.

country

character.

A data frame with 7 variables:

team

character.

conference

character.

position_group

character.

avg_rating

double.

total_rating

double.

commits

integer.

avg_stars

double.

A data frame with 4 variables:

year

integer.

rank

integer.

team

character.

points

character.

Details

At least one of year or team must be specified for the function to run

If you would like CFB recruiting information for teams, please see the cfbfastR::cfbd_recruiting_team() function

If you would like to get cfb recruiting information based on position groups during a time period for all FBS teams, please see the cfbfastR::cfbd_recruiting_position() function.

If you would like CFB recruiting information for players, please see the cfbfastR::cfbd_recruiting_player() function

If you would like CFB recruiting information for teams, please see the cfbfastR::cfbd_recruiting_team() function

If you would like CFB recruiting information for players, please see the cfbfastR::cfbd_recruiting_player() function

If you would like to get CFB recruiting information based on position groups during a time period for all FBS teams, please see the cfbfastR::cfbd_recruiting_position() function.

Examples

if (FALSE) {
  cfbd_recruiting_player(2018, team = "Texas")

  cfbd_recruiting_player(2016, recruit_type = 'JUCO')

  cfbd_recruiting_player(2020, recruit_type = 'HighSchool', position = 'OT', state = 'FL')
}
if (FALSE) {
  cfbd_recruiting_position(2018, team="Texas")

  cfbd_recruiting_position(2016, 2020, team="Virginia")

  cfbd_recruiting_position(2015, 2020, conference = "SEC")
}
if (FALSE) {
  cfbd_recruiting_team(2018, team = "Texas")

  cfbd_recruiting_team(2016, team = "Virginia")

  cfbd_recruiting_team(2016, team = "Texas A&M")

  cfbd_recruiting_team(2011)
}