cfbfastR: Functions to Access College Football Play by Play Data

An R package for working with College Football Data. It is an R API wrapper around <https://collegefootballdata.com/> made available by the cfbfastR team. It provides users the capability to access a plethora of endpoints, and supplement that data with additional information (Expected Points Added/Win Probability added).

See also

Author

Maintainer: Saiem Gilani saiem.gilani@gmail.com

Authors:

Akshay Easwaran akeaswaran@me.com

Jared Lee 13jaredlee@gmail.com

Eric Hess eric.hess.13@gmail.com

Other contributors:

Michael Egle mjegle@iastate.edu [contributor]

Nate Manzo natem67@gmail.com [contributor]

Jason DeLoach jdeloach5@gmail.com [contributor]

Tej Seth tejseth@umich.edu [contributor]

Conor McQuiston mcqconor@umich.edu [contributor]

Tan Ho tan@tanho.ca [contributor]

Keegan Abdoo kabdoo@draft.ly [contributor]

Matt Spencer maatspencer@gmail.com [contributor]