CFBD Coaches Endpoint Overview

Coach information search A coach search function which provides coaching records and school history for a given coach

cfbd_coaches(first = "Nick", last = "Saban", team = "alabama")

cfbd_coaches(
  first = NULL,
  last = NULL,
  team = NULL,
  year = NULL,
  min_year = NULL,
  max_year = NULL
)

Arguments

first
(String optional): First name for the coach you are trying to look up

last
(String optional): Last name for the coach you are trying to look up

team
(String optional): Team - Select a valid team, D1 football

year
(Integer optional): Year, 4 digit format (YYYY).

min_year
(Integer optional): Minimum Year filter (inclusive), 4 digit format (YYYY).

max_year
(Integer optional): Maximum Year filter (inclusive), 4 digit format (YYYY)

Value

first_name:character.

First name of coach.

last_name:character.

Last name of coach.

hire_date:character.

Hire date of coach.

school:character.

School of coach.

year:integer.

Season of record.

games:integer.

Games as coach.

wins:integer.

Wins for the season.

losses:integer.

Losses for the season.

ties:integer.

Ties for the season.

preseason_rank:integer.

Preseason rank for the school of coach.

postseason_rank:integer.

Postseason rank for the school of coach.

srs:character.

Simple Rating System adjustment for team.

sp_overall:character.

Bill Connelly's SP+ overall for team.

sp_offense:character.

Bill Connelly's SP+ offense for team.

sp_defense:character.

Bill Connelly's SP+ defense for team.

Examples

# \donttest{
  cfbd_coaches(first = "Nick", last = "Saban", team = "alabama")

#>     first_name   last_name                hire_date  school year games wins
#> 1           M.     Griffin                     <NA> Alabama 1900     5    2
#> 2         G.H.      Harvey                     <NA> Alabama 1901     5    2
#> 3          Eli      Abbott                     <NA> Alabama 1902     8    4
#> 4         J.O.    Heyworth                     <NA> Alabama 1902     8    4
#> 5         W.B.      Blount                     <NA> Alabama 1903     7    3
#> 6         W.B.      Blount                     <NA> Alabama 1904    10    7
#> 7         Jack Leavenworth                     <NA> Alabama 1905    10    6
#> 8       J.W.H.     Pollard                     <NA> Alabama 1906     6    5
#> 9       J.W.H.     Pollard                     <NA> Alabama 1907     8    5
#> 10      J.W.H.     Pollard                     <NA> Alabama 1908     8    6
#> 11      J.W.H.     Pollard                     <NA> Alabama 1909     8    5
#> 12         Guy      Lowman                     <NA> Alabama 1910     8    4
#> 13        D.V.      Graves                     <NA> Alabama 1911     9    5
#> 14        D.V.      Graves                     <NA> Alabama 1912     9    5
#> 15        D.V.      Graves                     <NA> Alabama 1913     9    6
#> 16       Tubby        Long                     <NA> Alabama 1914     9    5
#> 17      Thomas      Kelley                     <NA> Alabama 1915     8    6
#> 18      Thomas      Kelley                     <NA> Alabama 1916     9    6
#> 19      Thomas      Kelley                     <NA> Alabama 1917     8    5
#> 20         Xen       Scott                     <NA> Alabama 1919     9    8
#> 21         Xen       Scott                     <NA> Alabama 1920    11   10
#> 22         Xen       Scott                     <NA> Alabama 1921    11    5
#> 23         Xen       Scott                     <NA> Alabama 1922    10    6
#> 24     Wallace        Wade                     <NA> Alabama 1923    10    7
#> 25     Wallace        Wade                     <NA> Alabama 1924     9    8
#> 26     Wallace        Wade                     <NA> Alabama 1925    10   10
#> 27     Wallace        Wade                     <NA> Alabama 1926    10    9
#> 28     Wallace        Wade                     <NA> Alabama 1927    10    5
#> 29     Wallace        Wade                     <NA> Alabama 1928     9    6
#> 30     Wallace        Wade                     <NA> Alabama 1929     9    6
#> 31     Wallace        Wade                     <NA> Alabama 1930    10   10
#> 32       Frank      Thomas                     <NA> Alabama 1931    10    9
#> 33       Frank      Thomas                     <NA> Alabama 1932    10    8
#> 34       Frank      Thomas                     <NA> Alabama 1933     9    7
#> 35       Frank      Thomas                     <NA> Alabama 1934    10   10
#> 36       Frank      Thomas                     <NA> Alabama 1935     9    6
#> 37       Frank      Thomas                     <NA> Alabama 1936     9    8
#> 38       Frank      Thomas                     <NA> Alabama 1937    10    9
#> 39       Frank      Thomas                     <NA> Alabama 1938     9    7
#> 40       Frank      Thomas                     <NA> Alabama 1939     9    5
#> 41       Frank      Thomas                     <NA> Alabama 1940     9    7
#> 42       Frank      Thomas                     <NA> Alabama 1941    11    9
#> 43       Frank      Thomas                     <NA> Alabama 1942    11    8
#> 44       Frank      Thomas                     <NA> Alabama 1944     9    5
#> 45       Frank      Thomas                     <NA> Alabama 1945    10   10
#> 46       Frank      Thomas                     <NA> Alabama 1946    11    7
#> 47         Red        Drew                     <NA> Alabama 1947    11    8
#> 48         Red        Drew                     <NA> Alabama 1948    11    6
#> 49         Red        Drew                     <NA> Alabama 1949    10    6
#> 50         Red        Drew                     <NA> Alabama 1950    11    9
#> 51         Red        Drew                     <NA> Alabama 1951    11    5
#> 52         Red        Drew                     <NA> Alabama 1952    12   10
#> 53         Red        Drew                     <NA> Alabama 1953    12    6
#> 54         Red        Drew                     <NA> Alabama 1954    11    4
#> 55        Ears   Whitworth                     <NA> Alabama 1955    10    0
#> 56        Ears   Whitworth                     <NA> Alabama 1956    10    2
#> 57        Ears   Whitworth                     <NA> Alabama 1957    10    2
#> 58        Bear      Bryant                     <NA> Alabama 1958    10    5
#> 59        Bear      Bryant                     <NA> Alabama 1959    11    7
#> 60        Bear      Bryant                     <NA> Alabama 1960    11    8
#> 61        Bear      Bryant                     <NA> Alabama 1961    11   11
#> 62        Bear      Bryant                     <NA> Alabama 1962    11   10
#> 63        Bear      Bryant                     <NA> Alabama 1963    11    9
#> 64        Bear      Bryant                     <NA> Alabama 1964    11   10
#> 65        Bear      Bryant                     <NA> Alabama 1965    11    9
#> 66        Bear      Bryant                     <NA> Alabama 1966    11   11
#> 67        Bear      Bryant                     <NA> Alabama 1967    11    8
#> 68        Bear      Bryant                     <NA> Alabama 1968    11    8
#> 69        Bear      Bryant                     <NA> Alabama 1969    11    6
#> 70        Bear      Bryant                     <NA> Alabama 1970    12    6
#> 71        Bear      Bryant                     <NA> Alabama 1971    12   11
#> 72        Bear      Bryant                     <NA> Alabama 1972    12   10
#> 73        Bear      Bryant                     <NA> Alabama 1973    12   11
#> 74        Bear      Bryant                     <NA> Alabama 1974    12   11
#> 75        Bear      Bryant                     <NA> Alabama 1975    12   11
#> 76        Bear      Bryant                     <NA> Alabama 1976    12    9
#> 77        Bear      Bryant                     <NA> Alabama 1977    12   11
#> 78        Bear      Bryant                     <NA> Alabama 1978    12   11
#> 79        Bear      Bryant                     <NA> Alabama 1979    12   12
#> 80        Bear      Bryant                     <NA> Alabama 1980    12   10
#> 81        Bear      Bryant                     <NA> Alabama 1981    12    9
#> 82        Bear      Bryant                     <NA> Alabama 1982    12    8
#> 83         Ray     Perkins                     <NA> Alabama 1983    12    8
#> 84         Ray     Perkins                     <NA> Alabama 1984    11    5
#> 85         Ray     Perkins                     <NA> Alabama 1985    12    9
#> 86         Ray     Perkins                     <NA> Alabama 1986    13   10
#> 87        Bill       Curry                     <NA> Alabama 1987    12    7
#> 88        Bill       Curry                     <NA> Alabama 1988    12    9
#> 89        Bill       Curry                     <NA> Alabama 1989    12   10
#> 90        Gene   Stallings                     <NA> Alabama 1990    12    7
#> 91        Gene   Stallings                     <NA> Alabama 1991    12   11
#> 92        Gene   Stallings                     <NA> Alabama 1992    13   13
#> 93        Gene   Stallings                     <NA> Alabama 1993    13    9
#> 94        Gene   Stallings                     <NA> Alabama 1994    13   12
#> 95        Gene   Stallings                     <NA> Alabama 1995    11    8
#> 96        Gene   Stallings                     <NA> Alabama 1996    13   10
#> 97        Mike      DuBose                     <NA> Alabama 1997    11    4
#> 98        Mike      DuBose                     <NA> Alabama 1998    12    7
#> 99        Mike      DuBose                     <NA> Alabama 1999    13   10
#> 100       Mike      DuBose                     <NA> Alabama 2000    11    3
#> 101     Dennis  Franchione                     <NA> Alabama 2001    12    7
#> 102     Dennis  Franchione                     <NA> Alabama 2002    13   10
#> 103       Mike       Shula                     <NA> Alabama 2003    13    4
#> 104       Mike       Shula                     <NA> Alabama 2004    12    6
#> 105       Mike       Shula                     <NA> Alabama 2005    12   10
#> 106       Mike       Shula                     <NA> Alabama 2006    13    6
#> 107       Nick       Saban 2007-01-03T00:00:00.000Z Alabama 2007    13    7
#> 108       Nick       Saban 2007-01-03T00:00:00.000Z Alabama 2008    14   12
#> 109       Nick       Saban 2007-01-03T00:00:00.000Z Alabama 2009    14   14
#> 110       Nick       Saban 2007-01-03T00:00:00.000Z Alabama 2010    13   10
#> 111       Nick       Saban 2007-01-03T00:00:00.000Z Alabama 2011    13   12
#> 112       Nick       Saban 2007-01-03T00:00:00.000Z Alabama 2012    14   13
#> 113       Nick       Saban 2007-01-03T00:00:00.000Z Alabama 2013    13   11
#> 114       Nick       Saban 2007-01-03T00:00:00.000Z Alabama 2014    14   12
#> 115       Nick       Saban 2007-01-03T00:00:00.000Z Alabama 2015    15   14
#> 116       Nick       Saban 2007-01-03T00:00:00.000Z Alabama 2016    15   14
#> 117       Nick       Saban 2007-01-03T00:00:00.000Z Alabama 2017    14   13
#> 118       Nick       Saban 2007-01-03T00:00:00.000Z Alabama 2018    15   14
#> 119       Nick       Saban 2007-01-03T00:00:00.000Z Alabama 2019    13   11
#> 120       Nick       Saban 2007-01-03T00:00:00.000Z Alabama 2020    13   13
#> 121       Nick       Saban 2007-01-03T00:00:00.000Z Alabama 2021     0    0
#>     losses ties preseason_rank postseason_rank   srs sp_overall sp_offense
#> 1        3    0              0               0  <NA>       <NA>       <NA>
#> 2        1    2              0               0  <NA>       <NA>       <NA>
#> 3        4    0              0               0 -16.7       <NA>       <NA>
#> 4        4    0              0               0 -16.7       <NA>       <NA>
#> 5        4    0              0               0 -20.4       <NA>       <NA>
#> 6        3    0              0               0 -12.7       <NA>       <NA>
#> 7        4    0              0               0  -3.9       <NA>       <NA>
#> 8        1    0              0               0  -5.9       <NA>       <NA>
#> 9        1    2              0               0  -7.1       <NA>       <NA>
#> 10       1    1              0               0 -11.0       <NA>       <NA>
#> 11       1    2              0               0  -3.2       <NA>       <NA>
#> 12       4    0              0               0 -13.0       <NA>       <NA>
#> 13       2    2              0               0  -3.5       <NA>       <NA>
#> 14       3    1              0               0 -16.0       <NA>       <NA>
#> 15       3    0              0               0  -6.0       <NA>       <NA>
#> 16       4    0              0               0  -7.8       <NA>       <NA>
#> 17       2    0              0               0   1.5       <NA>       <NA>
#> 18       3    0              0               0 -14.5       <NA>       <NA>
#> 19       2    1              0               0  -0.6       <NA>       <NA>
#> 20       1    0              0               0  12.6       <NA>       <NA>
#> 21       1    0              0               0   7.1       <NA>       <NA>
#> 22       4    2              0               0 -13.2       <NA>       <NA>
#> 23       3    1              0               0   4.1       <NA>       <NA>
#> 24       2    1             NA              NA   3.8       <NA>       <NA>
#> 25       1    0             NA              NA  22.3       <NA>       <NA>
#> 26       0    0             NA              NA  21.7       <NA>       <NA>
#> 27       0    1             NA              NA  21.4       <NA>       <NA>
#> 28       4    1             NA              NA   2.0       <NA>       <NA>
#> 29       3    0             NA              NA  13.7       <NA>       <NA>
#> 30       3    0             NA              NA   9.3       <NA>       <NA>
#> 31       0    0             NA              NA  27.0       <NA>       <NA>
#> 32       1    0             NA              NA  20.3       <NA>       <NA>
#> 33       2    0             NA              NA  15.8       <NA>       <NA>
#> 34       1    1             NA              NA  11.3       <NA>       <NA>
#> 35       0    0             NA              NA  30.2       <NA>       <NA>
#> 36       2    1             NA              NA  19.5       <NA>       <NA>
#> 37       0    1             NA               4  20.0       <NA>       <NA>
#> 38       1    0             NA               4  18.3       <NA>       <NA>
#> 39       1    1             NA              13  13.0       <NA>       <NA>
#> 40       3    1             NA              NA  15.3       <NA>       <NA>
#> 41       2    0             NA              NA   9.2       <NA>       <NA>
#> 42       2    0             NA              20  19.1       <NA>       <NA>
#> 43       3    0             NA              10  23.4       <NA>       <NA>
#> 44       2    2             NA              NA  12.4       <NA>       <NA>
#> 45       0    0             NA               2  36.4       <NA>       <NA>
#> 46       4    0             NA              NA  14.3       <NA>       <NA>
#> 47       3    0             NA               6  16.8       <NA>       <NA>
#> 48       4    1             NA              NA  12.0       <NA>       <NA>
#> 49       3    1             NA              NA  13.1       <NA>       <NA>
#> 50       2    0             NA              16  18.0       <NA>       <NA>
#> 51       6    0              9              NA   4.1       <NA>       <NA>
#> 52       2    0             NA               9  24.2       <NA>       <NA>
#> 53       3    3              5              13  13.4       <NA>       <NA>
#> 54       5    2             14              NA   8.9       <NA>       <NA>
#> 55      10    0              0               0  -6.4       <NA>       <NA>
#> 56       7    1              0               0   0.3       <NA>       <NA>
#> 57       7    1              0               0  -1.8       <NA>       <NA>
#> 58       4    1             NA              NA  10.3       <NA>       <NA>
#> 59       2    2             NA              10  11.0       <NA>       <NA>
#> 60       1    2             NA               9  15.8       <NA>       <NA>
#> 61       0    0              3               1  27.6       <NA>       <NA>
#> 62       1    0              3               5  29.0       <NA>       <NA>
#> 63       2    0              3               8  20.0       <NA>       <NA>
#> 64       1    0              6               1  21.3       <NA>       <NA>
#> 65       1    1              5               1  24.5       <NA>       <NA>
#> 66       0    0              1               3  28.5       <NA>       <NA>
#> 67       2    1              2               8  15.5       <NA>       <NA>
#> 68       3    0             10              17  13.8       <NA>       <NA>
#> 69       5    0             13              NA  12.5       <NA>       <NA>
#> 70       5    1             16              NA  17.4       19.7       39.2
#> 71       1    0             16               4  33.0         37       42.1
#> 72       2    0              7               7  31.1       31.8         43
#> 73       1    0              6               4  31.5       32.6       44.6
#> 74       1    0              4               5  25.8       29.5       35.7
#> 75       1    0              2               3  26.6       28.8       36.2
#> 76       3    0              6              11  21.4         26       35.7
#> 77       1    0              6               2  28.5       32.5       41.3
#> 78       1    0              1               1  26.3       33.9         42
#> 79       0    0              2               1  26.4       27.3       35.6
#> 80       2    0              2               6  26.8       28.7       36.2
#> 81       2    1              4               7  18.0       20.2         32
#> 82       4    0              3              NA  18.6       20.4       37.8
#> 83       4    0             13              15  18.6       22.1         40
#> 84       6    0              9               0   8.0        9.2         28
#> 85       2    1              0              13  19.2       25.2       36.8
#> 86       3    0              5               9  20.8       26.4       36.1
#> 87       5    0             NA              NA  13.1       18.6       32.3
#> 88       3    0             14              17  13.6       16.6       32.9
#> 89       2    0             16               9  17.5         23       39.4
#> 90       5    0             12              NA   8.2         13       27.1
#> 91       1    0             22               5  13.1       17.3       29.9
#> 92       0    0              9               1  20.7       27.2       35.4
#> 93       3    1              2              14  14.0       18.6         32
#> 94       1    0             12               5  12.8       19.9       31.1
#> 95       3    0             10              21   6.8          9       28.7
#> 96       3    0             15              11  11.6       17.1       31.2
#> 97       7    0             15               0   5.2        7.8       30.5
#> 98       5    0              0               0   3.2          9       30.9
#> 99       3    0             20               8  17.6       26.5       41.9
#> 100      8    0              3               0   2.9        7.9       29.5
#> 101      5    0             25              NA  10.4       11.8       32.9
#> 102      3    0             NA              11  17.7       23.6       37.6
#> 103      9    0              0               0   6.3       16.2       38.4
#> 104      6    0              0               0   7.8        9.2       27.8
#> 105      2    0              0               8  12.1       19.8       28.1
#> 106      7    0              0               0   7.0       11.4       28.7
#> 107      6    0             NA              NA   8.3       12.8       30.1
#> 108      2    0             24               6  17.9         23       34.2
#> 109      0    0              5               1  27.5       30.2         35
#> 110      3    0              1              10  24.1       30.1       43.9
#> 111      1    0              2               1  30.5       32.5       39.3
#> 112      1    0              2               1  32.2       36.5       45.7
#> 113      2    0              1               7  25.4       34.1       45.6
#> 114      2    0              2               4  26.5       34.3       44.5
#> 115      1    0              3               1  27.3         29       37.7
#> 116      1    0              1               2  33.6       31.7         40
#> 117      1    0              1               1  26.3       33.5       38.7
#> 118      1    0              1              NA  33.5       36.3       50.9
#> 119      2    0              2               8  28.8       32.4       47.7
#> 120      0    0              3               1  31.1       28.3       46.8
#> 121      0    0              1              NA  23.6       25.5       41.9
#>     sp_defense
#> 1         <NA>
#> 2         <NA>
#> 3         <NA>
#> 4         <NA>
#> 5         <NA>
#> 6         <NA>
#> 7         <NA>
#> 8         <NA>
#> 9         <NA>
#> 10        <NA>
#> 11        <NA>
#> 12        <NA>
#> 13        <NA>
#> 14        <NA>
#> 15        <NA>
#> 16        <NA>
#> 17        <NA>
#> 18        <NA>
#> 19        <NA>
#> 20        <NA>
#> 21        <NA>
#> 22        <NA>
#> 23        <NA>
#> 24        <NA>
#> 25        <NA>
#> 26        <NA>
#> 27        <NA>
#> 28        <NA>
#> 29        <NA>
#> 30        <NA>
#> 31        <NA>
#> 32        <NA>
#> 33        <NA>
#> 34        <NA>
#> 35        <NA>
#> 36        <NA>
#> 37        <NA>
#> 38        <NA>
#> 39        <NA>
#> 40        <NA>
#> 41        <NA>
#> 42        <NA>
#> 43        <NA>
#> 44        <NA>
#> 45        <NA>
#> 46        <NA>
#> 47        <NA>
#> 48        <NA>
#> 49        <NA>
#> 50        <NA>
#> 51        <NA>
#> 52        <NA>
#> 53        <NA>
#> 54        <NA>
#> 55        <NA>
#> 56        <NA>
#> 57        <NA>
#> 58        <NA>
#> 59        <NA>
#> 60        <NA>
#> 61        <NA>
#> 62        <NA>
#> 63        <NA>
#> 64        <NA>
#> 65        <NA>
#> 66        <NA>
#> 67        <NA>
#> 68        <NA>
#> 69        <NA>
#> 70        21.3
#> 71         6.5
#> 72        11.6
#> 73        13.1
#> 74         6.6
#> 75         7.2
#> 76        10.7
#> 77         9.3
#> 78         8.2
#> 79         8.6
#> 80         7.7
#> 81        12.1
#> 82        18.4
#> 83        19.6
#> 84        19.5
#> 85        12.3
#> 86        10.8
#> 87        15.3
#> 88        17.7
#> 89          18
#> 90        14.3
#> 91        13.9
#> 92         8.6
#> 93        14.4
#> 94          12
#> 95        20.3
#> 96        15.3
#> 97          24
#> 98        23.3
#> 99        17.4
#> 100       22.7
#> 101       22.7
#> 102       14.4
#> 103         24
#> 104       19.2
#> 105        8.3
#> 106       16.5
#> 107       17.2
#> 108       10.9
#> 109        6.5
#> 110       14.2
#> 111        6.4
#> 112       10.7
#> 113       12.3
#> 114       10.7
#> 115        9.1
#> 116        8.6
#> 117        5.6
#> 118       14.2
#> 119       14.6
#> 120       18.6
#> 121       16.7

# }