CFBD Coaches Endpoint Overview

Coach information search A coach search function which provides coaching records and school history for a given coach

cfbd_coaches(first = "Nick", last = "Saban", team = "alabama")  

cfbd_coaches(
  first = NULL,
  last = NULL,
  team = NULL,
  year = NULL,
  min_year = NULL,
  max_year = NULL,
  verbose = FALSE
)

Arguments

first
(String optional): First name for the coach you are trying to look up

last
(String optional): Last name for the coach you are trying to look up

team
(String optional): Team - Select a valid team, D1 football

year
(Integer optional): Year, 4 digit format (YYYY).

min_year
(Integer optional): Minimum Year filter (inclusive), 4 digit format (YYYY).

max_year
(Integer optional): Maximum Year filter (inclusive), 4 digit format (YYYY)

verbose
Logical parameter (TRUE/FALSE, default: FALSE) to return warnings and messages from function

Source

https://api.collegefootballdata.com/coaches

Value

cfbd_coaches() - A data frame with coach information with the following columns:

first_name:character.

First name of coach.

last_name:character.

Last name of coach.

school:character.

School of coach.

year:integer.

Season of record.

games:integer.

Games as coach.

wins:integer.

Wins for the season.

losses:integer.

Losses for the season.

ties:integer.

Ties for the season.

preseason_rank:integer.

Preseason rank for the school of coach.

postseason_rank:integer.

Postseason rank for the school of coach.

srs:character.

Simple Rating System adjustment for team.

sp_overall:character.

Bill Connelly's SP+ overall for team.

sp_offense:character.

Bill Connelly's SP+ offense for team.

sp_defense:character.

Bill Connelly's SP+ defense for team.

Examples

# \donttest{
  cfbd_coaches(first = "Nick", last = "Saban", team = "alabama")

#>     first_name   last_name  school year games wins losses ties preseason_rank
#> 1           M.     Griffin Alabama 1900     5    2      3    0              0
#> 2         G.H.      Harvey Alabama 1901     5    2      1    2              0
#> 3          Eli      Abbott Alabama 1902     8    4      4    0              0
#> 4         J.O.    Heyworth Alabama 1902     8    4      4    0              0
#> 5         W.B.      Blount Alabama 1903     7    3      4    0              0
#> 6         W.B.      Blount Alabama 1904    10    7      3    0              0
#> 7         Jack Leavenworth Alabama 1905    10    6      4    0              0
#> 8       J.W.H.     Pollard Alabama 1906     6    5      1    0              0
#> 9       J.W.H.     Pollard Alabama 1907     8    5      1    2              0
#> 10      J.W.H.     Pollard Alabama 1908     8    6      1    1              0
#> 11      J.W.H.     Pollard Alabama 1909     8    5      1    2              0
#> 12         Guy      Lowman Alabama 1910     8    4      4    0              0
#> 13        D.V.      Graves Alabama 1911     9    5      2    2              0
#> 14        D.V.      Graves Alabama 1912     9    5      3    1              0
#> 15        D.V.      Graves Alabama 1913     9    6      3    0              0
#> 16       Tubby        Long Alabama 1914     9    5      4    0              0
#> 17      Thomas      Kelley Alabama 1915     8    6      2    0              0
#> 18      Thomas      Kelley Alabama 1916     9    6      3    0              0
#> 19      Thomas      Kelley Alabama 1917     8    5      2    1              0
#> 20         Xen       Scott Alabama 1919     9    8      1    0              0
#> 21         Xen       Scott Alabama 1920    11   10      1    0              0
#> 22         Xen       Scott Alabama 1921    11    5      4    2              0
#> 23         Xen       Scott Alabama 1922    10    6      3    1              0
#> 24     Wallace        Wade Alabama 1923    10    7      2    1             NA
#> 25     Wallace        Wade Alabama 1924     9    8      1    0             NA
#> 26     Wallace        Wade Alabama 1925    10   10      0    0             NA
#> 27     Wallace        Wade Alabama 1926    10    9      0    1             NA
#> 28     Wallace        Wade Alabama 1927    10    5      4    1             NA
#> 29     Wallace        Wade Alabama 1928     9    6      3    0             NA
#> 30     Wallace        Wade Alabama 1929     9    6      3    0             NA
#> 31     Wallace        Wade Alabama 1930    10   10      0    0             NA
#> 32       Frank      Thomas Alabama 1931    10    9      1    0             NA
#> 33       Frank      Thomas Alabama 1932    10    8      2    0             NA
#> 34       Frank      Thomas Alabama 1933     9    7      1    1             NA
#> 35       Frank      Thomas Alabama 1934    10   10      0    0             NA
#> 36       Frank      Thomas Alabama 1935     9    6      2    1             NA
#> 37       Frank      Thomas Alabama 1936     9    8      0    1             NA
#> 38       Frank      Thomas Alabama 1937    10    9      1    0             NA
#> 39       Frank      Thomas Alabama 1938     9    7      1    1             NA
#> 40       Frank      Thomas Alabama 1939     9    5      3    1             NA
#> 41       Frank      Thomas Alabama 1940     9    7      2    0             NA
#> 42       Frank      Thomas Alabama 1941    11    9      2    0             NA
#> 43       Frank      Thomas Alabama 1942    11    8      3    0             NA
#> 44       Frank      Thomas Alabama 1944     9    5      2    2             NA
#> 45       Frank      Thomas Alabama 1945    10   10      0    0             NA
#> 46       Frank      Thomas Alabama 1946    11    7      4    0             NA
#> 47         Red        Drew Alabama 1947    11    8      3    0             NA
#> 48         Red        Drew Alabama 1948    11    6      4    1             NA
#> 49         Red        Drew Alabama 1949    10    6      3    1             NA
#> 50         Red        Drew Alabama 1950    11    9      2    0             NA
#> 51         Red        Drew Alabama 1951    11    5      6    0              9
#> 52         Red        Drew Alabama 1952    12   10      2    0             NA
#> 53         Red        Drew Alabama 1953    12    6      3    3              5
#> 54         Red        Drew Alabama 1954    11    4      5    2             14
#> 55        Ears   Whitworth Alabama 1955    10    0     10    0              0
#> 56        Ears   Whitworth Alabama 1956    10    2      7    1              0
#> 57        Ears   Whitworth Alabama 1957    10    2      7    1              0
#> 58        Bear      Bryant Alabama 1958    10    5      4    1             NA
#> 59        Bear      Bryant Alabama 1959    11    7      2    2             NA
#> 60        Bear      Bryant Alabama 1960    11    8      1    2             NA
#> 61        Bear      Bryant Alabama 1961    11   11      0    0              3
#> 62        Bear      Bryant Alabama 1962    11   10      1    0              3
#> 63        Bear      Bryant Alabama 1963    11    9      2    0              3
#> 64        Bear      Bryant Alabama 1964    11   10      1    0              6
#> 65        Bear      Bryant Alabama 1965    11    9      1    1              5
#> 66        Bear      Bryant Alabama 1966    11   11      0    0              1
#> 67        Bear      Bryant Alabama 1967    11    8      2    1              2
#> 68        Bear      Bryant Alabama 1968    11    8      3    0             10
#> 69        Bear      Bryant Alabama 1969    11    6      5    0             13
#> 70        Bear      Bryant Alabama 1970    12    6      5    1             16
#> 71        Bear      Bryant Alabama 1971    12   11      1    0             16
#> 72        Bear      Bryant Alabama 1972    12   10      2    0              7
#> 73        Bear      Bryant Alabama 1973    12   11      1    0              6
#> 74        Bear      Bryant Alabama 1974    12   11      1    0              4
#> 75        Bear      Bryant Alabama 1975    12   11      1    0              2
#> 76        Bear      Bryant Alabama 1976    12    9      3    0              6
#> 77        Bear      Bryant Alabama 1977    12   11      1    0              6
#> 78        Bear      Bryant Alabama 1978    12   11      1    0              1
#> 79        Bear      Bryant Alabama 1979    12   12      0    0              2
#> 80        Bear      Bryant Alabama 1980    12   10      2    0              2
#> 81        Bear      Bryant Alabama 1981    12    9      2    1              4
#> 82        Bear      Bryant Alabama 1982    12    8      4    0              3
#> 83         Ray     Perkins Alabama 1983    12    8      4    0             13
#> 84         Ray     Perkins Alabama 1984    11    5      6    0              9
#> 85         Ray     Perkins Alabama 1985    12    9      2    1              0
#> 86         Ray     Perkins Alabama 1986    13   10      3    0              5
#> 87        Bill       Curry Alabama 1987    12    7      5    0             NA
#> 88        Bill       Curry Alabama 1988    12    9      3    0             14
#> 89        Bill       Curry Alabama 1989    12   10      2    0             16
#> 90        Gene   Stallings Alabama 1990    12    7      5    0             12
#> 91        Gene   Stallings Alabama 1991    12   11      1    0             22
#> 92        Gene   Stallings Alabama 1992    13   13      0    0              9
#> 93        Gene   Stallings Alabama 1993    13    9      3    1              2
#> 94        Gene   Stallings Alabama 1994    13   12      1    0             12
#> 95        Gene   Stallings Alabama 1995    11    8      3    0             10
#> 96        Gene   Stallings Alabama 1996    13   10      3    0             15
#> 97        Mike      DuBose Alabama 1997    11    4      7    0             15
#> 98        Mike      DuBose Alabama 1998    12    7      5    0              0
#> 99        Mike      DuBose Alabama 1999    13   10      3    0             20
#> 100       Mike      DuBose Alabama 2000    11    3      8    0              3
#> 101     Dennis  Franchione Alabama 2001    12    7      5    0             25
#> 102     Dennis  Franchione Alabama 2002    13   10      3    0             NA
#> 103       Mike       Shula Alabama 2003    13    4      9    0              0
#> 104       Mike       Shula Alabama 2004    12    6      6    0              0
#> 105       Mike       Shula Alabama 2005    12   10      2    0              0
#> 106       Mike       Shula Alabama 2006    13    6      7    0              0
#> 107       Nick       Saban Alabama 2007    13    7      6    0             NA
#> 108       Nick       Saban Alabama 2008    14   12      2    0             24
#> 109       Nick       Saban Alabama 2009    14   14      0    0              5
#> 110       Nick       Saban Alabama 2010    13   10      3    0              1
#> 111       Nick       Saban Alabama 2011    13   12      1    0              2
#> 112       Nick       Saban Alabama 2012    14   13      1    0              2
#> 113       Nick       Saban Alabama 2013    13   11      2    0              1
#> 114       Nick       Saban Alabama 2014    14   12      2    0              2
#> 115       Nick       Saban Alabama 2015    15   14      1    0              3
#> 116       Nick       Saban Alabama 2016    15   14      1    0              1
#> 117       Nick       Saban Alabama 2017    14   13      1    0              1
#> 118       Nick       Saban Alabama 2018    15   14      1    0              1
#> 119       Nick       Saban Alabama 2019    13   11      2    0              2
#> 120       Nick       Saban Alabama 2020    13   13      0    0              3
#>     postseason_rank   srs sp_overall sp_offense sp_defense
#> 1                 0  <NA>       <NA>       <NA>       <NA>
#> 2                 0  <NA>       <NA>       <NA>       <NA>
#> 3                 0 -16.7       <NA>       <NA>       <NA>
#> 4                 0 -16.7       <NA>       <NA>       <NA>
#> 5                 0 -20.4       <NA>       <NA>       <NA>
#> 6                 0 -12.7       <NA>       <NA>       <NA>
#> 7                 0  -3.9       <NA>       <NA>       <NA>
#> 8                 0  -5.9       <NA>       <NA>       <NA>
#> 9                 0  -7.1       <NA>       <NA>       <NA>
#> 10                0 -11.0       <NA>       <NA>       <NA>
#> 11                0  -3.2       <NA>       <NA>       <NA>
#> 12                0 -13.0       <NA>       <NA>       <NA>
#> 13                0  -3.5       <NA>       <NA>       <NA>
#> 14                0 -16.0       <NA>       <NA>       <NA>
#> 15                0  -6.0       <NA>       <NA>       <NA>
#> 16                0  -7.8       <NA>       <NA>       <NA>
#> 17                0   1.5       <NA>       <NA>       <NA>
#> 18                0 -14.5       <NA>       <NA>       <NA>
#> 19                0  -0.6       <NA>       <NA>       <NA>
#> 20                0  12.6       <NA>       <NA>       <NA>
#> 21                0   7.1       <NA>       <NA>       <NA>
#> 22                0 -13.2       <NA>       <NA>       <NA>
#> 23                0   4.1       <NA>       <NA>       <NA>
#> 24               NA   3.8       <NA>       <NA>       <NA>
#> 25               NA  22.3       <NA>       <NA>       <NA>
#> 26               NA  21.7       <NA>       <NA>       <NA>
#> 27               NA  21.4       <NA>       <NA>       <NA>
#> 28               NA   2.0       <NA>       <NA>       <NA>
#> 29               NA  13.7       <NA>       <NA>       <NA>
#> 30               NA   9.3       <NA>       <NA>       <NA>
#> 31               NA  27.0       <NA>       <NA>       <NA>
#> 32               NA  20.3       <NA>       <NA>       <NA>
#> 33               NA  15.8       <NA>       <NA>       <NA>
#> 34               NA  11.3       <NA>       <NA>       <NA>
#> 35               NA  30.2       <NA>       <NA>       <NA>
#> 36               NA  19.5       <NA>       <NA>       <NA>
#> 37                4  20.0       <NA>       <NA>       <NA>
#> 38                4  18.3       <NA>       <NA>       <NA>
#> 39               13  13.0       <NA>       <NA>       <NA>
#> 40               NA  15.3       <NA>       <NA>       <NA>
#> 41               NA   9.2       <NA>       <NA>       <NA>
#> 42               20  19.1       <NA>       <NA>       <NA>
#> 43               10  23.4       <NA>       <NA>       <NA>
#> 44               NA  12.4       <NA>       <NA>       <NA>
#> 45                2  36.4       <NA>       <NA>       <NA>
#> 46               NA  14.3       <NA>       <NA>       <NA>
#> 47                6  16.8       <NA>       <NA>       <NA>
#> 48               NA  12.0       <NA>       <NA>       <NA>
#> 49               NA  13.1       <NA>       <NA>       <NA>
#> 50               16  18.0       <NA>       <NA>       <NA>
#> 51               NA   4.1       <NA>       <NA>       <NA>
#> 52                9  24.2       <NA>       <NA>       <NA>
#> 53               13  13.4       <NA>       <NA>       <NA>
#> 54               NA   8.9       <NA>       <NA>       <NA>
#> 55                0  -6.4       <NA>       <NA>       <NA>
#> 56                0   0.3       <NA>       <NA>       <NA>
#> 57                0  -1.8       <NA>       <NA>       <NA>
#> 58               NA  10.3       <NA>       <NA>       <NA>
#> 59               10  11.0       <NA>       <NA>       <NA>
#> 60                9  15.8       <NA>       <NA>       <NA>
#> 61                1  27.6       <NA>       <NA>       <NA>
#> 62                5  29.0       <NA>       <NA>       <NA>
#> 63                8  20.0       <NA>       <NA>       <NA>
#> 64                1  21.3       <NA>       <NA>       <NA>
#> 65                1  24.5       <NA>       <NA>       <NA>
#> 66                3  28.5       <NA>       <NA>       <NA>
#> 67                8  15.5       <NA>       <NA>       <NA>
#> 68               17  13.8       <NA>       <NA>       <NA>
#> 69               NA  12.5       <NA>       <NA>       <NA>
#> 70               NA  17.4       19.7       39.2       21.3
#> 71                4  33.0         37       42.1        6.5
#> 72                7  31.1       31.8         43       11.6
#> 73                4  31.5       32.6       44.6       13.1
#> 74                5  25.8       29.5       35.7        6.6
#> 75                3  26.6       28.8       36.2        7.2
#> 76               11  21.4         26       35.7       10.7
#> 77                2  28.5       32.5       41.3        9.3
#> 78                1  26.3       33.9         42        8.2
#> 79                1  26.4       27.3       35.6        8.6
#> 80                6  26.8       28.7       36.2        7.7
#> 81                7  18.0       20.2         32       12.1
#> 82               NA  18.6       20.4       37.8       18.4
#> 83               15  18.6       22.1         40       19.6
#> 84                0   8.0        9.2         28       19.5
#> 85               13  19.2       25.2       36.8       12.3
#> 86                9  20.8       26.4       36.1       10.8
#> 87               NA  13.1       18.6       32.3       15.3
#> 88               17  13.6       16.6       32.9       17.7
#> 89                9  17.5         23       39.4         18
#> 90               NA   8.2         13       27.1       14.3
#> 91                5  13.1       17.3       29.9       13.9
#> 92                1  20.7       27.2       35.4        8.6
#> 93               14  14.0       18.6         32       14.4
#> 94                5  12.8       19.9       31.1         12
#> 95               21   6.8          9       28.7       20.3
#> 96               11  11.6       17.1       31.2       15.3
#> 97                0   5.2        7.8       30.5         24
#> 98                0   3.2          9       30.9       23.3
#> 99                8  17.6       26.5       41.9       17.4
#> 100               0   2.9        7.9       29.5       22.7
#> 101              NA  10.4       11.8       32.9       22.7
#> 102              11  17.7       23.6       37.6       14.4
#> 103               0   6.3       16.2       38.4         24
#> 104               0   7.8        9.2       27.8       19.2
#> 105               8  12.1       19.8       28.1        8.3
#> 106               0   7.0       11.4       28.7       16.5
#> 107              NA   8.3       12.8       30.1       17.2
#> 108               6  17.9         23       34.2       10.9
#> 109               1  27.5       30.2         35        6.5
#> 110              10  24.1       30.1       43.9       14.2
#> 111               1  30.5       32.5       39.3        6.4
#> 112               1  32.2       36.5       45.7       10.7
#> 113               7  25.4       34.1       45.6       12.3
#> 114               4  26.5       34.3       44.5       10.7
#> 115               1  27.3         29       37.7        9.1
#> 116               2  33.6       31.7         40        8.6
#> 117               1  26.3       33.5       38.7        5.6
#> 118              NA  33.5       36.3       50.9       14.2
#> 119               8  28.8       32.4       47.7       14.6
#> 120               1  31.1       28.3       46.8       18.6

# }