Get results information from games.

Get results information from games.

cfbd_game_info(
  year,
  week = NULL,
  season_type = "regular",
  team = NULL,
  home_team = NULL,
  away_team = NULL,
  conference = NULL,
  game_id = NULL,
  quarter_scores = FALSE,
  verbose = FALSE
)

Arguments

year
(Integer required): Year, 4 digit format(YYYY)

week
(Integer optional): Week - values from 1-15, 1-14 for seasons pre-playoff (i.e. 2013 or earlier)

season_type
(String default regular): Select Season Type: regular, postseason, or both

team
(String optional): D-I Team

home_team
(String optional): Home D-I Team

away_team
(String optional): Away D-I Team

conference
(String optional): Conference abbreviation - Select a valid FBS conference
Conference abbreviations P5: ACC, B12, B1G, SEC, PAC
Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC

game_id
(Integer optional): Game ID filter for querying a single game

quarter_scores
(Logical default FALSE): This is a parameter to return the list columns that give the score at each quarter: home_line_scores and away_line_scores.
I have defaulted the parameter to false so that you will not have to go to the trouble of dropping it.

verbose
(Logical, default: FALSE) to return warnings and messages from function

Source

https://api.collegefootballdata.com/games

Value

cfbd_game_info() - A data frame with 22 variables:

game_id: integer.

Referencing game id.

season: integer.

Season of the game.

week: integer.

Game week.

season_type: character.

Season type of the game.

start_date: character.

Game date.

start_time_tbd: logical.

TRUE/FALSE flag for if the game's start time is to be determined.

neutral_site: logical.

TRUE/FALSE flag for the game taking place at a neutral site.

conference_game: logical.

TRUE/FALSE flag for this game qualifying as a conference game.

attendance: integer.

Reported attendance at the game.

venue_id: integer.

Referencing venue id.

venue: character.

Venue name.

home_id: integer.

Home team referencing id.

home_team: character.

Home team name.

home_conference: character.

Home team conference.

home_points: integer.

Home team points.

home_post_win_prob: character.

Home team post-game win probability.

away_id: integer.

Away team referencing id.

away_team: character.

Away team name.

away_conference: character.

Away team conference.

away_points: integer.

Away team points.

away_post_win_prob: character.

Away team post-game win probability.

excitement_index: character.

Game excitement index.

Examples

# \donttest{
  cfbd_game_info(2018, week = 1)

#>      game_id season week season_type               start_date start_time_tbd
#> 1  401013357   2018    1     regular 2018-08-25T21:30:00.000Z             NA
#> 2  401014972   2018    1     regular 2018-08-25T23:00:00.000Z             NA
#> 3  401022510   2018    1     regular 2018-08-25T23:30:00.000Z             NA
#> 4  401013437   2018    1     regular 2018-08-26T02:00:00.000Z             NA
#> 5  401021651   2018    1     regular 2018-08-30T23:00:00.000Z             NA
#> 6  401013328   2018    1     regular 2018-08-30T23:00:00.000Z             NA
#> 7  401019470   2018    1     regular 2018-08-30T23:00:00.000Z             NA
#> 8  401020671   2018    1     regular 2018-08-30T23:00:00.000Z             NA
#> 9  401021655   2018    1     regular 2018-08-31T00:00:00.000Z             NA
#> 10 401014979   2018    1     regular 2018-08-31T00:00:00.000Z             NA
#> 11 401012675   2018    1     regular 2018-08-31T00:00:00.000Z             NA
#> 12 401013086   2018    1     regular 2018-08-31T00:00:00.000Z             NA
#> 13 401013342   2018    1     regular 2018-08-31T00:00:00.000Z             NA
#> 14 401013489   2018    1     regular 2018-08-31T00:00:00.000Z             NA
#> 15 401012258   2018    1     regular 2018-08-31T00:30:00.000Z             NA
#> 16 401022511   2018    1     regular 2018-08-31T02:00:00.000Z             NA
#> 17 401020673   2018    1     regular 2018-08-31T22:00:00.000Z             NA
#> 18 401020672   2018    1     regular 2018-08-31T22:30:00.000Z             NA
#> 19 401012880   2018    1     regular 2018-08-31T23:00:00.000Z             NA
#> 20 401013089   2018    1     regular 2018-08-31T23:00:00.000Z             NA
#> 21 401022517   2018    1     regular 2018-09-01T01:00:00.000Z             NA
#> 22 401022512   2018    1     regular 2018-09-01T01:00:00.000Z             NA
#> 23 401013354   2018    1     regular 2018-09-01T01:00:00.000Z             NA
#> 24 401012680   2018    1     regular 2018-09-01T01:30:00.000Z             NA
#> 25 401012820   2018    1     regular 2018-09-01T16:00:00.000Z             NA
#> 26 401013072   2018    1     regular 2018-09-01T16:00:00.000Z             NA
#> 27 401012804   2018    1     regular 2018-09-01T16:00:00.000Z             NA
#> 28 401012256   2018    1     regular 2018-09-01T16:00:00.000Z             NA
#> 29 401014977   2018    1     regular 2018-09-01T16:00:00.000Z             NA
#> 30 401012253   2018    1     regular 2018-09-01T16:00:00.000Z             NA
#> 31 401012682   2018    1     regular 2018-09-01T16:00:00.000Z             NA
#> 32 401012872   2018    1     regular 2018-09-01T16:00:00.000Z             NA
#> 33 401019473   2018    1     regular 2018-09-01T16:00:00.000Z             NA
#> 34 401013091   2018    1     regular 2018-09-01T16:00:00.000Z             NA
#> 35 401013065   2018    1     regular 2018-09-01T16:00:00.000Z             NA
#> 36 401013088   2018    1     regular 2018-09-01T16:20:00.000Z             NA
#> 37 401013090   2018    1     regular 2018-09-01T16:30:00.000Z             NA
#> 38 401013087   2018    1     regular 2018-09-01T17:00:00.000Z             NA
#> 39 401020675   2018    1     regular 2018-09-01T18:00:00.000Z             NA
#> 40 401022514   2018    1     regular 2018-09-01T18:00:00.000Z             NA
#> 41 401013349   2018    1     regular 2018-09-01T19:30:00.000Z             NA
#> 42 401012683   2018    1     regular 2018-09-01T19:30:00.000Z             NA
#> 43 401012248   2018    1     regular 2018-09-01T19:30:00.000Z             NA
#> 44 401012251   2018    1     regular 2018-09-01T19:30:00.000Z             NA
#> 45 401012250   2018    1     regular 2018-09-01T19:30:00.000Z             NA
#> 46 401013092   2018    1     regular 2018-09-01T19:30:00.000Z             NA
#> 47 401012257   2018    1     regular 2018-09-01T19:30:00.000Z             NA
#> 48 401012862   2018    1     regular 2018-09-01T19:40:00.000Z             NA
#> 49 401012678   2018    1     regular 2018-09-01T20:00:00.000Z             NA
#> 50 401012255   2018    1     regular 2018-09-01T20:00:00.000Z             NA
#> 51 401012684   2018    1     regular 2018-09-01T20:00:00.000Z             NA
#> 52 401012247   2018    1     regular 2018-09-01T20:00:00.000Z             NA
#> 53 401021656   2018    1     regular 2018-09-01T22:00:00.000Z             NA
#> 54 401014975   2018    1     regular 2018-09-01T22:00:00.000Z             NA
#> 55 401013094   2018    1     regular 2018-09-01T22:00:00.000Z             NA
#> 56 401014973   2018    1     regular 2018-09-01T22:00:00.000Z             NA
#> 57 401013448   2018    1     regular 2018-09-01T22:00:00.000Z             NA
#> 58 401021653   2018    1     regular 2018-09-01T22:00:00.000Z             NA
#> 59 401019472   2018    1     regular 2018-09-01T22:00:00.000Z             NA
#> 60 401020674   2018    1     regular 2018-09-01T22:00:00.000Z             NA
#> 61 401014978   2018    1     regular 2018-09-01T23:00:00.000Z             NA
#> 62 401021654   2018    1     regular 2018-09-01T23:00:00.000Z             NA
#> 63 401014974   2018    1     regular 2018-09-01T23:00:00.000Z             NA
#> 64 401012851   2018    1     regular 2018-09-01T23:00:00.000Z             NA
#> 65 401012679   2018    1     regular 2018-09-01T23:00:00.000Z             NA
#> 66 401019476   2018    1     regular 2018-09-01T23:00:00.000Z             NA
#> 67 401021652   2018    1     regular 2018-09-01T23:00:00.000Z             NA
#> 68 401013057   2018    1     regular 2018-09-01T23:00:00.000Z             NA
#> 69 401013048   2018    1     regular 2018-09-01T23:00:00.000Z             NA
#> 70 401020676   2018    1     regular 2018-09-01T23:00:00.000Z             NA
#> 71 401019474   2018    1     regular 2018-09-01T23:00:00.000Z             NA
#> 72 401012888   2018    1     regular 2018-09-01T23:30:00.000Z             NA
#> 73 401012259   2018    1     regular 2018-09-01T23:30:00.000Z             NA
#> 74 401014976   2018    1     regular 2018-09-01T23:30:00.000Z             NA
#> 75 401012249   2018    1     regular 2018-09-01T23:30:00.000Z             NA
#> 76 401012254   2018    1     regular 2018-09-01T23:30:00.000Z             NA
#> 77 401014980   2018    1     regular 2018-09-01T23:30:00.000Z             NA
#> 78 401013026   2018    1     regular 2018-09-02T00:00:00.000Z             NA
#> 79 401012246   2018    1     regular 2018-09-02T00:00:00.000Z             NA
#> 80 401012681   2018    1     regular 2018-09-02T00:00:00.000Z             NA
#> 81 401022515   2018    1     regular 2018-09-02T00:00:00.000Z             NA
#> 82 401022516   2018    1     regular 2018-09-02T02:00:00.000Z             NA
#> 83 401012677   2018    1     regular 2018-09-02T02:30:00.000Z             NA
#> 84 401012676   2018    1     regular 2018-09-02T02:45:00.000Z             NA
#> 85 401019475   2018    1     regular 2018-09-02T03:00:00.000Z             NA
#> 86 401079833   2018    1     regular 2018-09-02T19:30:00.000Z             NA
#> 87 401013095   2018    1     regular 2018-09-02T23:30:00.000Z             NA
#> 88 401013096   2018    1     regular 2018-09-04T00:00:00.000Z             NA
#>    neutral_site conference_game attendance venue_id
#> 1         FALSE           FALSE       8684     3985
#> 2         FALSE           FALSE      20050     3895
#> 3         FALSE            TRUE      31007     5388
#> 4         FALSE           FALSE      20322     3601
#> 5         FALSE           FALSE          0     4246
#> 6         FALSE           FALSE          0     3953
#> 7         FALSE            TRUE          0     3892
#> 8         FALSE           FALSE          0     3919
#> 9         FALSE           FALSE          0     3817
#> 10        FALSE           FALSE          0     3803
#> 11        FALSE           FALSE          0      587
#> 12        FALSE           FALSE          0     4729
#> 13        FALSE            TRUE          0     3907
#> 14        FALSE           FALSE          0     3646
#> 15        FALSE           FALSE          0     3795
#> 16        FALSE           FALSE          0     3935
#> 17        FALSE           FALSE          0     3980
#> 18        FALSE           FALSE          0     3912
#> 19        FALSE           FALSE          0     3936
#> 20        FALSE           FALSE          0     3982
#> 21        FALSE           FALSE          0     3816
#> 22        FALSE           FALSE          0     3940
#> 23        FALSE           FALSE          0      347
#> 24         TRUE           FALSE          0     3937
#> 25        FALSE           FALSE          0     3754
#> 26        FALSE           FALSE          0     3616
#> 27        FALSE           FALSE          0     3832
#> 28        FALSE           FALSE          0     3994
#> 29        FALSE           FALSE          0     3895
#> 30         TRUE           FALSE          0     3891
#> 31        FALSE           FALSE          0     3861
#> 32         TRUE           FALSE          0     3719
#> 33        FALSE           FALSE          0     3806
#> 34        FALSE           FALSE          0     3670
#> 35        FALSE           FALSE          0     3835
#> 36        FALSE           FALSE          0     3836
#> 37        FALSE           FALSE          0      721
#> 38        FALSE           FALSE          0     3615
#> 39        FALSE           FALSE          0     3876
#> 40        FALSE           FALSE          0     3713
#> 41        FALSE           FALSE     105232     3632
#> 42        FALSE           FALSE          0     3984
#> 43         TRUE           FALSE          0     5348
#> 44        FALSE           FALSE          0     3683
#> 45        FALSE           FALSE          0     3917
#> 46        FALSE           FALSE          0     3752
#> 47         TRUE           FALSE          0     3628
#> 48        FALSE           FALSE          0     3793
#> 49        FALSE           FALSE          0      477
#> 50        FALSE           FALSE          0     3838
#> 51        FALSE           FALSE          0     3831
#> 52        FALSE           FALSE          0     3887
#> 53        FALSE           FALSE          0     3975
#> 54        FALSE           FALSE          0     3996
#> 55        FALSE           FALSE          0     3923
#> 56        FALSE           FALSE          0     4418
#> 57        FALSE           FALSE          0     3992
#> 58        FALSE           FALSE          0     3608
#> 59        FALSE           FALSE          0     3886
#> 60        FALSE           FALSE          0     3965
#> 61        FALSE           FALSE          0     3815
#> 62        FALSE           FALSE          0     3666
#> 63        FALSE           FALSE          0     3797
#> 64        FALSE           FALSE          0      218
#> 65        FALSE           FALSE          0     1056
#> 66        FALSE           FALSE          0     3674
#> 67        FALSE           FALSE          0     3625
#> 68        FALSE           FALSE          0     3636
#> 69        FALSE           FALSE          0     3833
#> 70        FALSE           FALSE          0     3739
#> 71        FALSE           FALSE          0     3805
#> 72        FALSE           FALSE          0     3855
#> 73        FALSE           FALSE          0     3973
#> 74        FALSE           FALSE          0     3825
#> 75        FALSE           FALSE          0     3634
#> 76        FALSE           FALSE          0     3693
#> 77        FALSE           FALSE          0     3946
#> 78        FALSE           FALSE          0     4727
#> 79         TRUE           FALSE      57280     4013
#> 80        FALSE           FALSE          0     3626
#> 81        FALSE           FALSE          0     3971
#> 82        FALSE           FALSE          0     3660
#> 83        FALSE           FALSE          0     3947
#> 84        FALSE           FALSE          0     3619
#> 85        FALSE           FALSE          0     3610
#> 86        FALSE           FALSE          0     3699
#> 87         TRUE           FALSE          0     3687
#> 88        FALSE            TRUE          0     3697
#>                                                       venue home_id
#> 1                             Warren McGuirk Alumni Stadium     113
#> 2                                              Rice Stadium     242
#> 3                                            Canvas Stadium      36
#> 4                                    Aggie Memorial Stadium     166
#> 5                                     Georgia State Stadium    2247
#> 6                                          TCF Bank Stadium     135
#> 7                                          Rentschler Field      41
#> 8                                         Scheumann Stadium    2050
#> 9                                            Malone Stadium    2433
#> 10                                             Legion Field       5
#> 11                                      Rice-Eccles Stadium     254
#> 12                           Benson Field at Yulman Stadium    2655
#> 13                                         Ross-Ade Stadium    2509
#> 14                                    Boone Pickens Stadium     197
#> 15                                               Kyle Field     245
#> 16                                          Spartan Stadium      23
#> 17                                            Waldo Stadium    2711
#> 18                                        Rynearson Stadium    2199
#> 19                                          Spartan Stadium     127
#> 20                                     Wallace Wade Stadium     150
#> 21                                           Mackay Stadium    2440
#> 22                                         Stanford Stadium      24
#> 23                                     Camp Randall Stadium     275
#> 24                      Sports Authority Field at Mile High      36
#> 25                             High Point Solutions Stadium     164
#> 26                                   Amon G. Carter Stadium    2628
#> 27                                         Memorial Stadium     356
#> 28                                   Williams-Brice Stadium    2579
#> 29                                             Rice Stadium     242
#> 30                                              NRG Stadium    2641
#> 31                                             Ohio Stadium     194
#> 32                                              FedEx Field     120
#> 33                                  Lincoln Financial Field     218
#> 34                                    Carter-Finley Stadium     152
#> 35                 Gaylord Family Oklahoma Memorial Stadium     201
#> 36                                         Memorial Stadium     228
#> 37               Bobby Dodd Stadium at Historic Grant Field      59
#> 38                                           Alumni Stadium     103
#> 39                                            Peden Stadium     195
#> 40                                           Falcon Stadium    2005
#> 41                                           Beaver Stadium     213
#> 42                                     War Memorial Stadium    2751
#> 43                                    Mercedes-Benz Stadium       2
#> 44                                     Commonwealth Stadium      96
#> 45                                          Sanford Stadium      61
#> 46                                              Heinz Field     221
#> 47                                  Bank of America Stadium    2633
#> 48                                          Kinnick Stadium    2294
#> 49                            Los Angeles Memorial Coliseum      30
#> 50                                             Faurot Field     142
#> 51                              California Memorial Stadium      25
#> 52 Donald W. Reynolds Razorback Stadium Frank Broyles Field       8
#> 53                                Veterans Memorial Stadium    2653
#> 54                                            Yager Stadium     193
#> 55                                            Scott Stadium     258
#> 56                                 Jerry Richardson Stadium    2429
#> 57                                         Williams Stadium    2335
#> 58                                          Paulson Stadium     290
#> 59                                    Raymond James Stadium      58
#> 60                                               UB Stadium    2084
#> 61                                    M. M. Roberts Stadium    2572
#> 62                                              Cajun Field     309
#> 63                                     Ladd Peebles Stadium       6
#> 64                                              FIU Stadium    2229
#> 65                                                Rose Bowl      26
#> 66                    Skelly Field at H. A. Chapman Stadium     202
#> 67                                  Centennial Bank Stadium    2032
#> 68                      Bill Snyder Family Football Stadium    2306
#> 69                                         Memorial Stadium    2305
#> 70                                               Glass Bowl    2649
#> 71                            Liberty Bowl Memorial Stadium     235
#> 72                                       Notre Dame Stadium      87
#> 73                                       Vanderbilt Stadium     238
#> 74                                           Apogee Stadium     249
#> 75                                 Ben Hill Griffin Stadium      57
#> 76                                       Davis Wade Stadium     344
#> 77                                         Sun Bowl Stadium    2638
#> 78                                           McLane Stadium     239
#> 79                                    Camping World Stadium     333
#> 80                                           Autzen Stadium    2483
#> 81                                       University Stadium     167
#> 82                                          Bulldog Stadium     278
#> 83                                        Sun Devil Stadium       9
#> 84                                          Arizona Stadium      12
#> 85                                            Aloha Stadium      62
#> 86                                    Dowdy-Ficklen Stadium     151
#> 87                                             AT&T Stadium      99
#> 88              Bobby Bowden Field at Doak Campbell Stadium      52
#>                home_team   home_conference home_points   home_post_win_prob
#> 1                  UMass  FBS Independents          63   0.9984542422162311
#> 2                   Rice    Conference USA          31  0.41611197976904707
#> 3         Colorado State     Mountain West          34   0.3317397678892768
#> 4       New Mexico State  FBS Independents           7  0.00427210984632397
#> 5          Georgia State          Sun Belt          24    0.869590187971506
#> 6              Minnesota           Big Ten          48   0.9999638364577065
#> 7            Connecticut American Athletic          17 0.004203463433869054
#> 8             Ball State      Mid-American          42   0.9999720937904223
#> 9       Louisiana Monroe          Sun Belt          34   0.5720859305615573
#> 10                   UAB    Conference USA          52    0.999997345694123
#> 11                  Utah            Pac-12          41   0.9997998688829266
#> 12                Tulane American Athletic          17   0.2196613212325033
#> 13                Purdue           Big Ten          27  0.15206580002046932
#> 14        Oklahoma State            Big 12          58   0.9992698560515085
#> 15             Texas A&M               SEC          59   0.9999928023514462
#> 16        San José State     Mountain West          38   0.2255818432658651
#> 17      Western Michigan      Mid-American          42  0.10069975562117696
#> 18      Eastern Michigan      Mid-American          51   0.9995510164694754
#> 19        Michigan State           Big Ten          38    0.906183590742433
#> 20                  Duke               ACC          34   0.9973846191967987
#> 21                Nevada     Mountain West          72   0.9999946772201084
#> 22              Stanford            Pac-12          31   0.9937677403365777
#> 23             Wisconsin           Big Ten          34   0.9999658482595446
#> 24        Colorado State     Mountain West          13 0.003935607921436278
#> 25               Rutgers           Big Ten          35   0.9992581309620951
#> 26                   TCU            Big 12          55   0.9999921756502518
#> 27              Illinois           Big Ten          31   0.8340153107033903
#> 28        South Carolina               SEC          49   0.9999813927842912
#> 29                  Rice    Conference USA          27  0.05005196033053275
#> 30            Texas Tech            Big 12          27 0.007730744808323005
#> 31            Ohio State           Big Ten          77   0.9999435276188426
#> 32              Maryland           Big Ten          34   0.9131500976054453
#> 33                Temple American Athletic          17  0.08018018010651985
#> 34              NC State               ACC          24    0.839090960760694
#> 35              Oklahoma            Big 12          63   0.9999928028169979
#> 36               Clemson               ACC          48   0.9999840784279569
#> 37          Georgia Tech               ACC          41   0.9999959877845358
#> 38        Boston College               ACC          55   0.9998255691998653
#> 39                  Ohio      Mid-American          38   0.4527258093446935
#> 40             Air Force     Mountain West          38   0.9999947029448708
#> 41            Penn State           Big Ten          45    0.980509601522245
#> 42               Wyoming     Mountain West          19 0.005358724710402157
#> 43                Auburn               SEC          21   0.6081821946558714
#> 44              Kentucky               SEC          35   0.9948967001914114
#> 45               Georgia               SEC          45    0.999992203255419
#> 46            Pittsburgh               ACC          33   0.9997365885384938
#> 47             Tennessee               SEC          14 0.004273317866598812
#> 48                  Iowa           Big Ten          33   0.9987837404630044
#> 49                   USC            Pac-12          43   0.9967409790322357
#> 50              Missouri               SEC          51    0.995968150359603
#> 51            California            Pac-12          24   0.8968200177849878
#> 52              Arkansas               SEC          55   0.9967431808920783
#> 53                  Troy          Sun Belt          20  0.05923029255824867
#> 54            Miami (OH)      Mid-American          28   0.2914353097621623
#> 55              Virginia               ACC          42     0.99897198899527
#> 56             Charlotte    Conference USA          34    0.999984164464906
#> 57               Liberty  FBS Independents          52   0.9998402730839259
#> 58      Georgia Southern          Sun Belt          37   0.9987286659646343
#> 59         South Florida American Athletic          34   0.9999637417382397
#> 60               Buffalo      Mid-American          48   0.9999142322720109
#> 61  Southern Mississippi    Conference USA          55   0.9994952216812634
#> 62             Louisiana          Sun Belt          49   0.9997927217865997
#> 63         South Alabama          Sun Belt          26  0.17346633829827812
#> 64 Florida International    Conference USA          28  0.08816900970878677
#> 65                  UCLA            Pac-12          17   0.2196041237734069
#> 66                 Tulsa American Athletic          38   0.9977510896828313
#> 67        Arkansas State          Sun Belt          48    0.998884580500493
#> 68          Kansas State            Big 12          27   0.5711801005392997
#> 69                Kansas            Big 12          23  0.15578688342620697
#> 70                Toledo      Mid-American          66   0.9999873144404265
#> 71               Memphis American Athletic          66   0.9999958345824962
#> 72            Notre Dame  FBS Independents          24     0.92783043145901
#> 73            Vanderbilt               SEC          35   0.9996203330946147
#> 74           North Texas    Conference USA          46   0.9987373302064423
#> 75               Florida               SEC          53   0.9998257838347568
#> 76     Mississippi State               SEC          63   0.9999955955222241
#> 77                  UTEP    Conference USA          10 0.006031035556264542
#> 78                Baylor            Big 12          55   0.9998295720824828
#> 79               Alabama               SEC          51   0.9999255686466408
#> 80                Oregon            Pac-12          58   0.9998393673219237
#> 81            New Mexico     Mountain West          62   0.9991581186438564
#> 82          Fresno State     Mountain West          79   0.9995402711314478
#> 83         Arizona State            Pac-12          49   0.9999920070738963
#> 84               Arizona            Pac-12          23    0.252378224443892
#> 85               Hawai'i     Mountain West          59   0.9796414129712154
#> 86         East Carolina American Athletic          23   0.4801521413827188
#> 87                   LSU               SEC          33   0.9559552996558128
#> 88         Florida State               ACC           3  0.31246770456927847
#>    away_id                away_team   away_conference away_points
#> 1     2184                 Duquesne              <NA>          15
#> 2     2504             Prairie View              <NA>          28
#> 3       62                  Hawai'i     Mountain West          43
#> 4     2751                  Wyoming     Mountain West          29
#> 5      338           Kennesaw State              <NA>          20
#> 6      166         New Mexico State  FBS Independents          10
#> 7     2116                      UCF American Athletic          56
#> 8     2115      Central Connecticut              <NA>           6
#> 9     2545   Southeastern Louisiana              <NA>          31
#> 10    2542           Savannah State              <NA>           0
#> 11    2692              Weber State              <NA>          10
#> 12     154              Wake Forest               ACC          23
#> 13      77             Northwestern           Big Ten          31
#> 14    2623           Missouri State              <NA>          17
#> 15    2466       Northwestern State              <NA>           7
#> 16     302                 UC Davis              <NA>          44
#> 17     183                 Syracuse               ACC          55
#> 18    2405                 Monmouth              <NA>          17
#> 19     328               Utah State     Mountain West          31
#> 20     349                     Army  FBS Independents          14
#> 21    2502           Portland State              <NA>          19
#> 22      21          San Diego State     Mountain West          10
#> 23      98         Western Kentucky    Conference USA           3
#> 24      38                 Colorado            Pac-12          45
#> 25     326              Texas State          Sun Belt           7
#> 26    2582                 Southern              <NA>           7
#> 27    2309               Kent State      Mid-American          24
#> 28     324         Coastal Carolina          Sun Belt          15
#> 29     248                  Houston American Athletic          45
#> 30     145                 Ole Miss               SEC          47
#> 31     204             Oregon State            Pac-12          31
#> 32     251                    Texas            Big 12          29
#> 33     222                Villanova              <NA>          19
#> 34     256            James Madison              <NA>          13
#> 35    2226         Florida Atlantic    Conference USA          14
#> 36     231                   Furman              <NA>           7
#> 37    2016             Alcorn State              <NA>           0
#> 38     113                    UMass  FBS Independents          21
#> 39      47                   Howard              <NA>          32
#> 40    2619              Stony Brook              <NA>           0
#> 41    2026        Appalachian State          Sun Belt          38
#> 42     265         Washington State            Pac-12          41
#> 43     264               Washington            Pac-12          16
#> 44    2117         Central Michigan      Mid-American          20
#> 45    2046              Austin Peay              <NA>           0
#> 46     399                   Albany              <NA>           7
#> 47     277            West Virginia            Big 12          40
#> 48    2459        Northern Illinois      Mid-American           7
#> 49    2439                     UNLV     Mountain West          21
#> 50    2630                UT Martin              <NA>          14
#> 51     153           North Carolina               ACC          17
#> 52    2197         Eastern Illinois              <NA>          20
#> 53      68              Boise State     Mountain West          56
#> 54     276                 Marshall    Conference USA          35
#> 55     257                 Richmond              <NA>          13
#> 56    2230                  Fordham              <NA>          10
#> 57     295             Old Dominion    Conference USA          10
#> 58    2569     South Carolina State              <NA>           6
#> 59    2210                     Elon              <NA>          14
#> 60    2169           Delaware State              <NA>          10
#> 61    2296            Jackson State              <NA>           7
#> 62    2755                Grambling              <NA>          17
#> 63    2348           Louisiana Tech    Conference USA          30
#> 64      84                  Indiana           Big Ten          38
#> 65    2132               Cincinnati American Athletic          26
#> 66    2110         Central Arkansas              <NA>          27
#> 67    2546 Southeast Missouri State              <NA>          21
#> 68     233             South Dakota              <NA>          24
#> 69    2447                 Nicholls              <NA>          26
#> 70    2678                      VMI              <NA>           3
#> 71    2382                   Mercer              <NA>          14
#> 72     130                 Michigan           Big Ten          17
#> 73    2393         Middle Tennessee    Conference USA           7
#> 74    2567                      SMU American Athletic          23
#> 75    2127      Charleston Southern              <NA>           6
#> 76    2617        Stephen F. Austin              <NA>           6
#> 77    2464         Northern Arizona              <NA>          30
#> 78    2000        Abilene Christian              <NA>          27
#> 79      97               Louisville               ACC          14
#> 80     189            Bowling Green      Mid-American          24
#> 81    2916           Incarnate Word              <NA>          30
#> 82      70                    Idaho              <NA>          13
#> 83    2636           UT San Antonio    Conference USA           7
#> 84     252                      BYU  FBS Independents          28
#> 85    2426                     Navy American Athletic          41
#> 86    2448       North Carolina A&T              <NA>          28
#> 87    2390                    Miami               ACC          17
#> 88     259            Virginia Tech               ACC          24
#>          away_post_win_prob   excitement_index
#> 1      0.001545757783768864       0.2139101303
#> 2        0.5838880202309529       5.6012773743
#> 3        0.6682602321107232       4.5649361931
#> 4         0.995727890153676       2.7057118396
#> 5         0.130409812028494       5.8512641491
#> 6    0.00003616354229352137   3.06144440259999
#> 7         0.995796536566131  0.830061928399999
#> 8   0.000027906209577710506  0.760100617799999
#> 9        0.4279140694384427       8.4160129207
#> 10 0.0000026543058769812333       0.1406151783
#> 11    0.0002001311170733988        0.922426719
#> 12       0.7803386787674967       9.7884823983
#> 13       0.8479341999795307       6.6741715963
#> 14    0.0007301439484914685 0.0663753880999998
#> 15  0.000007197648553836089 0.0327134787999997
#> 16       0.7744181567341348       5.5811874724
#> 17        0.899300244378823       2.7679697627
#> 18    0.0004489835305245826  0.565053273800001
#> 19      0.09381640925756696       4.1366972475
#> 20    0.0026153808032013393   1.69744162409999
#> 21  0.000005322779891625551   1.08662059399999
#> 22     0.006232259663422335       2.8895508744
#> 23   0.00003415174045540148       0.2437023338
#> 24       0.9960643920785637         3.41109204
#> 25     0.000741869037904852   1.11697903199999
#> 26  0.000007824349748219994 0.0225499731999999
#> 27       0.1659846892966097   6.71988011069999
#> 28  0.000018607215708787983  0.182912192500001
#> 29       0.9499480396694673       3.9567502631
#> 30        0.992269255191677       3.6518173425
#> 31   0.00005647238115735753       0.1409114981
#> 32       0.0868499023945547   9.13837292739999
#> 33       0.9198198198934802   6.43306846939999
#> 34        0.160909039239306       3.8739046233
#> 35  0.000007197183002127794  0.451379604799999
#> 36    0.0000159215720431094 0.0250883832000001
#> 37   0.00000401221546419972 0.0701338293999992
#> 38   0.00017443080013468393       0.6622090949
#> 39       0.5472741906553065       3.8302068262
#> 40  0.000005297055129194206       0.8860730608
#> 41     0.019490398477755022   6.47689498139999
#> 42       0.9946412752895979       7.1429531895
#> 43       0.3918178053441286   7.48070415839999
#> 44    0.0051032998085885906   5.87733304869999
#> 45  0.000007796744580956805 0.0234045444000003
#> 46    0.0002634114615062133       0.2208688393
#> 47       0.9957266821334012   2.82384215719999
#> 48    0.0012162595369955564       3.0460114328
#> 49    0.0032590209677643145       2.4032706517
#> 50     0.004031849640397045 0.0905773775999997
#> 51      0.10317998221501223       3.2874759002
#> 52     0.003256819107921749  0.179641016199999
#> 53       0.9407697074417514   2.27508502569999
#> 54       0.7085646902378377       5.2033164459
#> 55     0.001028011004730045   1.87596885599999
#> 56  0.000015835535093988717       2.6062499703
#> 57   0.00015972691607413214       5.0482463785
#> 58    0.0012713340353657232  0.435713163799999
#> 59  0.000036258261760346855       0.5439158967
#> 60   0.00008576772798907228   1.62547624519999
#> 61    0.0005047783187366317  0.188652225199998
#> 62   0.00020727821340027486       1.0627373025
#> 63       0.8265336617017218   3.39524150609999
#> 64       0.9118309902912132       2.7499679181
#> 65       0.7803958762265931       8.7869201591
#> 66     0.002248910317168673       6.0252868305
#> 67    0.0011154194995069977        0.727200521
#> 68       0.4288198994607003   7.73617338379999
#> 69        0.844213116573793   9.41128064859999
#> 70  0.000012685559573455407 0.0266620140000006
#> 71  0.000004165417503809721       1.0251003832
#> 72      0.07216956854098999       4.2446507286
#> 73    0.0003796669053852675       3.6034339205
#> 74    0.0012626697935577136       2.0082988577
#> 75    0.0001742161652431573 0.0448168873999993
#> 76  0.000004404477775943327 0.0254067829000002
#> 77       0.9939689644437355   3.04488015159999
#> 78   0.00017042791751720188  0.146911048999999
#> 79   0.00007443135335916118  0.394188331699999
#> 80    0.0001606326780763112  0.738405284700001
#> 81    0.0008418813561436034  0.307032200699999
#> 82    0.0004597288685521672       0.4491663395
#> 83  0.000007992926103650433  0.574617268999999
#> 84        0.747621775556108   5.07418613049999
#> 85      0.02035858702878457       3.7631715548
#> 86       0.5198478586172812       6.9631012532
#> 87      0.04404470034418717       3.8470693333
#> 88       0.6875322954307215   3.91834934429999

  cfbd_game_info(2018, week = 7, conference = "Ind")

#>     game_id season week season_type               start_date start_time_tbd
#> 1 401013452   2018    7     regular 2018-10-13T18:00:00.000Z             NA
#> 2 401013148   2018    7     regular 2018-10-13T18:30:00.000Z             NA
#> 3 401013370   2018    7     regular 2018-10-13T19:30:00.000Z             NA
#> 4 401013442   2018    7     regular 2018-10-13T21:00:00.000Z             NA
#> 5 401016408   2018    7     regular 2018-10-14T02:15:00.000Z             NA
#>   neutral_site conference_game attendance venue_id                  venue
#> 1        FALSE           FALSE          0     3992       Williams Stadium
#> 2        FALSE           FALSE          0     3855     Notre Dame Stadium
#> 3         TRUE           FALSE          0     4738         Levi's Stadium
#> 4        FALSE           FALSE          0     3666            Cajun Field
#> 5        FALSE           FALSE          0     3801 LaVell Edwards Stadium
#>   home_id      home_team  home_conference home_points    home_post_win_prob
#> 1    2335        Liberty FBS Independents          22    0.5483899675354477
#> 2      87     Notre Dame FBS Independents          19    0.9661277906146866
#> 3      23 San José State    Mountain West           3 0.0036361807521906077
#> 4     309      Louisiana         Sun Belt          66    0.9983063510069449
#> 5     252            BYU FBS Independents          49    0.9978157683814658
#>   away_id        away_team  away_conference away_points    away_post_win_prob
#> 1    2653             Troy         Sun Belt          16   0.45161003246455234
#> 2     221       Pittsburgh              ACC          14  0.033872209385313434
#> 3     349             Army FBS Independents          52    0.9963638192478094
#> 4     166 New Mexico State FBS Independents          38 0.0016936489930551302
#> 5      62          Hawai'i    Mountain West          23 0.0021842316185342048
#>   excitement_index
#> 1     8.1162251011
#> 2     7.2536238281
#> 3     2.2159399162
#> 4 2.58899975529999
#> 5 1.08305180749999

  # 7 OTs LSU @ TAMU
  cfbd_game_info(2018, week = 13, team = "Texas A&M", quarter_scores = TRUE)

#>     game_id season week season_type               start_date start_time_tbd
#> 1 401012356   2018   13     regular 2018-11-25T00:30:00.000Z             NA
#>   neutral_site conference_game attendance venue_id      venue home_id home_team
#> 1        FALSE            TRUE          0     3795 Kyle Field     245 Texas A&M
#>   home_conference home_points home_scores_Q1 home_scores_Q2 home_scores_Q3
#> 1             SEC          74              7             10              7
#>   home_scores_Q4 home_scores_Q5 home_scores_Q6 home_scores_Q7 home_scores_Q8
#> 1              7              3              7              8              3
#>   home_scores_Q9 home_scores_Q10 home_scores_Q11 home_post_win_prob away_id
#> 1              6               8               8 0.6786258156883023      99
#>   away_team away_conference away_points away_scores_Q1 away_scores_Q2
#> 1       LSU             SEC          72              7              3
#>   away_scores_Q3 away_scores_Q4 away_scores_Q5 away_scores_Q6 away_scores_Q7
#> 1              7             14              3              7              8
#>   away_scores_Q8 away_scores_Q9 away_scores_Q10 away_scores_Q11
#> 1              3              6               8               6
#>    away_post_win_prob excitement_index
#> 1 0.32137418431169773    14.1962593954

# }