Get results information from games.

Get results information from games.

cfbd_game_info(
  year,
  week = NULL,
  season_type = "regular",
  team = NULL,
  home_team = NULL,
  away_team = NULL,
  conference = NULL,
  game_id = NULL,
  quarter_scores = FALSE
)

Arguments

year
(Integer required): Year, 4 digit format(YYYY)

week
(Integer optional): Week - values from 1-15, 1-14 for seasons pre-playoff (i.e. 2013 or earlier)

season_type
(String default regular): Select Season Type: regular, postseason, or both

team
(String optional): D-I Team

home_team
(String optional): Home D-I Team

away_team
(String optional): Away D-I Team

conference
(String optional): Conference abbreviation - Select a valid FBS conference
Conference abbreviations P5: ACC, B12, B1G, SEC, PAC
Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC

game_id
(Integer optional): Game ID filter for querying a single game

quarter_scores
(Logical default FALSE): This is a parameter to return the list columns that give the score at each quarter: home_line_scores and away_line_scores.
I have defaulted the parameter to false so that you will not have to go to the trouble of dropping it.

Value

cfbd_game_info() - A data frame with 22 variables:

game_id: integer.

Referencing game id.

season: integer.

Season of the game.

week: integer.

Game week.

season_type: character.

Season type of the game.

start_date: character.

Game date.

start_time_tbd: logical.

TRUE/FALSE flag for if the game's start time is to be determined.

neutral_site: logical.

TRUE/FALSE flag for the game taking place at a neutral site.

conference_game: logical.

TRUE/FALSE flag for this game qualifying as a conference game.

attendance: integer.

Reported attendance at the game.

venue_id: integer.

Referencing venue id.

venue: character.

Venue name.

home_id: integer.

Home team referencing id.

home_team: character.

Home team name.

home_conference: character.

Home team conference.

home_points: integer.

Home team points.

home_post_win_prob: character.

Home team post-game win probability.

home_pregame_elo: character.

Home team pre-game ELO rating.

home_postgame_elo: character.

Home team post-game ELO rating.

away_id: integer.

Away team referencing id.

away_team: character.

Away team name.

away_conference: character.

Away team conference.

away_points: integer.

Away team points.

away_post_win_prob: character.

Away team post-game win probability.

away_pregame_elo: character.

Away team pre-game ELO rating.

away_postgame_elo: character.

Away team post-game ELO rating.

excitement_index: character.

Game excitement index.

highlights: character.

Game highlight urls.

notes: character.

Game notes.

Examples

# \donttest{
  cfbd_game_info(2018, week = 1)
#>      game_id season week season_type               start_date start_time_tbd
#> 1  401013357   2018    1     regular 2018-08-25T21:30:00.000Z             NA
#> 2  401014972   2018    1     regular 2018-08-25T23:00:00.000Z             NA
#> 3  401022510   2018    1     regular 2018-08-25T23:30:00.000Z             NA
#> 4  401013437   2018    1     regular 2018-08-26T02:00:00.000Z             NA
#> 5  401020671   2018    1     regular 2018-08-30T23:00:00.000Z             NA
#> 6  401013328   2018    1     regular 2018-08-30T23:00:00.000Z             NA
#> 7  401019470   2018    1     regular 2018-08-30T23:00:00.000Z             NA
#> 8  401021651   2018    1     regular 2018-08-30T23:00:00.000Z             NA
#> 9  401014979   2018    1     regular 2018-08-31T00:00:00.000Z             NA
#> 10 401013489   2018    1     regular 2018-08-31T00:00:00.000Z             NA
#> 11 401021655   2018    1     regular 2018-08-31T00:00:00.000Z             NA
#> 12 401012675   2018    1     regular 2018-08-31T00:00:00.000Z             NA
#> 13 401013342   2018    1     regular 2018-08-31T00:00:00.000Z             NA
#> 14 401013086   2018    1     regular 2018-08-31T00:00:00.000Z             NA
#> 15 401012258   2018    1     regular 2018-08-31T00:30:00.000Z             NA
#> 16 401022511   2018    1     regular 2018-08-31T02:00:00.000Z             NA
#> 17 401020673   2018    1     regular 2018-08-31T22:00:00.000Z             NA
#> 18 401020672   2018    1     regular 2018-08-31T22:30:00.000Z             NA
#> 19 401013089   2018    1     regular 2018-08-31T23:00:00.000Z             NA
#> 20 401012880   2018    1     regular 2018-08-31T23:00:00.000Z             NA
#> 21 401013354   2018    1     regular 2018-09-01T01:00:00.000Z             NA
#> 22 401022512   2018    1     regular 2018-09-01T01:00:00.000Z             NA
#> 23 401022517   2018    1     regular 2018-09-01T01:00:00.000Z             NA
#> 24 401012680   2018    1     regular 2018-09-01T01:30:00.000Z             NA
#> 25 401012253   2018    1     regular 2018-09-01T16:00:00.000Z             NA
#> 26 401012820   2018    1     regular 2018-09-01T16:00:00.000Z             NA
#> 27 401012872   2018    1     regular 2018-09-01T16:00:00.000Z             NA
#> 28 401012256   2018    1     regular 2018-09-01T16:00:00.000Z             NA
#> 29 401013091   2018    1     regular 2018-09-01T16:00:00.000Z             NA
#> 30 401012682   2018    1     regular 2018-09-01T16:00:00.000Z             NA
#> 31 401014977   2018    1     regular 2018-09-01T16:00:00.000Z             NA
#> 32 401012804   2018    1     regular 2018-09-01T16:00:00.000Z             NA
#> 33 401013072   2018    1     regular 2018-09-01T16:00:00.000Z             NA
#> 34 401019473   2018    1     regular 2018-09-01T16:00:00.000Z             NA
#> 35 401013065   2018    1     regular 2018-09-01T16:00:00.000Z             NA
#> 36 401013088   2018    1     regular 2018-09-01T16:20:00.000Z             NA
#> 37 401013090   2018    1     regular 2018-09-01T16:30:00.000Z             NA
#> 38 401013087   2018    1     regular 2018-09-01T17:00:00.000Z             NA
#> 39 401022514   2018    1     regular 2018-09-01T18:00:00.000Z             NA
#> 40 401020675   2018    1     regular 2018-09-01T18:00:00.000Z             NA
#> 41 401012250   2018    1     regular 2018-09-01T19:30:00.000Z             NA
#> 42 401012248   2018    1     regular 2018-09-01T19:30:00.000Z             NA
#> 43 401012251   2018    1     regular 2018-09-01T19:30:00.000Z             NA
#> 44 401013349   2018    1     regular 2018-09-01T19:30:00.000Z             NA
#> 45 401012683   2018    1     regular 2018-09-01T19:30:00.000Z             NA
#> 46 401012257   2018    1     regular 2018-09-01T19:30:00.000Z             NA
#> 47 401013092   2018    1     regular 2018-09-01T19:30:00.000Z             NA
#> 48 401012862   2018    1     regular 2018-09-01T19:40:00.000Z             NA
#> 49 401012255   2018    1     regular 2018-09-01T20:00:00.000Z             NA
#> 50 401012247   2018    1     regular 2018-09-01T20:00:00.000Z             NA
#> 51 401012684   2018    1     regular 2018-09-01T20:00:00.000Z             NA
#> 52 401012678   2018    1     regular 2018-09-01T20:00:00.000Z             NA
#> 53 401014975   2018    1     regular 2018-09-01T22:00:00.000Z             NA
#> 54 401020674   2018    1     regular 2018-09-01T22:00:00.000Z             NA
#> 55 401014973   2018    1     regular 2018-09-01T22:00:00.000Z             NA
#> 56 401021656   2018    1     regular 2018-09-01T22:00:00.000Z             NA
#> 57 401019472   2018    1     regular 2018-09-01T22:00:00.000Z             NA
#> 58 401021653   2018    1     regular 2018-09-01T22:00:00.000Z             NA
#> 59 401013094   2018    1     regular 2018-09-01T22:00:00.000Z             NA
#> 60 401013448   2018    1     regular 2018-09-01T22:00:00.000Z             NA
#> 61 401013048   2018    1     regular 2018-09-01T23:00:00.000Z             NA
#> 62 401012851   2018    1     regular 2018-09-01T23:00:00.000Z             NA
#> 63 401013057   2018    1     regular 2018-09-01T23:00:00.000Z             NA
#> 64 401014978   2018    1     regular 2018-09-01T23:00:00.000Z             NA
#> 65 401019476   2018    1     regular 2018-09-01T23:00:00.000Z             NA
#> 66 401012679   2018    1     regular 2018-09-01T23:00:00.000Z             NA
#> 67 401019474   2018    1     regular 2018-09-01T23:00:00.000Z             NA
#> 68 401020676   2018    1     regular 2018-09-01T23:00:00.000Z             NA
#> 69 401021652   2018    1     regular 2018-09-01T23:00:00.000Z             NA
#> 70 401014974   2018    1     regular 2018-09-01T23:00:00.000Z             NA
#> 71 401021654   2018    1     regular 2018-09-01T23:00:00.000Z             NA
#> 72 401014976   2018    1     regular 2018-09-01T23:30:00.000Z             NA
#> 73 401012259   2018    1     regular 2018-09-01T23:30:00.000Z             NA
#> 74 401012888   2018    1     regular 2018-09-01T23:30:00.000Z             NA
#> 75 401012249   2018    1     regular 2018-09-01T23:30:00.000Z             NA
#> 76 401012254   2018    1     regular 2018-09-01T23:30:00.000Z             NA
#> 77 401014980   2018    1     regular 2018-09-01T23:30:00.000Z             NA
#> 78 401013026   2018    1     regular 2018-09-02T00:00:00.000Z             NA
#> 79 401022515   2018    1     regular 2018-09-02T00:00:00.000Z             NA
#> 80 401012681   2018    1     regular 2018-09-02T00:00:00.000Z             NA
#> 81 401012246   2018    1     regular 2018-09-02T00:00:00.000Z             NA
#> 82 401022516   2018    1     regular 2018-09-02T02:00:00.000Z             NA
#> 83 401012677   2018    1     regular 2018-09-02T02:30:00.000Z             NA
#> 84 401012676   2018    1     regular 2018-09-02T02:45:00.000Z             NA
#> 85 401019475   2018    1     regular 2018-09-02T03:00:00.000Z             NA
#> 86 401079833   2018    1     regular 2018-09-02T19:30:00.000Z             NA
#> 87 401013095   2018    1     regular 2018-09-02T23:30:00.000Z             NA
#> 88 401013096   2018    1     regular 2018-09-04T00:00:00.000Z             NA
#>    neutral_site conference_game attendance venue_id
#> 1         FALSE           FALSE       8684     3985
#> 2         FALSE           FALSE      20050     3895
#> 3         FALSE            TRUE      31007     5388
#> 4         FALSE           FALSE      20322     3601
#> 5         FALSE           FALSE          0     3919
#> 6         FALSE           FALSE          0     3953
#> 7         FALSE            TRUE          0     3892
#> 8         FALSE           FALSE          0     4246
#> 9         FALSE           FALSE          0     3803
#> 10        FALSE           FALSE          0     3646
#> 11        FALSE           FALSE          0     3817
#> 12        FALSE           FALSE          0      587
#> 13        FALSE            TRUE          0     3907
#> 14        FALSE           FALSE          0     4729
#> 15        FALSE           FALSE          0     3795
#> 16        FALSE           FALSE          0     3935
#> 17        FALSE           FALSE          0     3980
#> 18        FALSE           FALSE          0     3912
#> 19        FALSE           FALSE          0     3982
#> 20        FALSE           FALSE          0     3936
#> 21        FALSE           FALSE          0      347
#> 22        FALSE           FALSE          0     3940
#> 23        FALSE           FALSE          0     3816
#> 24         TRUE           FALSE          0     3937
#> 25         TRUE           FALSE          0     3891
#> 26        FALSE           FALSE          0     3754
#> 27         TRUE           FALSE          0     3719
#> 28        FALSE           FALSE          0     3994
#> 29        FALSE           FALSE          0     3670
#> 30        FALSE           FALSE          0     3861
#> 31        FALSE           FALSE          0     3895
#> 32        FALSE           FALSE          0     3832
#> 33        FALSE           FALSE          0     3616
#> 34        FALSE           FALSE          0     3806
#> 35        FALSE           FALSE          0     3835
#> 36        FALSE           FALSE          0     3836
#> 37        FALSE           FALSE          0      721
#> 38        FALSE           FALSE          0     3615
#> 39        FALSE           FALSE          0     3713
#> 40        FALSE           FALSE          0     3876
#> 41        FALSE           FALSE          0     3917
#> 42         TRUE           FALSE          0     5348
#> 43        FALSE           FALSE          0     3683
#> 44        FALSE           FALSE     105232     3632
#> 45        FALSE           FALSE          0     3984
#> 46         TRUE           FALSE          0     3628
#> 47        FALSE           FALSE          0     3752
#> 48        FALSE           FALSE          0     3793
#> 49        FALSE           FALSE          0     3838
#> 50        FALSE           FALSE          0     3887
#> 51        FALSE           FALSE          0     3831
#> 52        FALSE           FALSE          0      477
#> 53        FALSE           FALSE          0     3996
#> 54        FALSE           FALSE          0     3965
#> 55        FALSE           FALSE          0     4418
#> 56        FALSE           FALSE          0     3975
#> 57        FALSE           FALSE          0     3886
#> 58        FALSE           FALSE          0     3608
#> 59        FALSE           FALSE          0     3923
#> 60        FALSE           FALSE          0     3992
#> 61        FALSE           FALSE          0     3833
#> 62        FALSE           FALSE          0      218
#> 63        FALSE           FALSE          0     3636
#> 64        FALSE           FALSE          0     3815
#> 65        FALSE           FALSE          0     3674
#> 66        FALSE           FALSE          0     1056
#> 67        FALSE           FALSE          0     3805
#> 68        FALSE           FALSE          0     3739
#> 69        FALSE           FALSE          0     3625
#> 70        FALSE           FALSE          0     3797
#> 71        FALSE           FALSE          0     3666
#> 72        FALSE           FALSE          0     3825
#> 73        FALSE           FALSE          0     3973
#> 74        FALSE           FALSE          0     3855
#> 75        FALSE           FALSE          0     3634
#> 76        FALSE           FALSE          0     3693
#> 77        FALSE           FALSE          0     3946
#> 78        FALSE           FALSE          0     4727
#> 79        FALSE           FALSE          0     3971
#> 80        FALSE           FALSE          0     3626
#> 81         TRUE           FALSE      57280     4013
#> 82        FALSE           FALSE          0     3660
#> 83        FALSE           FALSE          0     3947
#> 84        FALSE           FALSE          0     3619
#> 85        FALSE           FALSE          0     3610
#> 86        FALSE           FALSE          0     3699
#> 87         TRUE           FALSE          0     3687
#> 88        FALSE            TRUE          0     3697
#>                                                       venue home_id
#> 1                             Warren McGuirk Alumni Stadium     113
#> 2                                              Rice Stadium     242
#> 3                                            Canvas Stadium      36
#> 4                                    Aggie Memorial Stadium     166
#> 5                                         Scheumann Stadium    2050
#> 6                                          TCF Bank Stadium     135
#> 7                                          Rentschler Field      41
#> 8                                     Georgia State Stadium    2247
#> 9                                              Legion Field       5
#> 10                                    Boone Pickens Stadium     197
#> 11                                           Malone Stadium    2433
#> 12                                      Rice-Eccles Stadium     254
#> 13                                         Ross-Ade Stadium    2509
#> 14                           Benson Field at Yulman Stadium    2655
#> 15                                               Kyle Field     245
#> 16                                          Spartan Stadium      23
#> 17                                            Waldo Stadium    2711
#> 18                                        Rynearson Stadium    2199
#> 19                                     Wallace Wade Stadium     150
#> 20                                          Spartan Stadium     127
#> 21                                     Camp Randall Stadium     275
#> 22                                         Stanford Stadium      24
#> 23                                           Mackay Stadium    2440
#> 24                      Sports Authority Field at Mile High      36
#> 25                                              NRG Stadium    2641
#> 26                             High Point Solutions Stadium     164
#> 27                                              FedEx Field     120
#> 28                                   Williams-Brice Stadium    2579
#> 29                                    Carter-Finley Stadium     152
#> 30                                             Ohio Stadium     194
#> 31                                             Rice Stadium     242
#> 32                                         Memorial Stadium     356
#> 33                                   Amon G. Carter Stadium    2628
#> 34                                  Lincoln Financial Field     218
#> 35                 Gaylord Family Oklahoma Memorial Stadium     201
#> 36                                         Memorial Stadium     228
#> 37               Bobby Dodd Stadium at Historic Grant Field      59
#> 38                                           Alumni Stadium     103
#> 39                                           Falcon Stadium    2005
#> 40                                            Peden Stadium     195
#> 41                                          Sanford Stadium      61
#> 42                                    Mercedes-Benz Stadium       2
#> 43                                     Commonwealth Stadium      96
#> 44                                           Beaver Stadium     213
#> 45                                     War Memorial Stadium    2751
#> 46                                  Bank of America Stadium    2633
#> 47                                              Heinz Field     221
#> 48                                          Kinnick Stadium    2294
#> 49                                             Faurot Field     142
#> 50 Donald W. Reynolds Razorback Stadium Frank Broyles Field       8
#> 51                              California Memorial Stadium      25
#> 52                            Los Angeles Memorial Coliseum      30
#> 53                                            Yager Stadium     193
#> 54                                               UB Stadium    2084
#> 55                                 Jerry Richardson Stadium    2429
#> 56                                Veterans Memorial Stadium    2653
#> 57                                    Raymond James Stadium      58
#> 58                                          Paulson Stadium     290
#> 59                                            Scott Stadium     258
#> 60                                         Williams Stadium    2335
#> 61                                         Memorial Stadium    2305
#> 62                                              FIU Stadium    2229
#> 63                      Bill Snyder Family Football Stadium    2306
#> 64                                    M. M. Roberts Stadium    2572
#> 65                    Skelly Field at H. A. Chapman Stadium     202
#> 66                                                Rose Bowl      26
#> 67                            Liberty Bowl Memorial Stadium     235
#> 68                                               Glass Bowl    2649
#> 69                                  Centennial Bank Stadium    2032
#> 70                                     Ladd Peebles Stadium       6
#> 71                                              Cajun Field     309
#> 72                                           Apogee Stadium     249
#> 73                                       Vanderbilt Stadium     238
#> 74                                       Notre Dame Stadium      87
#> 75                                 Ben Hill Griffin Stadium      57
#> 76                                       Davis Wade Stadium     344
#> 77                                         Sun Bowl Stadium    2638
#> 78                                           McLane Stadium     239
#> 79                                       University Stadium     167
#> 80                                           Autzen Stadium    2483
#> 81                                    Camping World Stadium     333
#> 82                                          Bulldog Stadium     278
#> 83                                        Sun Devil Stadium       9
#> 84                                          Arizona Stadium      12
#> 85                                            Aloha Stadium      62
#> 86                                    Dowdy-Ficklen Stadium     151
#> 87                                             AT&T Stadium      99
#> 88              Bobby Bowden Field at Doak Campbell Stadium      52
#>                home_team   home_conference home_points   home_post_win_prob
#> 1                  UMass  FBS Independents          63   0.9984542422162311
#> 2                   Rice    Conference USA          31  0.41611197976904707
#> 3         Colorado State     Mountain West          34   0.3317397678892768
#> 4       New Mexico State  FBS Independents           7  0.00427210984632397
#> 5             Ball State      Mid-American          42   0.9999720937904223
#> 6              Minnesota           Big Ten          48   0.9999638364577065
#> 7            Connecticut American Athletic          17 0.004203463433869054
#> 8          Georgia State          Sun Belt          24    0.869590187971506
#> 9                    UAB    Conference USA          52    0.999997345694123
#> 10        Oklahoma State            Big 12          58   0.9992698560515085
#> 11      Louisiana Monroe          Sun Belt          34   0.5720859305615573
#> 12                  Utah            Pac-12          41   0.9997998688829266
#> 13                Purdue           Big Ten          27  0.15206580002046932
#> 14                Tulane American Athletic          17   0.2196613212325033
#> 15             Texas A&M               SEC          59   0.9999928023514462
#> 16        San José State     Mountain West          38   0.2255818432658651
#> 17      Western Michigan      Mid-American          42  0.10069975562117696
#> 18      Eastern Michigan      Mid-American          51   0.9995510164694754
#> 19                  Duke               ACC          34   0.9973846191967987
#> 20        Michigan State           Big Ten          38    0.906183590742433
#> 21             Wisconsin           Big Ten          34   0.9999658482595446
#> 22              Stanford            Pac-12          31   0.9937677403365777
#> 23                Nevada     Mountain West          72   0.9999946772201084
#> 24        Colorado State     Mountain West          13 0.003935607921436278
#> 25            Texas Tech            Big 12          27 0.007730744808323005
#> 26               Rutgers           Big Ten          35   0.9992581309620951
#> 27              Maryland           Big Ten          34   0.9131500976054453
#> 28        South Carolina               SEC          49   0.9999813927842912
#> 29              NC State               ACC          24    0.839090960760694
#> 30            Ohio State           Big Ten          77   0.9999435276188426
#> 31                  Rice    Conference USA          27  0.05005196033053275
#> 32              Illinois           Big Ten          31   0.8340153107033903
#> 33                   TCU            Big 12          55   0.9999921756502518
#> 34                Temple American Athletic          17  0.08018018010651985
#> 35              Oklahoma            Big 12          63   0.9999928028169979
#> 36               Clemson               ACC          48   0.9999840784279569
#> 37          Georgia Tech               ACC          41   0.9999959877845358
#> 38        Boston College               ACC          55   0.9998255691998653
#> 39             Air Force     Mountain West          38   0.9999947029448708
#> 40                  Ohio      Mid-American          38   0.4527258093446935
#> 41               Georgia               SEC          45    0.999992203255419
#> 42                Auburn               SEC          21   0.6081821946558714
#> 43              Kentucky               SEC          35   0.9948967001914114
#> 44            Penn State           Big Ten          45    0.980509601522245
#> 45               Wyoming     Mountain West          19 0.005358724710402157
#> 46             Tennessee               SEC          14 0.004273317866598812
#> 47            Pittsburgh               ACC          33   0.9997365885384938
#> 48                  Iowa           Big Ten          33   0.9987837404630044
#> 49              Missouri               SEC          51    0.995968150359603
#> 50              Arkansas               SEC          55   0.9967431808920783
#> 51            California            Pac-12          24   0.8968200177849878
#> 52                   USC            Pac-12          43   0.9967409790322357
#> 53            Miami (OH)      Mid-American          28   0.2914353097621623
#> 54               Buffalo      Mid-American          48   0.9999142322720109
#> 55             Charlotte    Conference USA          34    0.999984164464906
#> 56                  Troy          Sun Belt          20  0.05923029255824867
#> 57         South Florida American Athletic          34   0.9999637417382397
#> 58      Georgia Southern          Sun Belt          37   0.9987286659646343
#> 59              Virginia               ACC          42     0.99897198899527
#> 60               Liberty  FBS Independents          52   0.9998402730839259
#> 61                Kansas            Big 12          23  0.15578688342620697
#> 62 Florida International    Conference USA          28  0.08816900970878677
#> 63          Kansas State            Big 12          27   0.5711801005392997
#> 64  Southern Mississippi    Conference USA          55   0.9994952216812634
#> 65                 Tulsa American Athletic          38   0.9977510896828313
#> 66                  UCLA            Pac-12          17   0.2196041237734069
#> 67               Memphis American Athletic          66   0.9999958345824962
#> 68                Toledo      Mid-American          66   0.9999873144404265
#> 69        Arkansas State          Sun Belt          48    0.998884580500493
#> 70         South Alabama          Sun Belt          26  0.17346633829827812
#> 71             Louisiana          Sun Belt          49   0.9997927217865997
#> 72           North Texas    Conference USA          46   0.9987373302064423
#> 73            Vanderbilt               SEC          35   0.9996203330946147
#> 74            Notre Dame  FBS Independents          24     0.92783043145901
#> 75               Florida               SEC          53   0.9998257838347568
#> 76     Mississippi State               SEC          63   0.9999955955222241
#> 77                  UTEP    Conference USA          10 0.006031035556264542
#> 78                Baylor            Big 12          55   0.9998295720824828
#> 79            New Mexico     Mountain West          62   0.9991581186438564
#> 80                Oregon            Pac-12          58   0.9998393673219237
#> 81               Alabama               SEC          51   0.9999255686466408
#> 82          Fresno State     Mountain West          79   0.9995402711314478
#> 83         Arizona State            Pac-12          49   0.9999920070738963
#> 84               Arizona            Pac-12          23    0.252378224443892
#> 85               Hawai'i     Mountain West          59   0.9796414129712154
#> 86         East Carolina American Athletic          23   0.4801521413827188
#> 87                   LSU               SEC          33   0.9559552996558128
#> 88         Florida State               ACC           3  0.31246770456927847
#>    home_pregame_elo home_postgame_elo away_id                away_team
#> 1                NA                NA    2184                 Duquesne
#> 2                NA                NA    2504             Prairie View
#> 3              1529              1482      62                  Hawai'i
#> 4              1339              1304    2751                  Wyoming
#> 5                NA                NA    2115      Central Connecticut
#> 6              1499              1582     166         New Mexico State
#> 7              1242              1190    2116                      UCF
#> 8                NA                NA     338           Kennesaw State
#> 9                NA                NA    2542           Savannah State
#> 10               NA                NA    2623           Missouri State
#> 11               NA                NA    2545   Southeastern Louisiana
#> 12               NA                NA    2692              Weber State
#> 13             1554              1556      77             Northwestern
#> 14             1391              1394     154              Wake Forest
#> 15               NA                NA    2466       Northwestern State
#> 16               NA                NA     302                 UC Davis
#> 17             1553              1502     183                 Syracuse
#> 18               NA                NA    2405                 Monmouth
#> 19             1593              1632     349                     Army
#> 20             1622              1623     328               Utah State
#> 21             1851              1886      98         Western Kentucky
#> 22             1757              1789      21          San Diego State
#> 23               NA                NA    2502           Portland State
#> 24             1482              1401      38                 Colorado
#> 25             1518              1477     145                 Ole Miss
#> 26             1231              1292     326              Texas State
#> 27             1301              1335     251                    Texas
#> 28             1572              1634     324         Coastal Carolina
#> 29               NA                NA     256            James Madison
#> 30             1915              1963     204             Oregon State
#> 31             1164              1160     248                  Houston
#> 32             1305              1305    2309               Kent State
#> 33               NA                NA    2582                 Southern
#> 34               NA                NA     222                Villanova
#> 35             1876              1992    2226         Florida Atlantic
#> 36               NA                NA     231                   Furman
#> 37               NA                NA    2016             Alcorn State
#> 38             1609              1674     113                    UMass
#> 39               NA                NA    2619              Stony Brook
#> 40               NA                NA      47                   Howard
#> 41               NA                NA    2046              Austin Peay
#> 42             1829              1840     264               Washington
#> 43             1453              1482    2117         Central Michigan
#> 44             1922              1917    2026        Appalachian State
#> 45             1572              1513     265         Washington State
#> 46             1411              1368     277            West Virginia
#> 47               NA                NA     399                   Albany
#> 48             1712              1756    2459        Northern Illinois
#> 49               NA                NA    2630                UT Martin
#> 50               NA                NA    2197         Eastern Illinois
#> 51             1556              1560     153           North Carolina
#> 52             1713              1730    2439                     UNLV
#> 53             1385              1379     276                 Marshall
#> 54               NA                NA    2169           Delaware State
#> 55               NA                NA    2230                  Fordham
#> 56             1594              1496      68              Boise State
#> 57               NA                NA    2210                     Elon
#> 58               NA                NA    2569     South Carolina State
#> 59               NA                NA     257                 Richmond
#> 60             1444              1541     295             Old Dominion
#> 61               NA                NA    2447                 Nicholls
#> 62             1377              1371      84                  Indiana
#> 63               NA                NA     233             South Dakota
#> 64               NA                NA    2296            Jackson State
#> 65               NA                NA    2110         Central Arkansas
#> 66             1471              1437    2132               Cincinnati
#> 67               NA                NA    2382                   Mercer
#> 68               NA                NA    2678                      VMI
#> 69               NA                NA    2546 Southeast Missouri State
#> 70             1233              1245    2348           Louisiana Tech
#> 71               NA                NA    2755                Grambling
#> 72             1335              1397    2567                      SMU
#> 73             1414              1488    2393         Middle Tennessee
#> 74             1722              1730     130                 Michigan
#> 75               NA                NA    2127      Charleston Southern
#> 76               NA                NA    2617        Stephen F. Austin
#> 77               NA                NA    2464         Northern Arizona
#> 78               NA                NA    2000        Abilene Christian
#> 79               NA                NA    2916           Incarnate Word
#> 80             1613              1664     189            Bowling Green
#> 81             1986              2062      97               Louisville
#> 82               NA                NA      70                    Idaho
#> 83             1498              1612    2636           UT San Antonio
#> 84             1513              1496     252                      BYU
#> 85             1264              1345    2426                     Navy
#> 86               NA                NA    2448       North Carolina A&T
#> 87             1736              1774    2390                    Miami
#> 88             1647              1603     259            Virginia Tech
#>      away_conference away_points       away_post_win_prob away_pregame_elo
#> 1               <NA>          15     0.001545757783768864               NA
#> 2               <NA>          28       0.5838880202309529               NA
#> 3      Mountain West          43       0.6682602321107232             1217
#> 4      Mountain West          29        0.995727890153676             1537
#> 5               <NA>           6  0.000027906209577710506               NA
#> 6   FBS Independents          10   0.00003616354229352137             1304
#> 7  American Athletic          56        0.995796536566131             1729
#> 8               <NA>          20        0.130409812028494               NA
#> 9               <NA>           0 0.0000026543058769812333               NA
#> 10              <NA>          17    0.0007301439484914685               NA
#> 11              <NA>          31       0.4279140694384427               NA
#> 12              <NA>          10    0.0002001311170733988               NA
#> 13           Big Ten          31       0.8479341999795307             1690
#> 14               ACC          23       0.7803386787674967             1590
#> 15              <NA>           7  0.000007197648553836089               NA
#> 16              <NA>          44       0.7744181567341348               NA
#> 17               ACC          55        0.899300244378823             1340
#> 18              <NA>          17    0.0004489835305245826               NA
#> 19  FBS Independents          14    0.0026153808032013393             1521
#> 20     Mountain West          31      0.09381640925756696             1475
#> 21    Conference USA           3   0.00003415174045540148             1436
#> 22     Mountain West          10     0.006232259663422335             1584
#> 23              <NA>          19  0.000005322779891625551               NA
#> 24            Pac-12          45       0.9960643920785637             1495
#> 25               SEC          47        0.992269255191677             1533
#> 26          Sun Belt           7     0.000741869037904852             1131
#> 27            Big 12          29       0.0868499023945547             1657
#> 28          Sun Belt          15  0.000018607215708787983             1315
#> 29              <NA>          13        0.160909039239306               NA
#> 30            Pac-12          31   0.00005647238115735753             1229
#> 31 American Athletic          45       0.9499480396694673             1574
#> 32      Mid-American          24       0.1659846892966097             1129
#> 33              <NA>           7  0.000007824349748219994               NA
#> 34              <NA>          19       0.9198198198934802               NA
#> 35    Conference USA          14  0.000007197183002127794             1642
#> 36              <NA>           7    0.0000159215720431094               NA
#> 37              <NA>           0   0.00000401221546419972               NA
#> 38  FBS Independents          21   0.00017443080013468393             1377
#> 39              <NA>           0  0.000005297055129194206               NA
#> 40              <NA>          32       0.5472741906553065               NA
#> 41              <NA>           0  0.000007796744580956805               NA
#> 42            Pac-12          16       0.3918178053441286             1839
#> 43      Mid-American          20    0.0051032998085885906             1424
#> 44          Sun Belt          38     0.019490398477755022             1652
#> 45            Pac-12          41       0.9946412752895979             1551
#> 46            Big 12          40       0.9957266821334012             1570
#> 47              <NA>           7    0.0002634114615062133               NA
#> 48      Mid-American           7    0.0012162595369955564             1515
#> 49              <NA>          14     0.004031849640397045               NA
#> 50              <NA>          20     0.003256819107921749               NA
#> 51               ACC          17      0.10317998221501223             1459
#> 52     Mountain West          21    0.0032590209677643145             1350
#> 53    Conference USA          35       0.7085646902378377             1475
#> 54              <NA>          10   0.00008576772798907228               NA
#> 55              <NA>          10  0.000015835535093988717               NA
#> 56     Mountain West          56       0.9407697074417514             1648
#> 57              <NA>          14  0.000036258261760346855               NA
#> 58              <NA>           6    0.0012713340353657232               NA
#> 59              <NA>          13     0.001028011004730045               NA
#> 60    Conference USA          10   0.00015972691607413214             1251
#> 61              <NA>          26        0.844213116573793               NA
#> 62           Big Ten          38       0.9118309902912132             1557
#> 63              <NA>          24       0.4288198994607003               NA
#> 64              <NA>           7    0.0005047783187366317               NA
#> 65              <NA>          27     0.002248910317168673               NA
#> 66 American Athletic          26       0.7803958762265931             1291
#> 67              <NA>          14  0.000004165417503809721               NA
#> 68              <NA>           3  0.000012685559573455407               NA
#> 69              <NA>          21    0.0011154194995069977               NA
#> 70    Conference USA          30       0.8265336617017218             1537
#> 71              <NA>          17   0.00020727821340027486               NA
#> 72 American Athletic          23    0.0012626697935577136             1387
#> 73    Conference USA           7    0.0003796669053852675             1421
#> 74           Big Ten          17      0.07216956854098999             1701
#> 75              <NA>           6    0.0001742161652431573               NA
#> 76              <NA>           6  0.000004404477775943327               NA
#> 77              <NA>          30       0.9939689644437355               NA
#> 78              <NA>          27   0.00017042791751720188               NA
#> 79              <NA>          30    0.0008418813561436034               NA
#> 80      Mid-American          24    0.0001606326780763112             1265
#> 81               ACC          14   0.00007443135335916118             1701
#> 82              <NA>          13    0.0004597288685521672               NA
#> 83    Conference USA           7  0.000007992926103650433             1429
#> 84  FBS Independents          28        0.747621775556108             1396
#> 85 American Athletic          41      0.02035858702878457             1633
#> 86              <NA>          28       0.5198478586172812               NA
#> 87               ACC          17      0.04404470034418717             1695
#> 88               ACC          24       0.6875322954307215             1731
#>    away_postgame_elo   excitement_index
#> 1                 NA       0.2139101303
#> 2                 NA       5.6012773743
#> 3               1264       4.5649361931
#> 4               1572       2.7057118396
#> 5                 NA  0.760100617799999
#> 6               1221   3.06144440259999
#> 7               1781  0.830061928399999
#> 8                 NA       5.8512641491
#> 9                 NA       0.1406151783
#> 10                NA 0.0663753880999998
#> 11                NA       8.4160129207
#> 12                NA        0.922426719
#> 13              1688       6.6741715963
#> 14              1587       9.7884823983
#> 15                NA 0.0327134787999997
#> 16                NA       5.5811874724
#> 17              1391       2.7679697627
#> 18                NA  0.565053273800001
#> 19              1482   1.69744162409999
#> 20              1474       4.1366972475
#> 21              1401       0.2437023338
#> 22              1552       2.8895508744
#> 23                NA   1.08662059399999
#> 24              1576         3.41109204
#> 25              1574       3.6518173425
#> 26              1070   1.11697903199999
#> 27              1623   9.13837292739999
#> 28              1253  0.182912192500001
#> 29                NA       3.8739046233
#> 30              1181       0.1409114981
#> 31              1578       3.9567502631
#> 32              1129   6.71988011069999
#> 33                NA 0.0225499731999999
#> 34                NA   6.43306846939999
#> 35              1526  0.451379604799999
#> 36                NA 0.0250883832000001
#> 37                NA 0.0701338293999992
#> 38              1312       0.6622090949
#> 39                NA       0.8860730608
#> 40                NA       3.8302068262
#> 41                NA 0.0234045444000003
#> 42              1828   7.48070415839999
#> 43              1395   5.87733304869999
#> 44              1657   6.47689498139999
#> 45              1610       7.1429531895
#> 46              1613   2.82384215719999
#> 47                NA       0.2208688393
#> 48              1471       3.0460114328
#> 49                NA 0.0905773775999997
#> 50                NA  0.179641016199999
#> 51              1455       3.2874759002
#> 52              1333       2.4032706517
#> 53              1481       5.2033164459
#> 54                NA   1.62547624519999
#> 55                NA       2.6062499703
#> 56              1746   2.27508502569999
#> 57                NA       0.5439158967
#> 58                NA  0.435713163799999
#> 59                NA   1.87596885599999
#> 60              1154       5.0482463785
#> 61                NA   9.41128064859999
#> 62              1563       2.7499679181
#> 63                NA   7.73617338379999
#> 64                NA  0.188652225199998
#> 65                NA       6.0252868305
#> 66              1325       8.7869201591
#> 67                NA       1.0251003832
#> 68                NA 0.0266620140000006
#> 69                NA        0.727200521
#> 70              1525   3.39524150609999
#> 71                NA       1.0627373025
#> 72              1325       2.0082988577
#> 73              1347       3.6034339205
#> 74              1693       4.2446507286
#> 75                NA 0.0448168873999993
#> 76                NA 0.0254067829000002
#> 77                NA   3.04488015159999
#> 78                NA  0.146911048999999
#> 79                NA  0.307032200699999
#> 80              1214  0.738405284700001
#> 81              1625  0.394188331699999
#> 82                NA       0.4491663395
#> 83              1315  0.574617268999999
#> 84              1413   5.07418613049999
#> 85              1552       3.7631715548
#> 86                NA       6.9631012532
#> 87              1657       3.8470693333
#> 88              1775   3.91834934429999
#>                                     highlights notes
#> 1  https://www.youtube.com/watch?v=uwMMvX-tU_U    NA
#> 2  https://www.youtube.com/watch?v=z2VFteZqBBw    NA
#> 3  https://www.youtube.com/watch?v=1FEy3ogBwpM    NA
#> 4  https://www.youtube.com/watch?v=S3V1Juvj160    NA
#> 5                                         <NA>    NA
#> 6  https://www.youtube.com/watch?v=hpuPjxkqUWI    NA
#> 7  https://www.youtube.com/watch?v=ZPI-FjeNw4A    NA
#> 8  https://www.youtube.com/watch?v=kUls9PpCkuc    NA
#> 9  https://www.youtube.com/watch?v=gKQoi3sPeic    NA
#> 10 https://www.youtube.com/watch?v=qi5_zqyEvDs    NA
#> 11                                        <NA>    NA
#> 12 https://www.youtube.com/watch?v=tJjO6_6SP2s    NA
#> 13 https://www.youtube.com/watch?v=6zZEHpc246A    NA
#> 14 https://www.youtube.com/watch?v=9Pgaey84Fhw    NA
#> 15 https://www.youtube.com/watch?v=Vk-bY7sQ4u8    NA
#> 16                                        <NA>    NA
#> 17 https://www.youtube.com/watch?v=9KWOUJL0s00    NA
#> 18                                        <NA>    NA
#> 19 https://www.youtube.com/watch?v=MBbw4OTY-i4    NA
#> 20 https://www.youtube.com/watch?v=BtDvvMyLLAg    NA
#> 21 https://www.youtube.com/watch?v=OX6SNSLJTxs    NA
#> 22 https://www.youtube.com/watch?v=ZqxQVQVhqtE    NA
#> 23 https://www.youtube.com/watch?v=eiXWWtKGd94    NA
#> 24 https://www.youtube.com/watch?v=IVggtfJsL8w    NA
#> 25 https://www.youtube.com/watch?v=aKDo9I4t61U    NA
#> 26 https://www.youtube.com/watch?v=lB6s-pGlo04    NA
#> 27 https://www.youtube.com/watch?v=kD6TOpVQ5i0    NA
#> 28 https://www.youtube.com/watch?v=j1z8dudbN9Y    NA
#> 29 https://www.youtube.com/watch?v=hTluPNefjSQ    NA
#> 30 https://www.youtube.com/watch?v=0SaqgglJe_E    NA
#> 31 https://www.youtube.com/watch?v=G5fwDmKe9cg    NA
#> 32 https://www.youtube.com/watch?v=2uMx20-ZiOE    NA
#> 33 https://www.youtube.com/watch?v=FC5iJjIe1X0    NA
#> 34 https://www.youtube.com/watch?v=2WuHl4W40Nw    NA
#> 35 https://www.youtube.com/watch?v=7oI3YL-nv_k    NA
#> 36 https://www.youtube.com/watch?v=XA6HmGetYJs    NA
#> 37 https://www.youtube.com/watch?v=n2MfSy_7yOU    NA
#> 38 https://www.youtube.com/watch?v=Rtj-zNHEd0c    NA
#> 39 https://www.youtube.com/watch?v=gM086JZoh68    NA
#> 40                                        <NA>    NA
#> 41 https://www.youtube.com/watch?v=Os0jOlTmAP4    NA
#> 42 https://www.youtube.com/watch?v=9RBQOKEFC0c    NA
#> 43                                        <NA>    NA
#> 44 https://www.youtube.com/watch?v=aMSClbb0CXw    NA
#> 45 https://www.youtube.com/watch?v=TBzMy1k0cSA    NA
#> 46 https://www.youtube.com/watch?v=wASVliw-Jrg    NA
#> 47 https://www.youtube.com/watch?v=PeUrYFpqo_A    NA
#> 48 https://www.youtube.com/watch?v=mn8JsJKjexY    NA
#> 49 https://www.youtube.com/watch?v=VuzsUbOUlC0    NA
#> 50 https://www.youtube.com/watch?v=gJr3StoaNTM    NA
#> 51 https://www.youtube.com/watch?v=2FHV0wgV0CE    NA
#> 52 https://www.youtube.com/watch?v=-YUtG0OfcLs    NA
#> 53 https://www.youtube.com/watch?v=5XJe-85pzBM    NA
#> 54                                        <NA>    NA
#> 55                                        <NA>    NA
#> 56 https://www.youtube.com/watch?v=FRnrJ3x_Z0Y    NA
#> 57 https://www.youtube.com/watch?v=A_OJM2hCXbc    NA
#> 58 https://www.youtube.com/watch?v=gJt4PIvFsxE    NA
#> 59 https://www.youtube.com/watch?v=dm_U31yHgH0    NA
#> 60                                        <NA>    NA
#> 61 https://www.youtube.com/watch?v=sC1A1nORhXo    NA
#> 62 https://www.youtube.com/watch?v=CLDKcBwCeEg    NA
#> 63 https://www.youtube.com/watch?v=gMbrkMbG29c    NA
#> 64 https://www.youtube.com/watch?v=c72cGNYWKhw    NA
#> 65 https://www.youtube.com/watch?v=vgHOuI_-Zz0    NA
#> 66 https://www.youtube.com/watch?v=zUuVo8okkTw    NA
#> 67 https://www.youtube.com/watch?v=YBrgW4EWIkQ    NA
#> 68                                        <NA>    NA
#> 69                                        <NA>    NA
#> 70                                        <NA>    NA
#> 71                                        <NA>    NA
#> 72 https://www.youtube.com/watch?v=4zlLX9-1Sqc    NA
#> 73                                        <NA>    NA
#> 74 https://www.youtube.com/watch?v=reeW5x_QkMU    NA
#> 75 https://www.youtube.com/watch?v=OqsR5HaF5Cg    NA
#> 76                                        <NA>    NA
#> 77 https://www.youtube.com/watch?v=2UMbpbcZuWY    NA
#> 78 https://www.youtube.com/watch?v=v1mQhDaVItY    NA
#> 79                                        <NA>    NA
#> 80 https://www.youtube.com/watch?v=ln8zA5g1xPo    NA
#> 81 https://www.youtube.com/watch?v=blF1MPQ5aaw    NA
#> 82 https://www.youtube.com/watch?v=Ez5o-Ng2KMA    NA
#> 83 https://www.youtube.com/watch?v=0DqaoprezwQ    NA
#> 84 https://www.youtube.com/watch?v=_no6VO2XeGI    NA
#> 85 https://www.youtube.com/watch?v=s880LSIAtMA    NA
#> 86 https://www.youtube.com/watch?v=6i-7cmf6uCU    NA
#> 87 https://www.youtube.com/watch?v=SxZrYfJeEG8    NA
#> 88 https://www.youtube.com/watch?v=xm_amg8qlrA    NA

  cfbd_game_info(2018, week = 7, conference = "Ind")
#>     game_id season week season_type               start_date start_time_tbd
#> 1 401013452   2018    7     regular 2018-10-13T18:00:00.000Z             NA
#> 2 401013148   2018    7     regular 2018-10-13T18:30:00.000Z             NA
#> 3 401013370   2018    7     regular 2018-10-13T19:30:00.000Z             NA
#> 4 401013442   2018    7     regular 2018-10-13T21:00:00.000Z             NA
#> 5 401016408   2018    7     regular 2018-10-14T02:15:00.000Z             NA
#>   neutral_site conference_game attendance venue_id                  venue
#> 1        FALSE           FALSE          0     3992       Williams Stadium
#> 2        FALSE           FALSE          0     3855     Notre Dame Stadium
#> 3         TRUE           FALSE          0     4738         Levi's Stadium
#> 4        FALSE           FALSE          0     3666            Cajun Field
#> 5        FALSE           FALSE          0     3801 LaVell Edwards Stadium
#>   home_id      home_team  home_conference home_points    home_post_win_prob
#> 1    2335        Liberty FBS Independents          22    0.5483899675354477
#> 2      87     Notre Dame FBS Independents          19    0.9661277906146866
#> 3      23 San José State    Mountain West           3 0.0036361807521906077
#> 4     309      Louisiana         Sun Belt          66    0.9983063510069449
#> 5     252            BYU FBS Independents          49    0.9978157683814658
#>   home_pregame_elo home_postgame_elo away_id        away_team  away_conference
#> 1             1336              1354    2653             Troy         Sun Belt
#> 2             1869              1858     221       Pittsburgh              ACC
#> 3             1041               978     349             Army FBS Independents
#> 4             1194              1259     166 New Mexico State FBS Independents
#> 5             1377              1440      62          Hawai'i    Mountain West
#>   away_points    away_post_win_prob away_pregame_elo away_postgame_elo
#> 1          16   0.45161003246455234             1548              1530
#> 2          14  0.033872209385313434             1444              1455
#> 3          52    0.9963638192478094             1628              1691
#> 4          38 0.0016936489930551302             1126              1061
#> 5          23 0.0021842316185342048             1351              1288
#>   excitement_index                                  highlights notes
#> 1     8.1162251011 https://www.youtube.com/watch?v=tC5LVoZMze4    NA
#> 2     7.2536238281 https://www.youtube.com/watch?v=Aij-ogHy23U    NA
#> 3     2.2159399162 https://www.youtube.com/watch?v=q3_V-NR-R2g    NA
#> 4 2.58899975529999                                        <NA>    NA
#> 5 1.08305180749999                                        <NA>    NA

  # 7 OTs LSU @ TAMU
  cfbd_game_info(2018, week = 13, team = "Texas A&M", quarter_scores = TRUE)
#>     game_id season week season_type               start_date start_time_tbd
#> 1 401012356   2018   13     regular 2018-11-25T00:30:00.000Z             NA
#>   neutral_site conference_game attendance venue_id      venue home_id home_team
#> 1        FALSE            TRUE          0     3795 Kyle Field     245 Texas A&M
#>   home_conference home_points home_scores_Q1 home_scores_Q2 home_scores_Q3
#> 1             SEC          74              7             10              7
#>   home_scores_Q4 home_scores_Q5 home_scores_Q6 home_scores_Q7 home_scores_Q8
#> 1              7              3              7              8              3
#>   home_scores_Q9 home_scores_Q10 home_scores_Q11 home_post_win_prob
#> 1              6               8               8 0.6786258156883023
#>   home_pregame_elo home_postgame_elo away_id away_team away_conference
#> 1             1722              1724      99       LSU             SEC
#>   away_points away_scores_Q1 away_scores_Q2 away_scores_Q3 away_scores_Q4
#> 1          72              7              3              7             14
#>   away_scores_Q5 away_scores_Q6 away_scores_Q7 away_scores_Q8 away_scores_Q9
#> 1              3              7              8              3              6
#>   away_scores_Q10 away_scores_Q11  away_post_win_prob away_pregame_elo
#> 1               8               6 0.32137418431169773             1836
#>   away_postgame_elo excitement_index
#> 1              1834    14.1962593954
#>                                    highlights notes
#> 1 https://www.youtube.com/watch?v=36oYtGl-n-s    NA
# }