Get player statistics by game

Get player statistics by game

cfbd_game_player_stats(
  year,
  week = NULL,
  season_type = "regular",
  team = NULL,
  conference = NULL,
  category = NULL,
  game_id = NULL,
  verbose = FALSE
)

Arguments

year
(Integer required): Year, 4 digit format(YYYY)

week
(Integer optional): Week - values from 1-15, 1-14 for seasons pre-playoff (i.e. 2013 or earlier)

season_type
(String default regular): Select Season Type: regular or postseason

team
(String optional): D-I Team

conference
(String optional): Conference abbreviation - Select a valid FBS conference
Conference abbreviations P5: ACC, B12, B1G, SEC, PAC
Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC

category
(String optional): Category filter (e.g defensive)
Offense: passing, receiving, rushing
Defense: defensive, fumbles, interceptions
Special Teams: punting, puntReturns, kicking, kickReturns

game_id
(Integer optional): Game ID filter for querying a single game Can be found using the cfbd_game_info() function

verbose
Logical parameter (TRUE/FALSE, default: FALSE) to return warnings and messages from function

Source

https://api.collegefootballdata.com/games/players

Value

cfbd_game_player_stats() - A data frame with 32 variables:

game_id: integer.

Referencing game id.

team: character.

Team name.

conference: character.

Conference of the team.

home_away: character.

Flag for if the team was the home or away team.

points: integer.

Team points.

category: character.

Statistic category.

athlete_id: character.

Athlete referencing id.

name: character.

Player name.

c_att: character.

Completions - Pass attempts.

yds: double.

Statistic yardage.

avg: double.

Average per statistic.

td: double.

Touchdowns scored.

int: double.

Interceptions thrown.

qbr: double.

Quarterback rating.

car: double.

Number of rushing carries.

long: double.

Longest carry/reception of the game.

rec: double.

Number of pass receptions.

no: double.

Player number.

fg: character.

Field goal attempts in the game.

pct: double.

Field goal percentage in the game.

xp: character.

Extra points kicked in the game.

pts: double.

Total kicking points in the game.

tb: double.

Touchbacks (for kicking) in the game.

in_20: double.

Punts inside the 20 yardline in the game.

fum: double.

Player fumbles in the game.

lost: double.

Player fumbles lost in the game.

tot: double.

Total tackles in the game.

solo: double.

Solo tackles in the game.

sacks: double.

Total sacks in the game.

tfl: double.

Total tackles for loss in the game.

pd: double.

Total passes defensed in the game.

qb_hur: double.

Total quarterback hurries in the game.

Examples

# \donttest{
  cfbd_game_player_stats(2018, week = 15, conference = "Ind")

#>      game_id team        conference home_away points      category athlete_id
#> 1  401013376 Navy American Athletic      away     10       passing    3916282
#> 2  401013376 Navy American Athletic      away     10       passing    3916251
#> 3  401013376 Navy American Athletic      away     10       passing    4039425
#> 4  401013376 Navy American Athletic      away     10       rushing    4039436
#> 5  401013376 Navy American Athletic      away     10       rushing    3916268
#> 6  401013376 Navy American Athletic      away     10       rushing    3916282
#> 7  401013376 Navy American Athletic      away     10       rushing    3916305
#> 8  401013376 Navy American Athletic      away     10       rushing    4243731
#> 9  401013376 Navy American Athletic      away     10       rushing    4243713
#> 10 401013376 Navy American Athletic      away     10       rushing    3916283
#> 11 401013376 Navy American Athletic      away     10       rushing    4039418
#> 12 401013376 Navy American Athletic      away     10       rushing    3916251
#> 13 401013376 Navy American Athletic      away     10       rushing    4371891
#> 14 401013376 Navy American Athletic      away     10     receiving    4259215
#> 15 401013376 Navy American Athletic      away     10     receiving    4039436
#> 16 401013376 Navy American Athletic      away     10     receiving    3916278
#> 17 401013376 Navy American Athletic      away     10   kickReturns    4039436
#> 18 401013376 Navy American Athletic      away     10   kickReturns    4039418
#> 19 401013376 Navy American Athletic      away     10   puntReturns    4243727
#> 20 401013376 Navy American Athletic      away     10       kicking    3916285
#> 21 401013376 Navy American Athletic      away     10       punting    4039465
#> 22 401013376 Navy American Athletic      away     10       fumbles    3916251
#> 23 401013376 Navy American Athletic      away     10       fumbles    3916282
#> 24 401013376 Navy American Athletic      away     10     defensive    3916276
#> 25 401013376 Navy American Athletic      away     10     defensive    3916271
#> 26 401013376 Navy American Athletic      away     10     defensive    3916290
#> 27 401013376 Navy American Athletic      away     10     defensive    3916296
#> 28 401013376 Navy American Athletic      away     10     defensive    2267927
#> 29 401013376 Navy American Athletic      away     10     defensive    4039448
#> 30 401013376 Navy American Athletic      away     10     defensive    3916303
#> 31 401013376 Navy American Athletic      away     10     defensive    4039423
#> 32 401013376 Navy American Athletic      away     10     defensive    4039466
#> 33 401013376 Navy American Athletic      away     10     defensive    4371866
#> 34 401013376 Navy American Athletic      away     10     defensive    3916293
#> 35 401013376 Navy American Athletic      away     10     defensive    3916309
#> 36 401013376 Navy American Athletic      away     10     defensive    3916268
#> 37 401013376 Navy American Athletic      away     10     defensive    4039416
#> 38 401013376 Navy American Athletic      away     10     defensive    4039447
#> 39 401013376 Navy American Athletic      away     10     defensive    4243753
#> 40 401013376 Army  FBS Independents      home     17       passing    4036923
#> 41 401013376 Army  FBS Independents      home     17       rushing    4036923
#> 42 401013376 Army  FBS Independents      home     17       rushing    3916803
#> 43 401013376 Army  FBS Independents      home     17       rushing    4036942
#> 44 401013376 Army  FBS Independents      home     17       rushing    3916821
#> 45 401013376 Army  FBS Independents      home     17       rushing    4036934
#> 46 401013376 Army  FBS Independents      home     17       rushing    4036949
#> 47 401013376 Army  FBS Independents      home     17       rushing    3916819
#> 48 401013376 Army  FBS Independents      home     17       rushing      -5258
#> 49 401013376 Army  FBS Independents      home     17     receiving    4036942
#> 50 401013376 Army  FBS Independents      home     17     receiving    4036969
#> 51 401013376 Army  FBS Independents      home     17     receiving    4258460
#> 52 401013376 Army  FBS Independents      home     17 interceptions    4036925
#> 53 401013376 Army  FBS Independents      home     17 interceptions    3916807
#> 54 401013376 Army  FBS Independents      home     17   kickReturns    4036942
#> 55 401013376 Army  FBS Independents      home     17   kickReturns    4258467
#> 56 401013376 Army  FBS Independents      home     17   puntReturns    3916807
#> 57 401013376 Army  FBS Independents      home     17       kicking    4077517
#> 58 401013376 Army  FBS Independents      home     17       punting    3916795
#> 59 401013376 Army  FBS Independents      home     17       fumbles    3916832
#> 60 401013376 Army  FBS Independents      home     17       fumbles    4260811
#> 61 401013376 Army  FBS Independents      home     17     defensive    4036925
#> 62 401013376 Army  FBS Independents      home     17     defensive    3916813
#> 63 401013376 Army  FBS Independents      home     17     defensive    3916860
#> 64 401013376 Army  FBS Independents      home     17     defensive    4036958
#> 65 401013376 Army  FBS Independents      home     17     defensive    4036959
#> 66 401013376 Army  FBS Independents      home     17     defensive    3916818
#> 67 401013376 Army  FBS Independents      home     17     defensive    4260811
#> 68 401013376 Army  FBS Independents      home     17     defensive    3916832
#> 69 401013376 Army  FBS Independents      home     17     defensive    4036924
#> 70 401013376 Army  FBS Independents      home     17     defensive    3916850
#> 71 401013376 Army  FBS Independents      home     17     defensive    4036944
#> 72 401013376 Army  FBS Independents      home     17     defensive    4258506
#> 73 401013376 Army  FBS Independents      home     17     defensive    3916819
#> 74 401013376 Army  FBS Independents      home     17     defensive    4036929
#> 75 401013376 Army  FBS Independents      home     17     defensive    4258457
#> 76 401013376 Army  FBS Independents      home     17     defensive    4258521
#> 77 401013376 Army  FBS Independents      home     17     defensive    3916807
#> 78 401013376 Army  FBS Independents      home     17     defensive    4036953
#>                     name c_att yds  avg td int qbr car long rec no   fg pct
#> 1           Garret Lewis  5/11  81  7.4  0   0  NA  NA   NA  NA NA <NA>  NA
#> 2              Zach Abey   0/5   0  0.0  0   2  NA  NA   NA  NA NA <NA>  NA
#> 3               OJ Davis   0/1   0  0.0  0   0  NA  NA   NA  NA NA <NA>  NA
#> 4          Malcolm Perry  <NA>  52  8.7  0  NA  NA   6   43  NA NA <NA>  NA
#> 5       Anthony Gargiulo  <NA>  38  4.8  0  NA  NA   8    7  NA NA <NA>  NA
#> 6           Garret Lewis  <NA>  14  2.3  1  NA  NA   6   13  NA NA <NA>  NA
#> 7             Tre Walker  <NA>  10  3.3  0  NA  NA   3    6  NA NA <NA>  NA
#> 8           Nelson Smith  <NA>   9  2.3  0  NA  NA   4    4  NA NA <NA>  NA
#> 9            Myles Fells  <NA>   5  5.0  0  NA  NA   1    5  NA NA <NA>  NA
#> 10           Mike Martin  <NA>   2  2.0  0  NA  NA   1    2  NA NA <NA>  NA
#> 11            Tazh Maloy  <NA>   2  2.0  0  NA  NA   1    2  NA NA <NA>  NA
#> 12             Zach Abey  <NA>   0  2.0  0  NA  NA   6    4  NA NA <NA>  NA
#> 13         Mychal Cooper  <NA>  -5 -5.0  0  NA  NA   1    0  NA NA <NA>  NA
#> 14 Keoni-Kordell Makekau  <NA>  52 26.0  0  NA  NA  NA   34   2 NA <NA>  NA
#> 15         Malcolm Perry  <NA>  18  9.0  0  NA  NA  NA   13   2 NA <NA>  NA
#> 16        Taylor Jackson  <NA>  11 11.0  0  NA  NA  NA   11   1 NA <NA>  NA
#> 17         Malcolm Perry  <NA>  49 24.5  0  NA  NA  NA   33  NA  2 <NA>  NA
#> 18            Tazh Maloy  <NA>  15 15.0  0  NA  NA  NA   15  NA  1 <NA>  NA
#> 19         Jake Springer  <NA>   8  4.0  0  NA  NA  NA    4  NA  2 <NA>  NA
#> 20      Bennett Moehring  <NA>  NA   NA NA  NA  NA  NA   46  NA NA  1/1 100
#> 21            Owen White  <NA> 145 36.3 NA  NA  NA  NA   43  NA  4 <NA>  NA
#> 22             Zach Abey  <NA>  NA   NA NA  NA  NA  NA   NA   0 NA <NA>  NA
#> 23          Garret Lewis  <NA>  NA   NA NA  NA  NA  NA   NA   0 NA <NA>  NA
#> 24         Taylor Heflin  <NA>  NA   NA  0  NA  NA  NA   NA  NA NA <NA>  NA
#> 25           Juan Hailey  <NA>  NA   NA  0  NA  NA  NA   NA  NA NA <NA>  NA
#> 26           Jarvis Polu  <NA>  NA   NA  0  NA  NA  NA   NA  NA NA <NA>  NA
#> 27       Hudson Sullivan  <NA>  NA   NA  0  NA  NA  NA   NA  NA NA <NA>  NA
#> 28         Sean Williams  <NA>  NA   NA  0  NA  NA  NA   NA  NA NA <NA>  NA
#> 29     Nizaire Cromartie  <NA>  NA   NA  0  NA  NA  NA   NA  NA NA <NA>  NA
#> 30    Anthony Villalobos  <NA>  NA   NA  0  NA  NA  NA   NA  NA NA <NA>  NA
#> 31             Elan Nash  <NA>  NA   NA  0  NA  NA  NA   NA  NA NA <NA>  NA
#> 32       Jackson Pittman  <NA>  NA   NA  0  NA  NA  NA   NA  NA NA <NA>  NA
#> 33           Diego Fagot  <NA>  NA   NA  0  NA  NA  NA   NA  NA NA <NA>  NA
#> 34            Jarid Ryan  <NA>  NA   NA  0  NA  NA  NA   NA  NA NA <NA>  NA
#> 35      Khaylan Williams  <NA>  NA   NA  0  NA  NA  NA   NA  NA NA <NA>  NA
#> 36      Anthony Gargiulo  <NA>  NA   NA  0  NA  NA  NA   NA  NA NA <NA>  NA
#> 37         Noruwa Obanor  <NA>  NA   NA  0  NA  NA  NA   NA  NA NA <NA>  NA
#> 38       Carter Bankston  <NA>  NA   NA  0  NA  NA  NA   NA  NA NA <NA>  NA
#> 39         Pepe Tanuvasa  <NA>  NA   NA  0  NA  NA  NA   NA  NA NA <NA>  NA
#> 40    Kelvin Hopkins Jr.   4/9  61  6.8  0   0  NA  NA   NA  NA NA <NA>  NA
#> 41    Kelvin Hopkins Jr.  <NA>  64  3.6  2  NA  NA  18   17  NA NA <NA>  NA
#> 42      Darnell Woolfolk  <NA>  62  3.9  0  NA  NA  16   18  NA NA <NA>  NA
#> 43           Kell Walker  <NA>  56 11.2  0  NA  NA   5   51  NA NA <NA>  NA
#> 44         Andy Davidson  <NA>  31  5.2  0  NA  NA   6   22  NA NA <NA>  NA
#> 45         Malik Hancock  <NA>   8  8.0  0  NA  NA   1    8  NA NA <NA>  NA
#> 46         Connor Slomka  <NA>   1  1.0  0  NA  NA   1    1  NA NA <NA>  NA
#> 47        Jordan Asberry  <NA>   1  0.5  0  NA  NA   2    2  NA NA <NA>  NA
#> 48                  Team  <NA>  -1 -1.0  0  NA  NA   1    0  NA NA <NA>  NA
#> 49           Kell Walker  <NA>  30 15.0  0  NA  NA  NA   25   2 NA <NA>  NA
#> 50           Glen Coates  <NA>  17 17.0  0  NA  NA  NA   17   1 NA <NA>  NA
#> 51           Zack Boobas  <NA>  14 14.0  0  NA  NA  NA   14   1 NA <NA>  NA
#> 52      Jaylon McClinton  <NA>  11   NA  0   1  NA  NA   NA  NA NA <NA>  NA
#> 53         Mike Reynolds  <NA>   0   NA  0   1  NA  NA   NA  NA NA <NA>  NA
#> 54           Kell Walker  <NA>  23 23.0  0  NA  NA  NA   23  NA  1 <NA>  NA
#> 55          Sandon McCoy  <NA>  15 15.0  0  NA  NA  NA   15  NA  1 <NA>  NA
#> 56         Mike Reynolds  <NA>   4  2.0  0  NA  NA  NA    3  NA  2 <NA>  NA
#> 57      John Abercrombie  <NA>  NA   NA NA  NA  NA  NA   33  NA NA  1/2  50
#> 58          Nick Schrage  <NA> 156 39.0 NA  NA  NA  NA   47  NA  4 <NA>  NA
#> 59       Kenneth Brinson  <NA>  NA   NA NA  NA  NA  NA   NA   1 NA <NA>  NA
#> 60          James Gibson  <NA>  NA   NA NA  NA  NA  NA   NA   1 NA <NA>  NA
#> 61      Jaylon McClinton  <NA>  NA   NA  0  NA  NA  NA   NA  NA NA <NA>  NA
#> 62       James Nachtigal  <NA>  NA   NA  0  NA  NA  NA   NA  NA NA <NA>  NA
#> 63         Wunmi Oyetuga  <NA>  NA   NA  0  NA  NA  NA   NA  NA NA <NA>  NA
#> 64           Amadeo West  <NA>  NA   NA  0  NA  NA  NA   NA  NA NA <NA>  NA
#> 65     Cole Christiansen  <NA>  NA   NA  0  NA  NA  NA   NA  NA NA <NA>  NA
#> 66      Chandler Ramirez  <NA>  NA   NA  0  NA  NA  NA   NA  NA NA <NA>  NA
#> 67          James Gibson  <NA>  NA   NA  0  NA  NA  NA   NA  NA NA <NA>  NA
#> 68       Kenneth Brinson  <NA>  NA   NA  0  NA  NA  NA   NA  NA NA <NA>  NA
#> 69          Elijah Riley  <NA>  NA   NA  0  NA  NA  NA   NA  NA NA <NA>  NA
#> 70        Raymond Wright  <NA>  NA   NA  0  NA  NA  NA   NA  NA NA <NA>  NA
#> 71         Donavan Lynch  <NA>  NA   NA  0  NA  NA  NA   NA  NA NA <NA>  NA
#> 72     Edriece Patterson  <NA>  NA   NA  0  NA  NA  NA   NA  NA NA <NA>  NA
#> 73        Jordan Asberry  <NA>  NA   NA  0  NA  NA  NA   NA  NA NA <NA>  NA
#> 74           Ryan Parker  <NA>  NA   NA  0  NA  NA  NA   NA  NA NA <NA>  NA
#> 75      Javhari Bourdeau  <NA>  NA   NA  0  NA  NA  NA   NA  NA NA <NA>  NA
#> 76        Landon Salyers  <NA>  NA   NA  0  NA  NA  NA   NA  NA NA <NA>  NA
#> 77         Mike Reynolds  <NA>  NA   NA  0  NA  NA  NA   NA  NA NA <NA>  NA
#> 78       Jacob Covington  <NA>  NA   NA  0  NA  NA  NA   NA  NA NA <NA>  NA
#>      xp pts tb in_20 fum lost tot solo sacks tfl pd qb_hur
#> 1  <NA>  NA NA    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 2  <NA>  NA NA    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 3  <NA>  NA NA    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 4  <NA>  NA NA    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 5  <NA>  NA NA    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 6  <NA>  NA NA    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 7  <NA>  NA NA    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 8  <NA>  NA NA    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 9  <NA>  NA NA    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 10 <NA>  NA NA    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 11 <NA>  NA NA    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 12 <NA>  NA NA    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 13 <NA>  NA NA    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 14 <NA>  NA NA    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 15 <NA>  NA NA    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 16 <NA>  NA NA    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 17 <NA>  NA NA    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 18 <NA>  NA NA    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 19 <NA>  NA NA    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 20  1/1   4 NA    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 21 <NA>  NA  0     0  NA   NA  NA   NA    NA  NA NA     NA
#> 22 <NA>  NA NA    NA   1    1  NA   NA    NA  NA NA     NA
#> 23 <NA>  NA NA    NA   1    1  NA   NA    NA  NA NA     NA
#> 24 <NA>  NA NA    NA  NA   NA  14    4   0.0 2.5  0      0
#> 25 <NA>  NA NA    NA  NA   NA  13    7   0.0 0.5  0      0
#> 26 <NA>  NA NA    NA  NA   NA   9    3   0.0 0.0  0      0
#> 27 <NA>  NA NA    NA  NA   NA   9    2   0.5 0.5  0      0
#> 28 <NA>  NA NA    NA  NA   NA   8    5   0.0 1.0  1      0
#> 29 <NA>  NA NA    NA  NA   NA   5    2   0.5 0.5  0      0
#> 30 <NA>  NA NA    NA  NA   NA   5    1   0.0 0.0  0      0
#> 31 <NA>  NA NA    NA  NA   NA   5    1   0.0 0.0  0      0
#> 32 <NA>  NA NA    NA  NA   NA   3    1   0.0 0.0  0      0
#> 33 <NA>  NA NA    NA  NA   NA   2    2   0.0 0.0  0      0
#> 34 <NA>  NA NA    NA  NA   NA   2    1   0.0 0.0  0      0
#> 35 <NA>  NA NA    NA  NA   NA   2    0   0.0 0.0  0      0
#> 36 <NA>  NA NA    NA  NA   NA   1    1   0.0 0.0  0      0
#> 37 <NA>  NA NA    NA  NA   NA   1    1   0.0 0.0  0      0
#> 38 <NA>  NA NA    NA  NA   NA   1    0   0.0 0.0  0      0
#> 39 <NA>  NA NA    NA  NA   NA   1    0   0.0 0.0  0      0
#> 40 <NA>  NA NA    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 41 <NA>  NA NA    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 42 <NA>  NA NA    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 43 <NA>  NA NA    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 44 <NA>  NA NA    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 45 <NA>  NA NA    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 46 <NA>  NA NA    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 47 <NA>  NA NA    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 48 <NA>  NA NA    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 49 <NA>  NA NA    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 50 <NA>  NA NA    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 51 <NA>  NA NA    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 52 <NA>  NA NA    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 53 <NA>  NA NA    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 54 <NA>  NA NA    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 55 <NA>  NA NA    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 56 <NA>  NA NA    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 57  2/2   5 NA    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 58 <NA>  NA  0     1  NA   NA  NA   NA    NA  NA NA     NA
#> 59 <NA>  NA NA    NA   0    0  NA   NA    NA  NA NA     NA
#> 60 <NA>  NA NA    NA   0    0  NA   NA    NA  NA NA     NA
#> 61 <NA>  NA NA    NA  NA   NA   9    6   0.0 1.0  0      0
#> 62 <NA>  NA NA    NA  NA   NA   8    3   0.0 0.0  1      0
#> 63 <NA>  NA NA    NA  NA   NA   7    1   0.0 0.5  0      0
#> 64 <NA>  NA NA    NA  NA   NA   6    3   0.0 0.0  0      0
#> 65 <NA>  NA NA    NA  NA   NA   6    1   0.0 1.5  0      0
#> 66 <NA>  NA NA    NA  NA   NA   5    3   0.0 0.0  0      0
#> 67 <NA>  NA NA    NA  NA   NA   4    1   0.0 0.0  0      0
#> 68 <NA>  NA NA    NA  NA   NA   3    2   1.0 1.0  0      0
#> 69 <NA>  NA NA    NA  NA   NA   3    1   0.0 1.5  2      0
#> 70 <NA>  NA NA    NA  NA   NA   3    0   0.0 0.0  0      0
#> 71 <NA>  NA NA    NA  NA   NA   2    1   0.0 0.0  0      0
#> 72 <NA>  NA NA    NA  NA   NA   2    0   0.0 0.5  0      0
#> 73 <NA>  NA NA    NA  NA   NA   1    1   0.0 0.0  0      0
#> 74 <NA>  NA NA    NA  NA   NA   1    1   0.0 0.0  0      0
#> 75 <NA>  NA NA    NA  NA   NA   1    1   0.0 0.0  0      0
#> 76 <NA>  NA NA    NA  NA   NA   1    1   0.0 0.0  0      0
#> 77 <NA>  NA NA    NA  NA   NA   1    0   0.0 0.0  3      0
#> 78 <NA>  NA NA    NA  NA   NA   1    0   0.0 0.0  0      0

  cfbd_game_player_stats(2013, week = 1, team = "Florida State", category = "passing")

#>     game_id          team conference home_away points category athlete_id
#> 1 332450221    Pittsburgh        ACC      home     13  passing     514116
#> 2 332450221 Florida State        ACC      away     41  passing     530308
#> 3 332450221 Florida State        ACC      away     41  passing     514124
#>             name c_att yds  avg td int  qbr car long rec no fg pct xp pts tb
#> 1     Tom Savage 15/28 201  7.2  1   2 20.3  NA   NA  NA NA NA  NA NA  NA NA
#> 2 Jameis Winston 25/27 356 13.2  4   0 97.3  NA   NA  NA NA NA  NA NA  NA NA
#> 3     Jake Coker   2/2  21 10.5  0   0 99.1  NA   NA  NA NA NA  NA NA  NA NA
#>   in_20 fum lost tot solo sacks tfl pd qb_hur
#> 1    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 2    NA  NA   NA  NA   NA    NA  NA NA     NA
#> 3    NA  NA   NA  NA   NA    NA  NA NA     NA

# }