Get player statistics by game

Get player statistics by game

cfbd_game_player_stats(
  year,
  week = NULL,
  season_type = "regular",
  team = NULL,
  conference = NULL,
  category = NULL,
  game_id = NULL,
  verbose = FALSE
)

Arguments

year
(Integer required): Year, 4 digit format(YYYY)

week
(Integer optional): Week - values from 1-15, 1-14 for seasons pre-playoff (i.e. 2013 or earlier)

season_type
(String default regular): Select Season Type: regular or postseason

team
(String optional): D-I Team

conference
(String optional): Conference abbreviation - Select a valid FBS conference
Conference abbreviations P5: ACC, B12, B1G, SEC, PAC
Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC

category
(String optional): Category filter (e.g defensive)
Offense: passing, receiving, rushing
Defense: defensive, fumbles, interceptions
Special Teams: punting, puntReturns, kicking, kickReturns

game_id
(Integer optional): Game ID filter for querying a single game Can be found using the cfbd_game_info() function

verbose
Logical parameter (TRUE/FALSE, default: FALSE) to return warnings and messages from function

Source

https://api.collegefootballdata.com/games/players

Value

cfbd_game_player_stats() - A data frame with 32 variables:

game_id: integer.

Referencing game id.

team: character.

Team name.

conference: character.

Conference of the team.

home_away: character.

Flag for if the team was the home or away team.

points: integer.

Team points.

category: character.

Statistic category.

athlete_id: character.

Athlete referencing id.

name: character.

Player name.

c_att: character.

Completions - Pass attempts.

yds: double.

Statistic yardage.

avg: double.

Average per statistic.

td: double.

Touchdowns scored.

int: double.

Interceptions thrown.

qbr: double.

Quarterback rating.

car: double.

Number of rushing carries.

long: double.

Longest carry/reception of the game.

rec: double.

Number of pass receptions.

no: double.

Player number.

fg: character.

Field goal attempts in the game.

pct: double.

Field goal percentage in the game.

xp: character.

Extra points kicked in the game.

pts: double.

Total kicking points in the game.

tb: double.

Touchbacks (for kicking) in the game.

in_20: double.

Punts inside the 20 yardline in the game.

fum: double.

Player fumbles in the game.

lost: double.

Player fumbles lost in the game.

tot: double.

Total tackles in the game.

solo: double.

Solo tackles in the game.

sacks: double.

Total sacks in the game.

tfl: double.

Total tackles for loss in the game.

pd: double.

Total passes defensed in the game.

qb_hur: double.

Total quarterback hurries in the game.

Examples

# \donttest{
  cfbd_game_player_stats(year = 2020, week = 15, team = "Alabama")

#> # A tibble: 65 x 32
#>     game_id team   conference home_away points category athlete_id name    c_att
#>       <int> <chr>  <chr>      <chr>      <int> <chr>    <chr>      <chr>   <lis>
#>  1   4.01e8 Alaba… SEC        away          52 defensi… 4241478    DeVont… <NUL…
#>  2   4.01e8 Alaba… SEC        away          52 fumbles  4685723    Roydel… <NUL…
#>  3   4.01e8 Alaba… SEC        away          52 fumbles  4567140    DJ Dale <NUL…
#>  4   4.01e8 Alaba… SEC        away          52 fumbles  4567123    Byron … <NUL…
#>  5   4.01e8 Alaba… SEC        away          52 fumbles  4372025    Jaylen… <NUL…
#>  6   4.01e8 Alaba… SEC        away          52 punting  4261098    Charli… <NUL…
#>  7   4.01e8 Alaba… SEC        away          52 kicking  4567104    Will R… <NUL…
#>  8   4.01e8 Alaba… SEC        away          52 puntRet… 4241478    DeVont… <NUL…
#>  9   4.01e8 Alaba… SEC        away          52 kickRet… 4241474    Brian … <NUL…
#> 10   4.01e8 Alaba… SEC        away          52 interce… 4692025    Brian … <NUL…
#> # … with 55 more rows, and 23 more variables: yds <list>, avg <list>,
#> #   td <list>, int <list>, qbr <lgl>, car <list>, long <list>, rec <list>,
#> #   no <list>, fg <list>, pct <list>, xp <list>, pts <list>, tb <list>,
#> #   in_20 <list>, fum <list>, lost <list>, tot <list>, solo <list>,
#> #   sacks <list>, tfl <list>, pd <list>, qb_hur <list>

  cfbd_game_player_stats(2013, week = 1, team = "Florida State", category = "passing")

#> # A tibble: 3 x 32
#>    game_id team    conference home_away points category athlete_id name    c_att
#>      <int> <chr>   <chr>      <chr>      <int> <chr>    <chr>      <chr>   <lis>
#> 1   3.32e8 Pittsb… ACC        home          13 passing  514116     Tom Sa… <chr…
#> 2   3.32e8 Florid… ACC        away          41 passing  530308     Jameis… <chr…
#> 3   3.32e8 Florid… ACC        away          41 passing  514124     Jake C… <chr…
#> # … with 23 more variables: yds <list>, avg <list>, td <list>, int <list>,
#> #   qbr <list>, car <lgl>, long <lgl>, rec <lgl>, no <lgl>, fg <lgl>,
#> #   pct <lgl>, xp <lgl>, pts <lgl>, tb <lgl>, in_20 <lgl>, fum <lgl>,
#> #   lost <lgl>, tot <lgl>, solo <lgl>, sacks <lgl>, tfl <lgl>, pd <lgl>,
#> #   qb_hur <lgl>

# }