Get player statistics by game

Get player statistics by game

Usage

cfbd_game_player_stats(
  year,
  week = NULL,
  season_type = "regular",
  team = NULL,
  conference = NULL,
  category = NULL,
  game_id = NULL
)

Arguments

year

(Integer required): Year, 4 digit format(YYYY)

week

(Integer optional): Week - values from 1-15, 1-14 for seasons pre-playoff (i.e. 2013 or earlier)

season_type

(String default regular): Select Season Type: regular or postseason

team

(String optional): D-I Team

conference

(String optional): Conference abbreviation - Select a valid FBS conference Conference abbreviations P5: ACC, B12, B1G, SEC, PAC Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC

category

(String optional): Category filter (e.g defensive) Offense: passing, receiving, rushing Defense: defensive, fumbles, interceptions Special Teams: punting, puntReturns, kicking, kickReturns

game_id

(Integer optional): Game ID filter for querying a single game Can be found using the cfbd_game_info() function

Value

cfbd_game_player_stats() - A data frame with 32 variables:

game_id: integer.

Referencing game id.

team: character.

Team name.

conference: character.

Conference of the team.

home_away: character.

Flag for if the team was the home or away team.

points: integer.

Team points.

category: character.

Statistic category.

athlete_id: character.

Athlete referencing id.

name: character.

Player name.

c_att: character.

Completions - Pass attempts.

yds: double.

Statistic yardage.

avg: double.

Average per statistic.

td: double.

Touchdowns scored.

int: double.

Interceptions thrown.

qbr: double.

Quarterback rating.

car: double.

Number of rushing carries.

long: double.

Longest carry/reception of the game.

rec: double.

Number of pass receptions.

no: double.

Player number.

fg: character.

Field goal attempts in the game.

pct: double.

Field goal percentage in the game.

xp: character.

Extra points kicked in the game.

pts: double.

Total kicking points in the game.

tb: double.

Touchbacks (for kicking) in the game.

in_20: double.

Punts inside the 20 yardline in the game.

fum: double.

Player fumbles in the game.

lost: double.

Player fumbles lost in the game.

tot: double.

Total tackles in the game.

solo: double.

Solo tackles in the game.

sacks: double.

Total sacks in the game.

tfl: double.

Total tackles for loss in the game.

pd: double.

Total passes defensed in the game.

qb_hur: double.

Total quarterback hurries in the game.

Examples

# \donttest{
  try(cfbd_game_player_stats(year = 2020, week = 15, team = "Alabama"))
#> ── Game player stats data from CollegeFootballData.com ─────── cfbfastR 1.9.1 ──
#> ℹ Data updated: 2022-09-04 02:07:23 UTC
#> # A tibble: 65 × 32
#>      game_id team    confer…¹ home_…² points categ…³ athle…⁴ name  c_att  yds   
#>        <int> <chr>   <chr>    <chr>    <int> <chr>   <chr>   <chr> <list> <list>
#>  1 401267164 Alabama SEC      away        52 defens… 4241478 DeVo… <NULL> <NULL>
#>  2 401267164 Alabama SEC      away        52 fumbles 4685723 Royd… <NULL> <NULL>
#>  3 401267164 Alabama SEC      away        52 fumbles 4567140 DJ D… <NULL> <NULL>
#>  4 401267164 Alabama SEC      away        52 fumbles 4567123 Byro… <NULL> <NULL>
#>  5 401267164 Alabama SEC      away        52 fumbles 4372025 Jayl… <NULL> <NULL>
#>  6 401267164 Alabama SEC      away        52 punting 4261098 Char… <NULL> <chr> 
#>  7 401267164 Alabama SEC      away        52 kicking 4567104 Will… <NULL> <NULL>
#>  8 401267164 Alabama SEC      away        52 puntRe… 4241478 DeVo… <NULL> <chr> 
#>  9 401267164 Alabama SEC      away        52 kickRe… 4241474 Bria… <NULL> <chr> 
#> 10 401267164 Alabama SEC      away        52 interc… 4692025 Bria… <NULL> <chr> 
#> # … with 55 more rows, 22 more variables: avg <list>, td <list>, int <list>,
#> #   qbr <lgl>, car <list>, long <list>, rec <list>, no <list>, fg <list>,
#> #   pct <list>, xp <list>, pts <list>, tb <list>, in_20 <list>, fum <list>,
#> #   lost <list>, tot <list>, solo <list>, sacks <list>, tfl <list>, pd <list>,
#> #   qb_hur <list>, and abbreviated variable names ¹​conference, ²​home_away,
#> #   ³​category, ⁴​athlete_id

  try(cfbd_game_player_stats(2013, week = 1, team = "Florida State", category = "passing"))
#> ── Game player stats data from CollegeFootballData.com ─────── cfbfastR 1.9.1 ──
#> ℹ Data updated: 2022-09-04 02:07:24 UTC
#> # A tibble: 3 × 32
#>     game_id team  confe…¹ home_…² points categ…³ athle…⁴ name  c_att yds   avg  
#>       <int> <chr> <chr>   <chr>    <int> <chr>   <chr>   <chr> <lis> <lis> <lis>
#> 1 332450221 Flor… ACC     away        41 passing 530308  Jame… <chr> <chr> <chr>
#> 2 332450221 Flor… ACC     away        41 passing 514124  Jake… <chr> <chr> <chr>
#> 3 332450221 Pitt… ACC     home        13 passing 514116  Tom … <chr> <chr> <chr>
#> # … with 21 more variables: td <list>, int <list>, qbr <list>, car <lgl>,
#> #   long <lgl>, rec <lgl>, no <lgl>, fg <lgl>, pct <lgl>, xp <lgl>, pts <lgl>,
#> #   tb <lgl>, in_20 <lgl>, fum <lgl>, lost <lgl>, tot <lgl>, solo <lgl>,
#> #   sacks <lgl>, tfl <lgl>, pd <lgl>, qb_hur <lgl>, and abbreviated variable
#> #   names ¹​conference, ²​home_away, ³​category, ⁴​athlete_id
# }