Get team statistics by game

Get team statistics by game

cfbd_game_team_stats(
  year,
  week = NULL,
  season_type = "regular",
  team = NULL,
  conference = NULL,
  game_id = NULL,
  rows_per_team = 1,
  verbose = FALSE
)

Arguments

year
(Integer required): Year, 4 digit format (YYYY)

week
(Integer optional): Week - values range from 1-15, 1-14 for seasons pre-playoff, i.e. 2013 or earlier

season_type
(String default: regular): Select Season Type - regular, postseason, or both

team
(String optional): D-I Team

conference
(String optional): Conference abbreviation - Select a valid FBS conference
Conference abbreviations P5: ACC, B12, B1G, SEC, PAC
Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC

game_id
(Integer optional): Game ID filter for querying a single game
Can be found using the cfbd_game_info() function

rows_per_team
(Integer default 1): Both Teams for each game on one or two row(s), Options: 1 or 2

verbose
Logical parameter (TRUE/FALSE, default: FALSE) to return warnings and messages from function

Source

https://api.collegefootballdata.com/games/teams

Value

cfbd_game_team_stats() - A data frame with 78 variables:

game_id: integer.

Referencing game id.

school: character.

Team name.

conference: character.

Conference of the team.

home_away: character.

Home/Away Flag.

points: integer.

Team points.

total_yards: character.

Team total yards.

net_passing_yards: character.

Team net passing yards.

completion_attempts:character.

Team completion attempts.

passing_tds: character.

Team passing touchdowns.

yards_per_pass: character.

Team game yards per pass.

passes_intercepted: character.

Team passes intercepted.

interception_yards: character.

Interception yards.

interception_tds: character.

Interceptions returned for a touchdown.

rushing_attempts: character.

Team rushing attempts. see also: ESTABLISH IT.

rushing_yards: character.

Team rushing yards.

rush_tds: character.

Team rushing touchdowns.

yards_per_rush_attempt: character.

Team yards per rush attempt.

first_downs: character.

First downs earned by the team.

third_down_eff: character.

Third down efficiency.

fourth_down_eff: character.

Fourth down efficiency.

punt_returns: character.

Team punt returns.

punt_return_yards: character.

Team punt return yards.

punt_return_tds: character.

Team punt return touchdowns.

kick_return_yards: character.

Team kick return yards.

kick_return_tds: character.

Team kick return touchdowns.

kick_returns: character.

Team kick returns.

kicking_points: character.

Team points from kicking the ball.

fumbles_recovered: character.

Team fumbles recovered.

fumbles_lost: character.

Team fumbles lost.

total_fumbles: character.

Team total fumbles.

tackles: character.

Team tackles.

tackles_for_loss: character.

Team tackles for a loss.

sacks: character.

Team sacks.

qb_hurries: character.

Team QB hurries.

interceptions: character.

Team interceptions.

passes_deflected: character.

Team passes deflected.

turnovers: character.

Team turnovers.

defensive_tds: character.

Team defensive touchdowns.

total_penalties_yards: character.

Team total penalty yards.

possession_time: character.

Team time of possession.

conference_allowed: character.

Conference of the opponent team.

home_away_allowed: character.

Flag for if the opponent was the home or away team.

points_allowed: integer.

Points for the opponent.

total_yards_allowed: character.

Opponent total yards.

net_passing_yards_allowed: character.

Opponent net passing yards.

completion_attempts_allowed: character.

Oppponent completion attempts.

passing_tds_allowed: character.

Opponent passing TDs.

yards_per_pass_allowed: character.

Opponent yards per pass allowed.

passes_intercepted_allowed: character.

Opponent passes intercepted.

interception_yards_allowed: character.

Opponent interception yards.

interception_tds_allowed: character.

Opponent interception TDs.

rushing_attempts_allowed: character.

Opponent rushing attempts.

rushing_yards_allowed: character.

Opponent rushing yards.

rush_tds_allowed: character.

Opponent rushing touchdownss.

yards_per_rush_attempt_allowed: character.

Opponent rushing yards per attempt.

first_downs_allowed: character.

Opponent first downs.

third_down_eff_allowed: character.

Opponent third down efficiency.

fourth_down_eff_allowed: character.

Opponent fourth down efficiency.

punt_returns_allowed: character.

Opponent punt returns.

punt_return_yards_allowed: character.

Opponent punt return yards.

punt_return_tds_allowed: character.

Opponent punt return touchdowns.

kick_return_yards_allowed: character.

Opponent kick return yards.

kick_return_tds_allowed: character.

Opponent kick return touchdowns.

kick_returns_allowed: character.

Opponent kick returns.

kicking_points_allowed: character.

Opponent points from kicking.

fumbles_recovered_allowed: character.

Opponent fumbles recovered.

fumbles_lost_allowed: character.

Opponent fumbles lost.

total_fumbles_allowed:character.

Opponent total number of fumbles.

tackles_allowed:character.

Opponent tackles.

tackles_for_loss_allowed: character.

Opponent tackles for loss.

sacks_allowed: character.

Opponent sacks.

qb_hurries_allowed: character.

Opponent quarterback hurries.

interceptions_allowed: character.

Opponent interceptions.

passes_deflected_allowed: character.

Opponent passes deflected.

turnovers_allowed: character.

Opponent turnovers.

defensive_tds_allowed: character.

Opponent defensive touchdowns.

total_penalties_yards_allowed: character.

Opponent total penalty yards.

possession_time_allowed: character.

Opponent time of possession.

Examples

# \donttest{
  cfbd_game_team_stats(2019, team = "LSU")

#> # A tibble: 13 x 78
#>      game_id school conference home_away points total_yards net_passing_yards
#>        <int> <chr>  <chr>      <chr>      <int> <chr>       <chr>            
#>  1 401110778 LSU    SEC        away          45 573         471              
#>  2 401110805 LSU    SEC        away          66 599         418              
#>  3 401110828 LSU    SEC        away          36 413         327              
#>  4 401110842 LSU    SEC        away          46 559         393              
#>  5 401110833 LSU    SEC        home          23 508         321              
#>  6 401110850 LSU    SEC        away          58 714         489              
#>  7 401110859 LSU    SEC        home          56 612         352              
#>  8 401110869 LSU    SEC        home          50 553         425              
#>  9 401110725 LSU    SEC        home          55 472         350              
#> 10 401110819 LSU    SEC        home          42 511         293              
#> 11 401110790 LSU    SEC        home          65 610         488              
#> 12 401110813 LSU    SEC        home          42 601         353              
#> 13 401132981 LSU    SEC        home          37 481         349              
#> # … with 71 more variables: completion_attempts <chr>, passing_tds <chr>,
#> #   yards_per_pass <chr>, passes_intercepted <chr>, interception_yards <chr>,
#> #   interception_tds <chr>, rushing_attempts <chr>, rushing_yards <chr>,
#> #   rush_tds <chr>, yards_per_rush_attempt <chr>, first_downs <chr>,
#> #   third_down_eff <chr>, fourth_down_eff <chr>, punt_returns <chr>,
#> #   punt_return_yards <chr>, punt_return_tds <chr>, kick_return_yards <chr>,
#> #   kick_return_tds <chr>, kick_returns <chr>, kicking_points <chr>,
#> #   fumbles_recovered <chr>, fumbles_lost <chr>, total_fumbles <chr>,
#> #   tackles <chr>, tackles_for_loss <chr>, sacks <chr>, qb_hurries <chr>,
#> #   interceptions <chr>, passes_deflected <chr>, turnovers <chr>,
#> #   defensive_tds <chr>, total_penalties_yards <chr>, possession_time <chr>,
#> #   conference_allowed <chr>, home_away_allowed <chr>, points_allowed <int>,
#> #   total_yards_allowed <chr>, net_passing_yards_allowed <chr>,
#> #   completion_attempts_allowed <chr>, passing_tds_allowed <chr>,
#> #   yards_per_pass_allowed <chr>, passes_intercepted_allowed <chr>,
#> #   interception_yards_allowed <chr>, interception_tds_allowed <chr>,
#> #   rushing_attempts_allowed <chr>, rushing_yards_allowed <chr>,
#> #   rush_tds_allowed <chr>, yards_per_rush_attempt_allowed <chr>,
#> #   first_downs_allowed <chr>, third_down_eff_allowed <chr>,
#> #   fourth_down_eff_allowed <chr>, punt_returns_allowed <chr>,
#> #   punt_return_yards_allowed <chr>, punt_return_tds_allowed <chr>,
#> #   kick_return_yards_allowed <chr>, kick_return_tds_allowed <chr>,
#> #   kick_returns_allowed <chr>, kicking_points_allowed <chr>,
#> #   fumbles_recovered_allowed <chr>, fumbles_lost_allowed <chr>,
#> #   total_fumbles_allowed <chr>, tackles_allowed <chr>,
#> #   tackles_for_loss_allowed <chr>, sacks_allowed <chr>,
#> #   qb_hurries_allowed <chr>, interceptions_allowed <chr>,
#> #   passes_deflected_allowed <chr>, turnovers_allowed <chr>,
#> #   defensive_tds_allowed <chr>, total_penalties_yards_allowed <chr>,
#> #   possession_time_allowed <chr>

  cfbd_game_team_stats(2013, team = "Florida State")

#> # A tibble: 13 x 78
#>      game_id school     conference home_away points total_yards net_passing_yar…
#>        <int> <chr>      <chr>      <chr>      <int> <chr>       <chr>           
#>  1 332450221 Florida S… ACC        away          41 533         377             
#>  2 332570052 Florida S… ACC        home          62 617         240             
#>  3 332640052 Florida S… ACC        home          54 492         226             
#>  4 332710103 Florida S… ACC        away          48 489         330             
#>  5 332780052 Florida S… ACC        home          63 614         431             
#>  6 332990052 Florida S… ACC        home          49 566         342             
#>  7 333060052 Florida S… ACC        home          41 517         325             
#>  8 333130154 Florida S… ACC        away          59 296         207             
#>  9 333200052 Florida S… ACC        home          59 523         298             
#> 10 333270052 Florida S… ACC        home          80 645         309             
#> 11 332920228 Florida S… ACC        away          51 565         444             
#> 12 333340057 Florida S… ACC        away          37 456         327             
#> 13 333410052 Florida S… ACC        home          45 569         330             
#> # … with 71 more variables: completion_attempts <chr>, passing_tds <chr>,
#> #   yards_per_pass <chr>, passes_intercepted <chr>, interception_yards <chr>,
#> #   interception_tds <chr>, rushing_attempts <chr>, rushing_yards <chr>,
#> #   rush_tds <chr>, yards_per_rush_attempt <chr>, first_downs <chr>,
#> #   third_down_eff <chr>, fourth_down_eff <chr>, punt_returns <chr>,
#> #   punt_return_yards <chr>, punt_return_tds <chr>, kick_return_yards <chr>,
#> #   kick_return_tds <chr>, kick_returns <chr>, kicking_points <chr>,
#> #   fumbles_recovered <chr>, fumbles_lost <chr>, total_fumbles <lgl>,
#> #   tackles <lgl>, tackles_for_loss <lgl>, sacks <lgl>, qb_hurries <lgl>,
#> #   interceptions <chr>, passes_deflected <lgl>, turnovers <chr>,
#> #   defensive_tds <lgl>, total_penalties_yards <chr>, possession_time <chr>,
#> #   conference_allowed <chr>, home_away_allowed <chr>, points_allowed <int>,
#> #   total_yards_allowed <chr>, net_passing_yards_allowed <chr>,
#> #   completion_attempts_allowed <chr>, passing_tds_allowed <chr>,
#> #   yards_per_pass_allowed <chr>, passes_intercepted_allowed <chr>,
#> #   interception_yards_allowed <chr>, interception_tds_allowed <chr>,
#> #   rushing_attempts_allowed <chr>, rushing_yards_allowed <chr>,
#> #   rush_tds_allowed <chr>, yards_per_rush_attempt_allowed <chr>,
#> #   first_downs_allowed <chr>, third_down_eff_allowed <chr>,
#> #   fourth_down_eff_allowed <chr>, punt_returns_allowed <chr>,
#> #   punt_return_yards_allowed <chr>, punt_return_tds_allowed <chr>,
#> #   kick_return_yards_allowed <chr>, kick_return_tds_allowed <chr>,
#> #   kick_returns_allowed <chr>, kicking_points_allowed <chr>,
#> #   fumbles_recovered_allowed <chr>, fumbles_lost_allowed <chr>,
#> #   total_fumbles_allowed <lgl>, tackles_allowed <lgl>,
#> #   tackles_for_loss_allowed <lgl>, sacks_allowed <lgl>,
#> #   qb_hurries_allowed <lgl>, interceptions_allowed <chr>,
#> #   passes_deflected_allowed <lgl>, turnovers_allowed <chr>,
#> #   defensive_tds_allowed <lgl>, total_penalties_yards_allowed <chr>,
#> #   possession_time_allowed <chr>

# }