CFBD Metrics Endpoint

CFBD Metrics Endpoint

Get team game averages for Predicted Points Added (PPA)

Get team game averages for Predicted Points Added (PPA)

Get player game averages for Predicted Points Added (PPA)

Get player season averages for Predicted Points Added (PPA)

Calculate Predicted Points using Down and Distance

Get team averages for Predicted Points Added (PPA)

Get Pre-game Win Probability Data from API

Get win probability chart data from API

cfbd_metrics_ppa_games(
  year,
  week = NULL,
  team = NULL,
  conference = NULL,
  excl_garbage_time = FALSE
)

cfbd_metrics_ppa_games(
  year,
  week = NULL,
  team = NULL,
  conference = NULL,
  excl_garbage_time = FALSE
)

cfbd_metrics_ppa_players_games(
  year = NULL,
  week = NULL,
  team = NULL,
  position = NULL,
  athlete_id = NULL,
  threshold = NULL,
  excl_garbage_time = FALSE
)

cfbd_metrics_ppa_players_season(
  year = NULL,
  team = NULL,
  conference = NULL,
  position = NULL,
  athlete_id = NULL,
  threshold = NULL,
  excl_garbage_time = FALSE
)

cfbd_metrics_ppa_predicted(down, distance)

cfbd_metrics_ppa_teams(
  year = 2019,
  team = NULL,
  conference = NULL,
  excl_garbage_time = FALSE
)

cfbd_metrics_wp_pregame(
  year = NULL,
  week = NULL,
  team = NULL,
  season_type = "regular"
)

cfbd_metrics_wp(game_id)

Arguments

year
(Integer optional): Year, 4 digit format (YYYY)

week
(Integer optional): Week - values from 1-15, 1-14 for seasons pre-playoff, i.e. 2013 or earlier

team
(String optional): D-I Team

conference
(String optional): Conference name - select a valid FBS conference
Conference names P5: ACC, Big 12, Big Ten, SEC, Pac-12
Conference names G5 and FBS Independents: Conference USA, Mid-American, Mountain West, FBS Independents, American Athletic

excl_garbage_time
(Logical default FALSE): Select whether to exclude Garbage Time (TRUE or FALSE)

position
(string optional): Position abbreviation of the player you are searching for.
Position Group - options include:
* Offense: QB, RB, FB, TE, OL, G, OT, C, WR
* Defense: DB, CB, S, LB, DE, DT, NT, DL
* Special Teams: K, P, LS, PK

athlete_id
(Integer optional): Athlete ID filter for querying a single athlete
Can be found using the cfbfastR::cfbd_player_info() function.

threshold
(Integer optional): Minimum threshold of plays.

down
(Integer required): Down filter

distance
(Integer required): Distance filter

season_type
(String default regular): Select Season Type: regular or postseason

game_id
(Integer required): Game ID filter for querying a single game
Can be found using the cfbfastR::cfbd_game_info() function

Source

https://api.collegefootballdata.com/ppa/players/games

https://api.collegefootballdata.com/ppa/players/season

https://api.collegefootballdata.com/ppa/predicted

https://api.collegefootballdata.com/ppa/teams

https://api.collegefootballdata.com/metrics/wp

Value

A data frame with 18 variables:

game_id

integer.

season

integer.

week

integer.

conference

character.

team

character.

opponent

character.

off_overall

character.

off_passing

character.

off_rushing

character.

off_first_down

character.

off_second_down

character.

off_third_down

character.

def_overall

character.

def_passing

character.

def_rushing

character.

def_first_down

character.

def_second_down

character.

def_third_down

character.

A data frame with 18 variables:

game_id

integer.

season

integer.

week

integer.

conference

character.

team

character.

opponent

character.

off_overall

character.

off_passing

character.

off_rushing

character.

off_first_down

character.

off_second_down

character.

off_third_down

character.

def_overall

character.

def_passing

character.

def_rushing

character.

def_first_down

character.

def_second_down

character.

def_third_down

character.

A data frame with 9 variables:

season

integer.

week

integer.

name

character.

position

character.

team

character.

opponent

character.

avg_PPA_all

double.

avg_PPA_pass

double.

avg_PPA_rush

double.

A data frame with 23 variables:

season

integer.

id

character.

name

character.

position

character.

team

character.

conference

character.

countable_plays

integer.

avg_PPA_all

double.

avg_PPA_pass

double.

avg_PPA_rush

double.

avg_PPA_first_down

double.

avg_PPA_second_down

double.

avg_PPA_third_down

double.

avg_PPA_standard_downs

double.

avg_PPA_passing_downs

double.

total_PPA_all

double.

total_PPA_pass

double.

total_PPA_rush

double.

total_PPA_first_down

double.

total_PPA_second_down

double.

total_PPA_third_down

double.

total_PPA_standard_downs

double.

total_PPA_passing_downs

double.

A data frame with 2 variables:

yard_line

integer.

predicted_points

character.

A data frame with 21 variables:

season

integer.

conference

character.

team

character.

off_overall

character.

off_passing

character.

off_rushing

character.

off_first_down

character.

off_second_down

character.

off_third_down

character.

off_cumulative_total

character.

off_cumulative_passing

character.

off_cumulative_rushing

character.

def_overall

character.

def_passing

character.

def_rushing

character.

def_first_down

character.

def_second_down

character.

def_third_down

character.

def_cumulative_total

character.

def_cumulative_passing

character.

def_cumulative_rushing

character.

A data frame with 9 variables:

season

integer.

season_type

character.

week

integer.

game_id

integer.

home_team

character.

away_team

character.

spread

integer.

home_win_prob

double.

away_win_prob

double.

A data frame with 16 variables:

play_id

character.

play_text

character.

home_id

integer.

home

character.

away_id

integer.

away

character.

spread

character.

home_ball

logical.

home_score

integer.

away_score

integer.

down

integer.

distance

integer.

home_win_prob

character.

away_win_prob

double.

play_number

integer.

yard_line

integer.

Examples

if (FALSE) {
  cfbd_metrics_ppa_games(year = 2019, team = 'TCU')
}
if (FALSE) {
  cfbd_metrics_ppa_games(year = 2019, team = 'TCU')
}
if (FALSE) {
  cfbd_metrics_ppa_players_games(year = 2019,week=3, team = 'TCU')
}
if (FALSE) {
  cfbd_metrics_ppa_players_season(year = 2019, team = 'TCU')
}
if (FALSE) {
  cfbd_metrics_ppa_predicted(down = 1, distance = 10)

  cfbd_metrics_ppa_predicted(down = 3, distance = 10)
}
if (FALSE) {
  cfbd_metrics_ppa_teams(year = 2019, team = 'TCU')
}
if (FALSE) {
  cfbd_metrics_wp_pregame(year = 2019, week=9, team='Texas A&M')
}
if (FALSE) {
  cfbd_metrics_wp(game_id = 401012356)
}