Get team game averages for predicted points added (PPA)

Get team game averages for predicted points added (PPA)

cfbd_metrics_ppa_games(
  year,
  week = NULL,
  team = NULL,
  conference = NULL,
  excl_garbage_time = FALSE
)

Arguments

year

(Integer required): Year, 4 digit format (YYYY)

week

(Integer optional): Week - values range from 1-15, 1-14 for seasons pre-playoff, i.e. 2013 or earlier

team

(String optional): D-I Team

conference

(String optional): Conference abbreviation - Select a valid FBS conference
Conference abbreviations P5: ACC, B12, B1G, SEC, PAC
Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC

excl_garbage_time

(Logical default FALSE): Select whether to exclude Garbage Time (TRUE or FALSE)

Value

cfbd_metrics_ppa_games() - A data frame with 18 variables:

game_id: integer.

Referencing game id.

season: integer.

Season of the game.

week: integer.

Game week of the season.

conference: character.

Conference of the team.

team: character.

Team name.

opponent: character.

Team Opponent.

off_overall: character.

Offense overall predicted points added (PPA).

off_passing: character.

Offense passing predicted points added (PPA).

off_rushing: character.

Offense rushing predicted points added (PPA).

off_first_down: character.

Offense 1st down predicted points added (PPA).

off_second_down: character.

Offense 2nd down predicted points added (PPA).

off_third_down: character.

Offense 3rd down predicted points added (PPA).

def_overall: character.

Defense overall predicted points added (PPA).

def_passing: character.

Defense passing predicted points added (PPA).

def_rushing: character.

Defense rushing predicted points added (PPA).

def_first_down: character.

Defense 1st down predicted points added (PPA).

def_second_down: character.

Defense 2nd down predicted points added (PPA).

def_third_down: character.

Defense 3rd down predicted points added (PPA).

Examples

# \donttest{
cfbd_metrics_ppa_games(year = 2019, team = "TCU")
#> ── PPA data from CollegeFootballData.com ───────────────────── cfbfastR 1.8.1 ──
#> ℹ Data updated: 2022-04-27 01:16:43 UTC
#> # A tibble: 12 × 18
#>      game_id season  week conference team  opponent      off_overall off_passing
#>        <int>  <int> <int> <chr>      <chr> <chr>         <chr>       <chr>      
#>  1 401112129   2019     1 Big 12     TCU   Arkansas-Pin… 0.10893335… 0.03259498…
#>  2 401112130   2019     3 Big 12     TCU   Purdue        0.08433374… -0.4461483…
#>  3 401112131   2019     4 Big 12     TCU   SMU           0.14467693… -0.0623022…
#>  4 401112100   2019     5 Big 12     TCU   Kansas        0.38938935… 0.78655661…
#>  5 401112088   2019     6 Big 12     TCU   Iowa State    0.22675780… 0.28269887…
#>  6 401112109   2019     8 Big 12     TCU   Kansas State  0.16269524… -0.1244139…
#>  7 401112132   2019     9 Big 12     TCU   Texas         0.20698084… 0.48632374…
#>  8 401112127   2019    10 Big 12     TCU   Oklahoma Sta… 0.17236682… -0.0570576…
#>  9 401112081   2019    11 Big 12     TCU   Baylor        -0.0574753… -0.1323649…
#> 10 401112133   2019    12 Big 12     TCU   Texas Tech    0.19928588… 0.34181791…
#> 11 401112120   2019    13 Big 12     TCU   Oklahoma      -0.0261717… -0.2697602…
#> 12 401112134   2019    14 Big 12     TCU   West Virginia 0.01175254… -0.1254802…
#> # … with 10 more variables: off_rushing <chr>, off_first_down <chr>,
#> #   off_second_down <chr>, off_third_down <chr>, def_overall <chr>,
#> #   def_passing <chr>, def_rushing <chr>, def_first_down <chr>,
#> #   def_second_down <chr>, def_third_down <chr>
# }