Get team game averages for predicted points added (PPA)

Get team game averages for predicted points added (PPA)

cfbd_metrics_ppa_games(
  year,
  week = NULL,
  team = NULL,
  conference = NULL,
  excl_garbage_time = FALSE,
  verbose = FALSE
)

Arguments

year
(Integer required): Year, 4 digit format (YYYY)

week
(Integer optional): Week - values range from 1-15, 1-14 for seasons pre-playoff, i.e. 2013 or earlier

team
(String optional): D-I Team

conference
(String optional): Conference abbreviation - Select a valid FBS conference
Conference abbreviations P5: ACC, B12, B1G, SEC, PAC
Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC

excl_garbage_time
(Logical default FALSE): Select whether to exclude Garbage Time (TRUE or FALSE)

verbose
Logical parameter (TRUE/FALSE, default: FALSE) to return warnings and messages from function

Value

cfbd_metrics_ppa_games() - A data frame with 18 variables:

game_id: integer.

Referencing game id.

season: integer.

Season of the game.

week: integer.

Game week of the season.

conference: character.

Conference of the team.

team: character.

Team name.

opponent: character.

Team Opponent.

off_overall: character.

Offense overall predicted points added (PPA).

off_passing: character.

Offense passing predicted points added (PPA).

off_rushing: character.

Offense rushing predicted points added (PPA).

off_first_down: character.

Offense 1st down predicted points added (PPA).

off_second_down: character.

Offense 2nd down predicted points added (PPA).

off_third_down: character.

Offense 3rd down predicted points added (PPA).

def_overall: character.

Defense overall predicted points added (PPA).

def_passing: character.

Defense passing predicted points added (PPA).

def_rushing: character.

Defense rushing predicted points added (PPA).

def_first_down: character.

Defense 1st down predicted points added (PPA).

def_second_down: character.

Defense 2nd down predicted points added (PPA).

def_third_down: character.

Defense 3rd down predicted points added (PPA).

Examples

# \donttest{
cfbd_metrics_ppa_games(year = 2019, team = "TCU")

#>      game_id season week conference team            opponent
#> 1  401112129   2019    1     Big 12  TCU Arkansas-Pine Bluff
#> 2  401112130   2019    3     Big 12  TCU              Purdue
#> 3  401112131   2019    4     Big 12  TCU                 SMU
#> 4  401112100   2019    5     Big 12  TCU              Kansas
#> 5  401112088   2019    6     Big 12  TCU          Iowa State
#> 6  401112109   2019    8     Big 12  TCU        Kansas State
#> 7  401112132   2019    9     Big 12  TCU               Texas
#> 8  401112127   2019   10     Big 12  TCU      Oklahoma State
#> 9  401112081   2019   11     Big 12  TCU              Baylor
#> 10 401112133   2019   12     Big 12  TCU          Texas Tech
#> 11 401112120   2019   13     Big 12  TCU            Oklahoma
#> 12 401112134   2019   14     Big 12  TCU       West Virginia
#>                off_overall             off_passing             off_rushing
#> 1   0.08241186305860345043  0.03259498097069323880  0.30213438066559038833
#> 2   0.08173250143654040193 -0.44614832954363769296  0.33405733843114364759
#> 3   0.12805173802739761500 -0.08846116290854666297  0.36654623910464983444
#> 4   0.37980477698230694080  0.78655661643275015429  0.20971350983392199190
#> 5   0.21605826335116850744  0.28269887706827934234  0.15006991320185294848
#> 6   0.17260128796495778211 -0.12441396665521391818  0.43365808495710325486
#> 7   0.20698084486653880582  0.48632374347332391703  0.07353648391980303126
#> 8   0.17328600369176078556 -0.05705769773821963411  0.49035105957558059478
#> 9  -0.06865833078897910918 -0.13236496476247898744 -0.01526747509881401072
#> 10  0.17834444243321469006  0.31688820792717983355  0.13886388898391054204
#> 11 -0.05083715878025127875 -0.31997837919867942500  0.26633614126601881357
#> 12 -0.06194181125385135376 -0.20241655536927595733  0.11270722477046994538
#>             off_first_down         off_second_down          off_third_down
#> 1  -0.20159843289860946000  0.16366920054237664080  0.58551403040472245861
#> 2   0.27079034102799532486 -0.19877397463694024037 -0.08798092910733494235
#> 3   0.07260215428529003182 -0.06396901512995885000  1.02548942572988437400
#> 4   0.19702763485829827474  0.11626148254744460733     1.16148005244326480
#> 5   0.14885129959463354878  0.01850230015257219126  0.49933360660355287214
#> 6   0.04571706610992616987  0.32407825963859998329  0.27545997769413610464
#> 7   0.09843147512640150979  0.07005023456243352130  0.84989565541578161538
#> 8  -0.11765281167206434385 -0.06220660210522395575  0.96745562278317352576
#> 9  -0.36428435389659998300 -0.11587095952937641082  0.24522191873786508971
#> 10  0.03237018224774046744 -0.08520291710110368125  0.90352478887495377417
#> 11  0.05220330145507693763  0.05839241748523671857 -0.20688487695469620444
#> 12 -0.07144631308950564344 -0.19295948713853326485  0.09172792916758818263
#>                def_overall             def_passing             def_rushing
#> 1  -0.06065759968477334484 -0.03764388628367385833 -0.08223295599830411344
#> 2  -0.07172604564015880316  0.06622756332186603298 -0.30497106530997632519
#> 3   0.13316672532550978955  0.33654610756914049314 -0.01161724745831185059
#> 4  -0.17844223106411302960 -0.25720076930439847667 -0.02092515458354213547
#> 5   0.27457173408857221869  0.58903083846771642960  0.05619735604749985000
#> 6   0.08554617077395194118  0.36586225279700669692 -0.09375661446094303290
#> 7   0.15950811843307340064  0.24220219320228677698  0.06891801753034372259
#> 8   0.26014468455171187026  0.31978024626672079772  0.24618755013223392286
#> 9   0.09026708922179553863  0.03988212379269154615  0.12501085214633910097
#> 10  0.37692569218292957043  0.56659192536248324229  0.11064627870036864676
#> 11  0.07710727868487275291 -0.07967090377387716350  0.12850996145823338125
#> 12 -0.17351382675033815778 -0.01114062893380378294 -0.37828870505071153219
#>             def_first_down         def_second_down          def_third_down
#> 1  -0.16700986874297808500 -0.13233293632154340080  0.28193805643817073388
#> 2  -0.19435683665003450510 -0.20710938245377283553  0.27746531214695174552
#> 3   0.00775800428980242268  0.06422480960939882393  0.63689744497234984926
#> 4  -0.60936102206902615189  0.23564420504661706825 -0.07560993467168079000
#> 5   0.39667026117745750625  0.29608133264754008944  0.44067763958463332556
#> 6  -0.46996319677320162973  0.39901038024699069232  0.60529024434759816200
#> 7  -0.25431076437183043387  0.32334694063194480250  0.62207345277435232806
#> 8   0.13728238726108948408  0.19193172774527143541  0.66940019560297404308
#> 9  -0.00841583671847412003  0.01105745573177564382  0.18312928633478108044
#> 10  0.26177355823045406688  0.46618515785912839313  0.67305311608670412164
#> 11 -0.12283850980323885605  0.54123639796762696521 -0.18017457946905594250
#> 12 -0.41363822208132530680 -0.04927759951683955077  0.24736137563005778823

# }