Get team game averages for predicted points added (PPA)

Get team game averages for predicted points added (PPA)

cfbd_metrics_ppa_games(
  year,
  week = NULL,
  team = NULL,
  conference = NULL,
  excl_garbage_time = FALSE
)

Arguments

year
(Integer required): Year, 4 digit format (YYYY)

week
(Integer optional): Week - values range from 1-15, 1-14 for seasons pre-playoff, i.e. 2013 or earlier

team
(String optional): D-I Team

conference
(String optional): Conference abbreviation - Select a valid FBS conference
Conference abbreviations P5: ACC, B12, B1G, SEC, PAC
Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC

excl_garbage_time
(Logical default FALSE): Select whether to exclude Garbage Time (TRUE or FALSE)

Value

cfbd_metrics_ppa_games() - A data frame with 18 variables:

game_id: integer.

Referencing game id.

season: integer.

Season of the game.

week: integer.

Game week of the season.

conference: character.

Conference of the team.

team: character.

Team name.

opponent: character.

Team Opponent.

off_overall: character.

Offense overall predicted points added (PPA).

off_passing: character.

Offense passing predicted points added (PPA).

off_rushing: character.

Offense rushing predicted points added (PPA).

off_first_down: character.

Offense 1st down predicted points added (PPA).

off_second_down: character.

Offense 2nd down predicted points added (PPA).

off_third_down: character.

Offense 3rd down predicted points added (PPA).

def_overall: character.

Defense overall predicted points added (PPA).

def_passing: character.

Defense passing predicted points added (PPA).

def_rushing: character.

Defense rushing predicted points added (PPA).

def_first_down: character.

Defense 1st down predicted points added (PPA).

def_second_down: character.

Defense 2nd down predicted points added (PPA).

def_third_down: character.

Defense 3rd down predicted points added (PPA).

Examples

# \donttest{
cfbd_metrics_ppa_games(year = 2019, team = "TCU")

#>      game_id season week conference team            opponent
#> 1  401112129   2019    1     Big 12  TCU Arkansas-Pine Bluff
#> 2  401112130   2019    3     Big 12  TCU              Purdue
#> 3  401112131   2019    4     Big 12  TCU                 SMU
#> 4  401112100   2019    5     Big 12  TCU              Kansas
#> 5  401112088   2019    6     Big 12  TCU          Iowa State
#> 6  401112109   2019    8     Big 12  TCU        Kansas State
#> 7  401112132   2019    9     Big 12  TCU               Texas
#> 8  401112127   2019   10     Big 12  TCU      Oklahoma State
#> 9  401112081   2019   11     Big 12  TCU              Baylor
#> 10 401112133   2019   12     Big 12  TCU          Texas Tech
#> 11 401112120   2019   13     Big 12  TCU            Oklahoma
#> 12 401112134   2019   14     Big 12  TCU       West Virginia
#>                off_overall             off_passing            off_rushing
#> 1   0.10893335058280822134  0.03259498097069323880 0.34336843527305105147
#> 2   0.08433374364670448048 -0.44614832954363769296 0.33344242040676509036
#> 3   0.14467693394729749883 -0.06230226939107379564 0.37882933476533863351
#> 4   0.38938935355899990281  0.78655661643275015429 0.22960034066096271717
#> 5   0.22675780664778231502  0.28269887706827934234 0.17641660582385020019
#> 6   0.16269524729117812819 -0.12441396665521391818 0.40801882902278008237
#> 7   0.20698084486653880582  0.48632374347332391703 0.07353648391980303126
#> 8   0.17236682128658921120 -0.05705769773821963411 0.46945628747567235097
#> 9  -0.05747532222199197810 -0.13236496476247898744 0.00273752470368853487
#> 10  0.19928588702312061925  0.34181791063511511880 0.15590995259388930520
#> 11 -0.02617175444970218620 -0.26976027101482437609 0.26322523360005291295
#> 12  0.01175254585700390189 -0.12548024258364589436 0.18154955719594856613
#>             off_first_down         off_second_down          off_third_down
#> 1  -0.15572709282725510111  0.20200200059644081926  0.58551403040472245861
#> 2   0.27079034102799532486 -0.19877397463694024037 -0.06641327487413172333
#> 3   0.07414469277764566324 -0.03137755373542027292  1.01271013528115006313
#> 4   0.19702763485829827474  0.11626148254744460733     1.12765513646249137
#> 5   0.14885129959463354878  0.02781246999064648170  0.52685680621475352733
#> 6   0.02739360652689168644  0.32407825963859998329  0.27545997769413610464
#> 7   0.09843147512640150979  0.07005023456243352130  0.84989565541578161538
#> 8  -0.09399410481716083374 -0.07442235215816016762  0.96745562278317352576
#> 9  -0.36428435389659998300 -0.11587095952937641082  0.26044540312940489763
#> 10  0.03237018224774046744 -0.08520291710110368125  0.92843848173564212704
#> 11  0.09134504183382242575  0.06769266782720553733 -0.20688487695469620444
#> 12  0.01775522507666778757 -0.15866311881612091462  0.18961212543283717311
#>                def_overall             def_passing             def_rushing
#> 1  -0.06065759968477334484 -0.03764388628367385833 -0.08223295599830411344
#> 2  -0.04759411021538597989  0.06622756332186603298 -0.23294227424374470586
#> 3   0.14368358559973889206  0.33654610756914049314  0.01025935330379594198
#> 4  -0.12317066204011410494 -0.20089735109133850323  0.02742479799663316675
#> 5   0.27457173408857221869  0.58903083846771642960  0.05619735604749985000
#> 6   0.09022320936271487204  0.36531325749846398904 -0.09375661446094303290
#> 7   0.19327830239265384363  0.28973552883496900684  0.08159726076029447183
#> 8   0.27035105992573229491  0.31978024626672079772  0.26213299502993556963
#> 9   0.10629089494247281474  0.08292326763474930214  0.10611305939622756427
#> 10  0.37692569218292957043  0.56659192536248324229  0.11064627870036864676
#> 11  0.10586763724386083393 -0.00263856878113831619  0.14261973928458635252
#> 12 -0.15736372545009683500  0.01306029566061039158 -0.37828870505071153219
#>             def_first_down         def_second_down          def_third_down
#> 1  -0.16700986874297808500 -0.13233293632154340080  0.28193805643817073388
#> 2  -0.19435683665003450510 -0.20710938245377283553  0.34427844721362958249
#> 3   0.00775800428980242268  0.06422480960939882393  0.65588172214177007680
#> 4  -0.53767496371280042284  0.30932426329684316424 -0.07560993467168079000
#> 5   0.39667026117745750625  0.29608133264754008944  0.44067763958463332556
#> 6  -0.43421969385552160235  0.39901038024699069232  0.60529024434759816200
#> 7  -0.14736371945133821848  0.32334694063194480250  0.62207345277435232806
#> 8   0.16375764907530455874  0.19193172774527143541  0.66940019560297404308
#> 9  -0.00841583671847412003  0.01901022893665299017  0.24739397737722054629
#> 10  0.26177355823045406688  0.46618515785912839313  0.67305311608670412164
#> 11 -0.12283850980323885605  0.57560783704024805710 -0.08913682822588874692
#> 12 -0.36464857290646254500 -0.04927759951683955077  0.24736137563005778823

# }