Get player season averages for predicted points added (PPA)

Get player season averages for predicted points added (PPA)

cfbd_metrics_ppa_players_season(
  year = NULL,
  team = NULL,
  conference = NULL,
  position = NULL,
  athlete_id = NULL,
  threshold = NULL,
  excl_garbage_time = FALSE
)

Arguments

year

(Integer required): Year, 4 digit format (YYYY)

team

(String optional): D-I Team

conference

(String optional): Conference abbreviation - S&P+ information by conference
Conference abbreviations P5: ACC, B12, B1G, SEC, PAC
Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC

position

(string optional): Position abbreviation of the player you are searching for.
Position Group - options include:

Offense: QB, RB, FB, TE, OL, G, OT, C, WR

Defense: DB, CB, S, LB, DE, DT, NT, DL

Special Teams: K, P, LS, PK

athlete_id

(Integer optional): Athlete ID filter for querying a single athlete
Can be found using the cfbd_player_info() function.

threshold

(Integer optional): Minimum threshold of plays.

excl_garbage_time

(Logical default FALSE): Select whether to exclude Garbage Time (TRUE or FALSE)

Value

cfbd_metrics_ppa_players_season() - A data frame with 23 variables:

season: integer.

Season.

athlete_id: character.

Athlete referencing id.

name: character.

Athlete name.

position: character.

Athlete Position.

team: character.

Team name.

conference: character.

Team conference.

countable_plays: integer.

Number of plays which can be counted.

avg_PPA_all: double.

Average overall predicted points added (PPA).

avg_PPA_pass: double.

Average passing predicted points added (PPA).

avg_PPA_rush: double.

Average rushing predicted points added (PPA).

avg_PPA_first_down: double.

Average 1st down predicted points added (PPA).

avg_PPA_second_down: double.

Average 2nd down predicted points added (PPA).

avg_PPA_third_down: double.

Average 3rd down predicted points added (PPA).

avg_PPA_standard_downs: double.

Average standard down predicted points added (PPA).

avg_PPA_passing_downs: double.

Average passing down predicted points added (PPA).

total_PPA_all: double.

Total overall predicted points added (PPA).

total_PPA_pass: double.

Total passing predicted points added (PPA).

total_PPA_rush: double.

Total rushing predicted points added (PPA).

total_PPA_first_down: double.

Total 1st down predicted points added (PPA).

total_PPA_second_down: double.

Total 2nd down predicted points added (PPA).

total_PPA_third_down: double.

Total 3rd down predicted points added (PPA).

total_PPA_standard_downs: double.

Total standard down predicted points added (PPA).

total_PPA_passing_downs: double.

Total passing down predicted points added (PPA).

Examples

# \donttest{
cfbd_metrics_ppa_players_season(year = 2019, team = "TCU")
#>    season athlete_id              name position team conference countable_plays
#> 1    2019    4427105        Max Duggan       QB  TCU     Big 12             466
#> 2    2019    4038533   Darius Anderson       RB  TCU     Big 12             177
#> 3    2019    4038539     Sewo Olonilua       RB  TCU     Big 12             164
#> 4    2019    4241802      Jalen Reagor       WR  TCU     Big 12              92
#> 5    2019    3886634       Alex Delton       QB  TCU     Big 12              66
#> 6    2019    4362430       Taye Barber       WR  TCU     Big 12              49
#> 7    2019    4362442         Pro Wells       TE  TCU     Big 12              28
#> 8    2019    4426635     Darwin Barlow       RB  TCU     Big 12              23
#> 9    2019    4362477      Derius Davis       WR  TCU     Big 12              21
#> 10   2019    4362478   Emari Demercado       RB  TCU     Big 12              18
#> 11   2019    4038556   Artayvious Lynn       TE  TCU     Big 12              16
#> 12   2019    4362489 John Stephens Jr.       WR  TCU     Big 12              15
#> 13   2019    4038536      Dylan Thomas       WR  TCU     Big 12              10
#> 14   2019    4362483   Tevailance Hunt       WR  TCU     Big 12               9
#> 15   2019    3676818  TreVontae Hights       WR  TCU     Big 12               9
#> 16   2019    4427217  Daimarqua Foster       RB  TCU     Big 12               8
#> 17   2019    4426965   Blair Conwright       WR  TCU     Big 12               4
#> 18   2019    4259331       Carter Ware       TE  TCU     Big 12               2
#> 19   2019    4241795   Al'Dontre Davis       WR  TCU     Big 12               1
#>    avg_PPA_all avg_PPA_pass avg_PPA_rush avg_PPA_first_down avg_PPA_second_down
#> 1        0.187        0.092        0.518              0.001               0.040
#> 2        0.076       -0.144        0.114              0.056              -0.005
#> 3        0.158       -0.104        0.224             -0.081               0.072
#> 4        0.299        0.347        0.030              0.397              -0.046
#> 5        0.022       -0.124        0.565              0.284              -0.449
#> 6        0.591        0.550        0.958              0.521               0.104
#> 7        0.642        0.642           NA              0.058               0.343
#> 8        0.166       -0.881        0.214             -0.170               0.373
#> 9        0.243        0.296       -0.075              0.144              -0.179
#> 10      -0.149           NA       -0.149              0.120              -0.407
#> 11       0.506        0.506           NA              0.273               0.662
#> 12       0.726        0.726           NA              0.658               0.139
#> 13       1.178        1.178           NA             -0.384               0.754
#> 14       0.385        0.385           NA                 NA               0.224
#> 15       0.462        0.462           NA              0.159                  NA
#> 16       0.365           NA        0.365              0.148              -0.381
#> 17       1.068        1.068           NA             -0.020               2.039
#> 18       0.808        0.808           NA             -0.362                  NA
#> 19       3.183        3.183           NA                 NA                  NA
#>    avg_PPA_third_down avg_PPA_standard_downs avg_PPA_passing_downs
#> 1               0.589                  0.050                 0.376
#> 2               1.060                  0.030                 0.294
#> 3               0.817                  0.134                 0.237
#> 4               0.663                  0.287                 0.318
#> 5               0.076                  0.043                -0.005
#> 6               0.922                  0.581                 0.607
#> 7               1.202                  0.092                 1.118
#> 8               0.235                  0.144                 0.209
#> 9               0.580                  0.242                 0.245
#> 10             -0.165                 -0.147                -0.164
#> 11              0.775                  0.392                 0.695
#> 12              1.088                  0.734                 0.717
#> 13              2.952                  0.457                 2.260
#> 14             -0.466                 -0.066                 0.747
#> 15              1.069                  0.159                 1.069
#> 16              2.946                 -0.047                 1.602
#> 17              2.273                 -0.020                 2.156
#> 18              1.979                  0.808                    NA
#> 19              3.183                     NA                 3.183
#>    total_PPA_all total_PPA_pass total_PPA_rush total_PPA_first_down
#> 1         87.252         33.428         53.824                0.143
#> 2         13.507         -3.744         17.250                5.605
#> 3         25.918         -3.418         29.336               -5.780
#> 4         27.500         27.076          0.424               16.691
#> 5          1.456         -6.459          7.915                6.809
#> 6         28.980         24.191          4.789               10.419
#> 7         17.965         17.965             NA                0.349
#> 8          3.823         -0.881          4.704               -1.363
#> 9          5.105          5.328         -0.224                1.579
#> 10        -2.673             NA         -2.673                0.839
#> 11         8.092          8.092             NA                1.911
#> 12        10.893         10.893             NA                3.947
#> 13        11.781         11.781             NA               -1.536
#> 14         3.469          3.469             NA                   NA
#> 15         4.162          4.162             NA                0.955
#> 16         2.923             NA          2.923                0.739
#> 17         4.272          4.272             NA               -0.039
#> 18         1.617          1.617             NA               -0.362
#> 19         3.183          3.183             NA                   NA
#>    total_PPA_second_down total_PPA_third_down total_PPA_standard_downs
#> 1                  6.099               77.143                   13.630
#> 2                 -0.345               10.602                    4.390
#> 3                  3.905               30.246                   16.895
#> 4                 -1.348               13.269                   16.366
#> 5                 -8.990                1.600                    1.603
#> 6                  1.044               17.517                   16.839
#> 7                  3.089               14.418                    1.197
#> 8                  4.482                0.704                    2.154
#> 9                 -0.536                4.062                    3.141
#> 10                -2.851               -0.661                   -2.345
#> 11                 4.631                1.549                    3.920
#> 12                 0.416                6.530                    5.875
#> 13                 1.509               11.808                    2.741
#> 14                 0.896               -1.864                   -0.265
#> 15                    NA                3.208                    0.955
#> 16                -0.762                2.946                   -0.281
#> 17                 2.039                2.273                   -0.039
#> 18                    NA                1.979                    1.617
#> 19                    NA                3.183                       NA
#>    total_PPA_passing_downs
#> 1                   73.622
#> 2                    9.117
#> 3                    9.023
#> 4                   11.135
#> 5                   -0.147
#> 6                   12.141
#> 7                   16.769
#> 8                    1.669
#> 9                    1.963
#> 10                  -0.328
#> 11                   4.171
#> 12                   5.018
#> 13                   9.040
#> 14                   3.734
#> 15                   3.208
#> 16                   3.204
#> 17                   4.311
#> 18                      NA
#> 19                   3.183
# }