Get college football play-by-play data.

Get college football play-by-play data.

cfbd_plays(
  year = 2020,
  season_type = "regular",
  week = 1,
  team = NULL,
  offense = NULL,
  defense = NULL,
  conference = NULL,
  offense_conference = NULL,
  defense_conference = NULL,
  play_type = NULL,
  verbose = FALSE
)

Arguments

year
Select year, (example: 2018)

season_type
Select Season Type (regular, postseason, both)

week
Select week, this is optional (also numeric)

team
Select team name (example: Texas, Texas A&M, Clemson)

offense
Select offense name (example: Texas, Texas A&M, Clemson)

defense
Select defense name (example: Texas, Texas A&M, Clemson)

conference
Select conference name (example: ACC, B1G, B12, SEC,
PAC, MAC, MWC, CUSA, Ind, SBC, AAC, Western, MVIAA, SWC, PCC, Big 6, etc.)

offense_conference
Select conference name (example: ACC, B1G, B12, SEC,
PAC, MAC, MWC, CUSA, Ind, SBC, AAC, Western, MVIAA, SWC, PCC, Big 6, etc.)

defense_conference
Select conference name (example: ACC, B1G, B12, SEC,
PAC, MAC, MWC, CUSA, Ind, SBC, AAC, Western, MVIAA, SWC, PCC, Big 6, etc.)

play_type
Select play type (example: see the cfbd_play_type_df)

verbose
Logical parameter (TRUE/FALSE, default: FALSE) to return warnings and messages from function

Source

https://api.collegefootballdata.com/plays

https://api.collegefootballdata.com/plays

Value

cfbd_plays() - A data frame with 29 columns:

play_id: character.

Referencing play id.

offense: character.

Offense on the field.

offense_conference: character.

Conference of the offense on the field.

defense: character.

Defense on the field.

defense_conference: character.

Conference of the defense on the field.

home: character.

Home team.

away: character.

Away team.

offense_score: integer.

Offense's post-play score.

defense_score: integer.

Defense's post-play score.

game_id: integer.

Referencing game id.

drive_id: character.

Referencing drive id.

drive_number: integer.

Drive number in the game.

play_number: integer.

Play number in the game.

period: integer.

Game period (quarter).

offense_timeouts: integer.

Timeouts for the offense at the end of the play.

defense_timeouts: integer.

Timeouts for the defense at the end of the play.

yard_line: integer.

Yard line (~0-50) of the play.

yards_to_goal: integer.

Yards to the goal line (~0-100).

down: integer.

Down of the play.

distance: integer.

Distance to the sticks, i.e. 1st down or goal-line in goal-to-go situations.

scoring: logical.

Scoring play flag.

yards_gained: integer.

Yards net gained by the offense on the play.

play_type: character.

Categorical label of the type of the play.

play_text: character.

A text description of the play.

ppa: character.

Predicted Points Added (calculated by CFBD).

clock.minutes: integer.

Minutes left on the clock.

clock.seconds: integer.

Seconds left on the clock.

Examples

if (FALSE) {
 year_vector <- 2020
 week_vector <- 1:3
 weekly_year_df <- expand.grid(year = year_vector, week = week_vector)
 tictoc::tic()
 year_split <- split(weekly_year_df, weekly_year_df$year)
 for (i in 1:length(year_split)) {
   i <- 1
   future::plan("multisession")
   progressr::with_progress({
      year_split[[i]] <- year_split[[i]] %>%
         dplyr::mutate(
            pbp = purrr::map2(
                .x = year,
                .y = week,
                cfbd_plays,
                season_type = "both"
            )
         )
   })
 }
 
 tictoc::toc()
 year_split <- lapply(year_split, function(x) {
     x %>% tidyr::unnest(pbp, names_repair = "minimal")
 })
 
 all_years <- dplyr::bind_rows(year_split)
 glimpse(all_years)
}