Get college football position group recruiting information.

Get college football position group recruiting information.

cfbd_recruiting_position(
  start_year = NULL,
  end_year = NULL,
  team = NULL,
  conference = NULL,
  verbose = FALSE
)

Arguments

start_year
(Integer optional): Start Year, 4 digit format (YYYY). Note: 2000 is the minimum value

end_year
(Integer optional): End Year, 4 digit format (YYYY). Note: 2020 is the maximum value currently

team
(String optional): Team - Select a valid team, D-I football

conference
(String optional): Conference abbreviation - Select a valid FBS conference
Conference abbreviations P5: ACC, B12, B1G, SEC, PAC
Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC

verbose
Logical parameter (TRUE/FALSE, default: FALSE) to return warnings and messages from function

Source

https://api.collegefootballdata.com/recruiting/groups

Value

cfbd_recruiting_position() - A data frame with 7 variables:

team: character.

Recruiting team.

conference: character.

Recruiting team conference.

position_group: character.

Position group of the recruits.

avg_rating: double.

Average rating of the recruits in the position group.

total_rating: double.

Sum of the ratings of the recruits in the position group.

commits: integer.

Number of commits in the position group.

avg_stars: double.

Average stars of the recruits in the position group.

Examples

# \donttest{
   cfbd_recruiting_position(2018, team = "Texas")

#>    team conference position_group avg_rating total_rating commits avg_stars
#> 1 Texas     Big 12 Defensive Back  0.9427063      15.0833      16  4.062500
#> 2 Texas     Big 12 Defensive Line  0.8996687      14.3947      16  3.500000
#> 3 Texas     Big 12     Linebacker  0.9024700       9.0247      10  3.600000
#> 4 Texas     Big 12 Offensive Line  0.9035769      11.7465      13  3.769231
#> 5 Texas     Big 12    Quarterback  0.9197667       5.5186       6  3.833333
#> 6 Texas     Big 12       Receiver  0.9105200      13.6578      15  3.600000
#> 7 Texas     Big 12   Running Back  0.9327000       3.7308       4  4.250000
#> 8 Texas     Big 12  Special Teams  0.8849333      10.6192      12  3.583333
#> 9 Texas     Big 12  All Positions  0.9106043      83.7756      92  3.728261

   cfbd_recruiting_position(2016, 2020, team = "Virginia")

#>       team conference position_group avg_rating total_rating commits avg_stars
#> 1 Virginia        ACC Defensive Back  0.8332857      11.6660      14  2.857143
#> 2 Virginia        ACC Defensive Line  0.8508000      14.4636      17  3.117647
#> 3 Virginia        ACC     Linebacker  0.8461053      16.0760      19  3.000000
#> 4 Virginia        ACC Offensive Line  0.8452158      16.0591      19  3.052632
#> 5 Virginia        ACC    Quarterback  0.8465500       5.0793       6  3.000000
#> 6 Virginia        ACC       Receiver  0.8401950      16.8039      20  2.950000
#> 7 Virginia        ACC   Running Back  0.8453714       5.9176       7  3.000000
#> 8 Virginia        ACC  Special Teams  0.8361111       7.5250       9  2.888889
#> 9 Virginia        ACC  All Positions  0.8431577      93.5905     111  2.990991

   cfbd_recruiting_position(2015, 2020, conference = "SEC")

#>                  team conference position_group avg_rating total_rating commits
#> 1             Alabama        SEC Defensive Back  0.9464875      22.7157      24
#> 2             Alabama        SEC Defensive Line  0.9543032      29.5834      31
#> 3             Alabama        SEC     Linebacker  0.9390235      15.9634      17
#> 4             Alabama        SEC Offensive Line  0.9327652      21.4536      23
#> 5             Alabama        SEC    Quarterback  0.9061444       8.1553       9
#> 6             Alabama        SEC       Receiver  0.9246435      21.2668      23
#> 7             Alabama        SEC   Running Back  0.9298214      13.0175      14
#> 8             Alabama        SEC  Special Teams  0.8863400       8.8634      10
#> 9            Arkansas        SEC Defensive Back  0.8627654      22.4319      26
#> 10           Arkansas        SEC Defensive Line  0.8787962      22.8487      26
#> 11           Arkansas        SEC     Linebacker  0.8500211      16.1504      19
#> 12           Arkansas        SEC Offensive Line  0.8516045      18.7353      22
#> 13           Arkansas        SEC    Quarterback  0.8730875       6.9847       8
#> 14           Arkansas        SEC       Receiver  0.8797947      16.7161      19
#> 15           Arkansas        SEC   Running Back  0.8774571       6.1422       7
#> 16           Arkansas        SEC  Special Teams  0.8589000       8.5890      10
#> 17             Auburn        SEC Defensive Back  0.8867389      15.9613      18
#> 18             Auburn        SEC Defensive Line  0.9239696      21.2513      23
#> 19             Auburn        SEC     Linebacker  0.9038409      19.8845      22
#> 20             Auburn        SEC Offensive Line  0.8966056      16.1389      18
#> 21             Auburn        SEC    Quarterback  0.8898800       4.4494       5
#> 22             Auburn        SEC       Receiver  0.9012440      22.5311      25
#> 23             Auburn        SEC   Running Back  0.9115455      10.0270      11
#> 24             Auburn        SEC  Special Teams  0.8934636      19.6562      22
#> 25            Florida        SEC Defensive Back  0.9018387      27.9570      31
#> 26            Florida        SEC Defensive Line  0.9034652      20.7797      23
#> 27            Florida        SEC     Linebacker  0.8898400      13.3476      15
#> 28            Florida        SEC Offensive Line  0.8827556      23.8344      27
#> 29            Florida        SEC    Quarterback  0.8856571       6.1996       7
#> 30            Florida        SEC       Receiver  0.9105240      22.7631      25
#> 31            Florida        SEC   Running Back  0.8969375       7.1755       8
#> 32            Florida        SEC  Special Teams  0.8470000       5.0820       6
#> 33            Georgia        SEC Defensive Back  0.9101815      24.5749      27
#> 34            Georgia        SEC Defensive Line  0.9288600      27.8658      30
#> 35            Georgia        SEC     Linebacker  0.9449941      16.0649      17
#> 36            Georgia        SEC Offensive Line  0.9216269      23.9623      26
#> 37            Georgia        SEC    Quarterback  0.9209000       7.3672       8
#> 38            Georgia        SEC       Receiver  0.9226769      23.9896      26
#> 39            Georgia        SEC   Running Back  0.9420125       7.5361       8
#> 40            Georgia        SEC  Special Teams  0.8812250      10.5747      12
#> 41           Kentucky        SEC Defensive Back  0.8647150      17.2943      20
#> 42           Kentucky        SEC Defensive Line  0.8664259      23.3935      27
#> 43           Kentucky        SEC     Linebacker  0.8635083      10.3621      12
#> 44           Kentucky        SEC Offensive Line  0.8633545      18.9938      22
#> 45           Kentucky        SEC    Quarterback  0.8507000       5.9549       7
#> 46           Kentucky        SEC       Receiver  0.8583818      18.8844      22
#> 47           Kentucky        SEC   Running Back  0.8748000       8.7480      10
#> 48           Kentucky        SEC  Special Teams  0.8524353      14.4914      17
#> 49                LSU        SEC Defensive Back  0.9398087      21.6156      23
#> 50                LSU        SEC Defensive Line  0.9160167      27.4805      30
#> 51                LSU        SEC     Linebacker  0.9289000      11.1468      12
#> 52                LSU        SEC Offensive Line  0.9100520      22.7513      25
#> 53                LSU        SEC    Quarterback  0.8728000       6.9824       8
#> 54                LSU        SEC       Receiver  0.9114040      22.7851      25
#> 55                LSU        SEC   Running Back  0.9257182      10.1829      11
#> 56                LSU        SEC  Special Teams  0.8538308      11.0998      13
#> 57  Mississippi State        SEC Defensive Back  0.8741818      19.2320      22
#> 58  Mississippi State        SEC Defensive Line  0.8818545      19.4008      22
#> 59  Mississippi State        SEC     Linebacker  0.9073500       7.2588       8
#> 60  Mississippi State        SEC Offensive Line  0.8599105      16.3383      19
#> 61  Mississippi State        SEC    Quarterback  0.8763167       5.2579       6
#> 62  Mississippi State        SEC       Receiver  0.8560118      14.5522      17
#> 63  Mississippi State        SEC   Running Back  0.8826143       6.1783       7
#> 64  Mississippi State        SEC  Special Teams  0.8568000      15.4224      18
#> 65           Missouri        SEC Defensive Back  0.8483348      19.5117      23
#> 66           Missouri        SEC Defensive Line  0.8602500      13.7640      16
#> 67           Missouri        SEC     Linebacker  0.8505937      13.6095      16
#> 68           Missouri        SEC Offensive Line  0.8572810      18.0029      21
#> 69           Missouri        SEC    Quarterback  0.8780167       5.2681       6
#> 70           Missouri        SEC       Receiver  0.8580182      18.8764      22
#> 71           Missouri        SEC   Running Back  0.8418875       6.7351       8
#> 72           Missouri        SEC  Special Teams  0.8499800       8.4998      10
#> 73           Ole Miss        SEC Defensive Back  0.8617200      17.2344      20
#> 74           Ole Miss        SEC Defensive Line  0.8743600      17.4872      20
#> 75           Ole Miss        SEC     Linebacker  0.8566867      12.8503      15
#> 76           Ole Miss        SEC Offensive Line  0.8736148      23.5876      27
#> 77           Ole Miss        SEC    Quarterback  0.8837667       7.9539       9
#> 78           Ole Miss        SEC       Receiver  0.9119895      17.3278      19
#> 79           Ole Miss        SEC   Running Back  0.8900500       8.9005      10
#> 80           Ole Miss        SEC  Special Teams  0.8908600       4.4543       5
#> 81     South Carolina        SEC Defensive Back  0.8699762      18.2695      21
#> 82     South Carolina        SEC Defensive Line  0.8877231      23.0808      26
#> 83     South Carolina        SEC     Linebacker  0.8668929      12.1365      14
#> 84     South Carolina        SEC Offensive Line  0.8646038      22.4797      26
#> 85     South Carolina        SEC    Quarterback  0.9170286       6.4192       7
#> 86     South Carolina        SEC       Receiver  0.8708417      20.9002      24
#> 87     South Carolina        SEC   Running Back  0.8799889       7.9199       9
#> 88     South Carolina        SEC  Special Teams  0.8698667      10.4384      12
#> 89          Tennessee        SEC Defensive Back  0.8865708      21.2777      24
#> 90          Tennessee        SEC Defensive Line  0.8988600      22.4715      25
#> 91          Tennessee        SEC     Linebacker  0.8986687      14.3787      16
#> 92          Tennessee        SEC Offensive Line  0.8976727      19.7488      22
#> 93          Tennessee        SEC    Quarterback  0.8861556       7.9754       9
#> 94          Tennessee        SEC       Receiver  0.8760739      20.1497      23
#> 95          Tennessee        SEC   Running Back  0.8638636       9.5025      11
#> 96          Tennessee        SEC  Special Teams  0.8606813      13.7709      16
#> 97          Texas A&M        SEC Defensive Back  0.9167524      19.2518      21
#> 98          Texas A&M        SEC Defensive Line  0.9068423      23.5779      26
#> 99          Texas A&M        SEC     Linebacker  0.8815350      17.6307      20
#> 100         Texas A&M        SEC Offensive Line  0.8876346      23.0785      26
#> 101         Texas A&M        SEC    Quarterback  0.8954750       7.1638       8
#> 102         Texas A&M        SEC       Receiver  0.9130464      25.5653      28
#> 103         Texas A&M        SEC   Running Back  0.8864250      10.6371      12
#> 104         Texas A&M        SEC  Special Teams  0.8424000       8.4240      10
#> 105        Vanderbilt        SEC Defensive Back  0.8578789      16.2997      19
#> 106        Vanderbilt        SEC Defensive Line  0.8307143      11.6300      14
#> 107        Vanderbilt        SEC     Linebacker  0.8476769      11.0198      13
#> 108        Vanderbilt        SEC Offensive Line  0.8466750      20.3202      24
#> 109        Vanderbilt        SEC    Quarterback  0.8761857       6.1333       7
#> 110        Vanderbilt        SEC       Receiver  0.8486955      18.6713      22
#> 111        Vanderbilt        SEC   Running Back  0.8454625       6.7637       8
#> 112        Vanderbilt        SEC  Special Teams  0.8177909       8.9957      11
#> 113           Alabama        SEC  All Positions  0.9339013     141.0190     151
#> 114          Arkansas        SEC  All Positions  0.8656810     118.5980     137
#> 115            Auburn        SEC  All Positions  0.9020812     129.9000     144
#> 116           Florida        SEC  All Positions  0.8953444     127.1390     142
#> 117           Georgia        SEC  All Positions  0.9216591     141.9350     154
#> 118          Kentucky        SEC  All Positions  0.8622073     118.1220     137
#> 119               LSU        SEC  All Positions  0.9118667     134.0440     147
#> 120 Mississippi State        SEC  All Positions  0.8709303     103.6410     119
#> 121          Missouri        SEC  All Positions  0.8546516     104.2670     122
#> 122          Ole Miss        SEC  All Positions  0.8783680     109.7960     125
#> 123    South Carolina        SEC  All Positions  0.8751381     121.6440     139
#> 124         Tennessee        SEC  All Positions  0.8854466     129.2750     146
#> 125         Texas A&M        SEC  All Positions  0.8962192     135.3290     151
#> 126        Vanderbilt        SEC  All Positions  0.8460483      99.8337     118
#>     avg_stars
#> 1    4.000000
#> 2    4.193548
#> 3    4.000000
#> 4    3.956522
#> 5    3.888889
#> 6    3.826087
#> 7    3.928571
#> 8    3.500000
#> 9    3.230769
#> 10   3.307692
#> 11   3.052632
#> 12   3.045455
#> 13   3.500000
#> 14   3.368421
#> 15   3.142857
#> 16   3.200000
#> 17   3.500000
#> 18   3.782609
#> 19   3.727273
#> 20   3.444444
#> 21   3.600000
#> 22   3.560000
#> 23   3.909091
#> 24   3.590909
#> 25   3.516129
#> 26   3.826087
#> 27   3.400000
#> 28   3.444444
#> 29   3.428571
#> 30   3.600000
#> 31   3.500000
#> 32   2.833333
#> 33   3.666667
#> 34   3.966667
#> 35   4.058824
#> 36   3.769231
#> 37   4.000000
#> 38   3.846154
#> 39   4.125000
#> 40   3.583333
#> 41   3.100000
#> 42   3.222222
#> 43   3.333333
#> 44   3.181818
#> 45   3.000000
#> 46   3.090909
#> 47   3.200000
#> 48   3.176471
#> 49   4.043478
#> 50   3.666667
#> 51   3.916667
#> 52   3.600000
#> 53   3.375000
#> 54   3.760000
#> 55   3.909091
#> 56   3.076923
#> 57   3.136364
#> 58   3.409091
#> 59   3.375000
#> 60   3.105263
#> 61   3.500000
#> 62   3.058824
#> 63   3.428571
#> 64   2.944444
#> 65   3.043478
#> 66   3.187500
#> 67   3.000000
#> 68   3.142857
#> 69   3.333333
#> 70   3.045455
#> 71   3.000000
#> 72   2.900000
#> 73   3.050000
#> 74   3.250000
#> 75   3.066667
#> 76   3.185185
#> 77   3.444444
#> 78   3.684211
#> 79   3.500000
#> 80   3.400000
#> 81   3.142857
#> 82   3.500000
#> 83   3.214286
#> 84   3.115385
#> 85   3.714286
#> 86   3.208333
#> 87   3.111111
#> 88   3.416667
#> 89   3.333333
#> 90   3.560000
#> 91   3.562500
#> 92   3.545455
#> 93   3.444444
#> 94   3.260870
#> 95   3.181818
#> 96   3.062500
#> 97   3.761905
#> 98   3.615385
#> 99   3.300000
#> 100  3.461538
#> 101  3.500000
#> 102  3.678571
#> 103  3.416667
#> 104  3.000000
#> 105  3.105263
#> 106  2.785714
#> 107  3.000000
#> 108  3.000000
#> 109  3.142857
#> 110  3.090909
#> 111  3.000000
#> 112  2.636364
#> 113  3.960265
#> 114  3.218978
#> 115  3.631944
#> 116  3.521127
#> 117  3.850649
#> 118  3.167883
#> 119  3.700680
#> 120  3.193277
#> 121  3.073770
#> 122  3.288000
#> 123  3.273381
#> 124  3.383562
#> 125  3.516556
#> 126  2.983051

# }