Get college football position group recruiting information.

Get college football position group recruiting information.

Usage

cfbd_recruiting_position(
  start_year = NULL,
  end_year = NULL,
  team = NULL,
  conference = NULL
)

Arguments

start_year

(Integer optional): Start Year, 4 digit format (YYYY). Note: 2000 is the minimum value

end_year

(Integer optional): End Year, 4 digit format (YYYY). Note: 2020 is the maximum value currently

team

(String optional): Team - Select a valid team, D-I football

conference

(String optional): Conference abbreviation - Select a valid FBS conference Conference abbreviations P5: ACC, B12, B1G, SEC, PAC Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC

Value

cfbd_recruiting_position() - A data frame with 7 variables:

team: character.

Recruiting team.

conference: character.

Recruiting team conference.

position_group: character.

Position group of the recruits.

avg_rating: double.

Average rating of the recruits in the position group.

total_rating: double.

Sum of the ratings of the recruits in the position group.

commits: integer.

Number of commits in the position group.

avg_stars: double.

Average stars of the recruits in the position group.

Examples

# \donttest{
  try(cfbd_recruiting_position(2018, team = "Texas"))
#> ── Recruiting position group info from CollegeFootballData.com ─────────────────
#> ℹ Data updated: 2022-11-24 08:46:16 UTC
#> # A tibble: 9 × 7
#>   team  conference position_group avg_rating total_rating commits avg_stars
#>   <chr> <chr>      <chr>               <dbl>        <dbl>   <int>     <dbl>
#> 1 Texas Big 12     Defensive Back      0.940        19.7       21      4.05
#> 2 Texas Big 12     Defensive Line      0.905        21.7       24      3.67
#> 3 Texas Big 12     Linebacker          0.901         9.91      11      3.55
#> 4 Texas Big 12     Offensive Line      0.913        18.3       20      3.9 
#> 5 Texas Big 12     Quarterback         0.921         6.45       7      3.86
#> 6 Texas Big 12     Receiver            0.913        14.6       16      3.62
#> 7 Texas Big 12     Running Back        0.931         4.65       5      4.2 
#> 8 Texas Big 12     Special Teams       0.872        14.0       16      3.38
#> 9 Texas Big 12     All Positions       0.911       109.       120      3.75

  try(cfbd_recruiting_position(2016, 2020, team = "Virginia"))
#> ── Recruiting position group info from CollegeFootballData.com ─────────────────
#> ℹ Data updated: 2022-11-24 08:46:16 UTC
#> # A tibble: 9 × 7
#>   team     conference position_group avg_rating total_rating commits avg_stars
#>   <chr>    <chr>      <chr>               <dbl>        <dbl>   <int>     <dbl>
#> 1 Virginia ACC        Defensive Back      0.833        11.7       14      2.86
#> 2 Virginia ACC        Defensive Line      0.851        14.5       17      3.12
#> 3 Virginia ACC        Linebacker          0.846        16.1       19      3   
#> 4 Virginia ACC        Offensive Line      0.845        16.1       19      3.05
#> 5 Virginia ACC        Quarterback         0.847         5.08       6      3   
#> 6 Virginia ACC        Receiver            0.840        16.8       20      2.95
#> 7 Virginia ACC        Running Back        0.845         5.92       7      3   
#> 8 Virginia ACC        Special Teams       0.836         7.52       9      2.89
#> 9 Virginia ACC        All Positions       0.843        93.6      111      2.99

  try(cfbd_recruiting_position(2015, 2020, conference = "SEC"))
#> ── Recruiting position group info from CollegeFootballData.com ─────────────────
#> ℹ Data updated: 2022-11-24 08:46:16 UTC
#> # A tibble: 126 × 7
#>    team     conference position_group avg_rating total_rating commits avg_stars
#>    <chr>    <chr>      <chr>               <dbl>        <dbl>   <int>     <dbl>
#>  1 Alabama  SEC        Defensive Back      0.946        22.7       24      4   
#>  2 Alabama  SEC        Defensive Line      0.954        29.6       31      4.19
#>  3 Alabama  SEC        Linebacker          0.939        16.0       17      4   
#>  4 Alabama  SEC        Offensive Line      0.933        21.5       23      3.96
#>  5 Alabama  SEC        Quarterback         0.906         8.16       9      3.89
#>  6 Alabama  SEC        Receiver            0.925        21.3       23      3.83
#>  7 Alabama  SEC        Running Back        0.930        13.0       14      3.93
#>  8 Alabama  SEC        Special Teams       0.886         8.86      10      3.5 
#>  9 Arkansas SEC        Defensive Back      0.863        22.4       26      3.23
#> 10 Arkansas SEC        Defensive Line      0.879        22.8       26      3.31
#> # … with 116 more rows
# }