CFBD Teams Endpoint

CFBD Teams Endpoint

Team Info Lookup Lists all teams in conference or all D-I teams if conference is left NULL Current support only for D-I

Get matchup history records between two teams.

Get matchup history between two teams.

Team Roster Get a teams full roster by year. If team is not selected, API returns rosters for every team from the selected year.

Get composite team talent rankings for all teams in a given year Extracts team talent composite as sourced from 247 rankings

cfbd_team_info(conference = NULL, only_fbs = TRUE, year = NULL)

cfbd_team_matchup_records(team1, team2, min_year = NULL, max_year = NULL)

cfbd_team_matchup(team1, team2, min_year = NULL, max_year = NULL)

cfbd_team_roster(year, team = NULL)

cfbd_team_talent(year = NULL)

Arguments

conference
(String optional): Conference abbreviation - Select a valid FBS conference
Conference abbreviations P5: ACC, B12, B1G, SEC, PAC,
Conference abbreviations G5 and FBS Independents: CUSA, MAC, MWC, Ind, SBC, AAC

only_fbs
(Logical default TRUE): Filter for only returning FBS teams for a given year.
If year is left blank while only_fbs is TRUE, then will return values for most current year

year
(Integer optional): Year 4 digit format (YYYY)

team1
(String required): D-I Team 1

team2
(String required): D-I Team 2

min_year
(Integer optional): Minimum of year range, 4 digit format (YYYY)

max_year
(Integer optional): Maximum of year range, 4 digit format (YYYY)

team
(String optional): Team, select a valid team in D-I football

Source

https://api.collegefootballdata.com/teams

https://api.collegefootballdata.com/teams/matchup

https://api.collegefootballdata.com/teams/matchup

https://api.collegefootballdata.com/roster

https://api.collegefootballdata.com/talent

Value

cfbfastR::cfbd_team_info() - A data frame with 12 variables:

team_id

integer.

school

character.

mascot

character.

abbreviation

character.

alt_name1

character.

alt_name2

character.

alt_name3

character.

conference

character.

division

character.

color

character.

alt_color

character.

logos

list.

cfbfastR::cfbd_team_matchup_records() - A data frame with 7 variables:

start_year

character.

end_year

character.

team1

character.

team1_wins

character.

team2

character.

team2_wins

character.

ties

character.

cfbfastR::cfbd_team_matchup() - A data frame with 11 variables:

season

integer.

week

integer.

season_type

character.

date

character.

neutral_site

logical.

venue

character.

home_team

character.

home_score

integer.

away_team

character.

away_score

integer.

winner

character.

cfbfastR::cfbd_team_roster() - A data frame with 12 variables:

athlete_id

character.

first_name

character.

last_name

character.

weight

integer.

height

integer.

jersey

integer.

year

integer.

position

character.

home_city

character.

home_state

character.

home_country

character.

team

character.

cfbfastR::cfbd_team_talent() - A data frame with 3 variables:

year

integer.

school

character.

talent

double.

Examples

if (FALSE) {
  cfbd_team_info(conference = "SEC")

  cfbd_team_info(conference = "Ind")

  cfbd_team_info(year = 2019)
}
if (FALSE) {
  cfbd_team_matchup_records('Texas','Oklahoma')

  cfbd_team_matchup_records('Texas A&M','TCU', min_year = 1975)
}

if (FALSE) {
  cfbd_team_matchup('Texas','Oklahoma')

  cfbd_team_matchup('Texas A&M','TCU')

  cfbd_team_matchup('Texas A&M','TCU', min_year = 1975)

  cfbd_team_matchup('Florida State', 'Florida', min_year = 1975)
}
if (FALSE) {
  cfbd_team_roster(year = 2013, team = "Florida State")
}
if (FALSE) {
  cfbd_team_talent()

  cfbd_team_talent(year = 2018)
}