ESPN Metrics

ESPN Metrics

Get win probability chart data from ESPN Graciously contributed by MrCaseB:

espn_metrics_wp(game_id)

Arguments

game_id
(Integer required): Game ID filter for querying a single game
Can be found using the cfbd_game_info() function

Source

https://gist.github.com/mrcaseb/0f868193affb4be152e8e82c43a4dc07

Value

espn_metrics_wp() - A data frame with 5 variables:

game_id: character.

Referencing game ID (should be same as game_id from other functions).

play_id: character.

Referencing play ID.

seconds_left: integer.

Seconds left in the game.

home_win_percentage: double.

The probability of the home team winning the game.

away_win_percentage: double.

The probability of the away team winning the game (calculated as 1 - home_win_percentage - tie_percentage).

tie_percentage: double.

The probability of the game ending the final period in a tie.

Examples

# \donttest{
espn_metrics_wp(game_id = 401114164)

#> 2021-06-27 17:25:35: Scraping ESPN win probability data for game_id '401114164'...

#>       game_id            play_id seconds_left home_win_percentage
#> 1   401114164 401114164101849902         3600               0.405
#> 2   401114164 401114164101854501         3594               0.409
#> 3   401114164 401114164101857601         3563               0.414
#> 4   401114164 401114164101865401         3525               0.372
#> 5   401114164 401114164101868701         3492               0.370
#> 6   401114164 401114164101869201         3487               0.418
#> 7   401114164 401114164101876901         3450               0.375
#> 8   401114164 401114164101877701         3442               0.395
#> 9   401114164 401114164101878001         3439               0.405
#> 10  401114164 401114164101878601         3433               0.437
#> 11  401114164 401114164101879101         3428               0.416
#> 12  401114164 401114164101884301         3416               0.454
#> 13  401114164 401114164101885401         3405               0.455
#> 14  401114164 401114164101888701         3372               0.488
#> 15  401114164 401114164101896601         3333               0.485
#> 16  401114164 401114164101899701         3302               0.528
#> 17  401114164 401114164101899702         3302               0.526
#> 18  401114164 401114164101907701         3262               0.566
#> 19  401114164 401114164101907703         3262               0.554
#> 20  401114164 401114164101908001         3259               0.576
#> 21  401114164 401114164101908901         3250               0.581
#> 22  401114164 401114164101916501         3214               0.602
#> 23  401114164 401114164101919301         3186               0.606
#> 24  401114164 401114164101924201         3177               0.611
#> 25  401114164 401114164101926701         3152               0.576
#> 26  401114164 401114164101928901         3130               0.576
#> 27  401114164 401114164101937201         3087               0.568
#> 28  401114164 401114164101946901         3030               0.645
#> 29  401114164 401114164101954501         2994               0.748
#> 30  401114164 401114164101954503         2994               0.747
#> 31  401114164 401114164101957801         2961               0.756
#> 32  401114164 401114164101958601         2953               0.766
#> 33  401114164 401114164101965001         2929               0.783
#> 34  401114164 401114164101967301         2906               0.779
#> 35  401114164 401114164101967901         2900               0.786
#> 36  401114164 401114164101975701         2862               0.791
#> 37  401114164 401114164101985701         2802               0.787
#> 38  401114164 401114164101987601         2783               0.734
#> 39  401114164 401114164101988901         2770               0.742
#> 40  401114164 401114164101994201         2757               0.735
#> 41  401114164 401114164101996901         2730               0.710
#> 42  401114164 401114164101998601         2713               0.730
#> 43  401114164 401114164101999901         2700               0.679
#> 44  401114164 401114164101999902         2700               0.679
#> 45  401114164 401114164102855601         2683               0.711
#> 46  401114164 401114164102857601         2663               0.628
#> 47  401114164 401114164102857603         2663               0.629
#> 48  401114164 401114164102858501         2654               0.628
#> 49  401114164 401114164102866001         2619               0.617
#> 50  401114164 401114164102866901         2610               0.646
#> 51  401114164 401114164102869901         2580               0.643
#> 52  401114164 401114164102876901         2550               0.672
#> 53  401114164 401114164102884701         2512               0.652
#> 54  401114164 401114164102885901         2500               0.691
#> 55  401114164 401114164102889001         2469               0.712
#> 56  401114164 401114164102897401         2425               0.710
#> 57  401114164 401114164102904901         2390               0.800
#> 58  401114164 401114164102904903         2390               0.806
#> 59  401114164 401114164102906601         2373               0.809
#> 60  401114164 401114164102908501         2354               0.795
#> 61  401114164 401114164102914101         2338               0.777
#> 62  401114164 401114164102915401         2325               0.759
#> 63  401114164 401114164102918401         2295               0.683
#> 64  401114164 401114164102919901         2280               0.652
#> 65  401114164 401114164102919903         2280               0.685
#> 66  401114164 401114164102926901         2250               0.677
#> 67  401114164 401114164102929901         2220               0.704
#> 68  401114164 401114164102939001         2169               0.680
#> 69  401114164 401114164102945401         2145               0.734
#> 70  401114164 401114164102948901         2110               0.715
#> 71  401114164 401114164102957601         2063               0.727
#> 72  401114164 401114164102965401         2025               0.768
#> 73  401114164 401114164102968901         1990               0.716
#> 74  401114164 401114164102974201         1977               0.712
#> 75  401114164 401114164102975101         1968               0.663
#> 76  401114164 401114164102975401         1965               0.647
#> 77  401114164 401114164102979001         1929               0.631
#> 78  401114164 401114164102984601         1913               0.574
#> 79  401114164 401114164102984602         1913               0.575
#> 80  401114164 401114164102985201         1907               0.442
#> 81  401114164 401114164102985203         1907               0.439
#> 82  401114164 401114164102986101         1898               0.460
#> 83  401114164 401114164102987901         1880               0.480
#> 84  401114164 401114164102994301         1856               0.478
#> 85  401114164 401114164102995001         1849               0.503
#> 86  401114164 401114164102997101         1828               0.665
#> 87  401114164 401114164102997103         1828               0.687
#> 88  401114164 401114164102998301         1816               0.688
#> 89  401114164 401114164102998901         1810               0.682
#> 90  401114164 401114164102999601         1803               0.676
#> 91  401114164 401114164102999901         1800               0.712
#> 92  401114164 401114164102999902         1800               0.685
#> 93  401114164 401114164103849901         1800               0.668
#> 94  401114164 401114164103855401         1785               0.692
#> 95  401114164 401114164103858001         1759               0.706
#> 96  401114164 401114164103858601         1753               0.638
#> 97  401114164 401114164103864901         1730               0.661
#> 98  401114164 401114164103867901         1700               0.684
#> 99  401114164 401114164103874901         1670               0.710
#> 100 401114164 401114164103875001         1669               0.716
#> 101 401114164 401114164103875501         1664               0.707
#> 102 401114164 401114164103876901         1650               0.768
#> 103 401114164 401114164103879401         1625               0.795
#> 104 401114164 401114164103884001         1619               0.833
#> 105 401114164 401114164103886101         1598               0.834
#> 106 401114164 401114164103889701         1562               0.866
#> 107 401114164 401114164103889703         1562               0.874
#> 108 401114164 401114164103894101         1558               0.887
#> 109 401114164 401114164103894801         1551               0.895
#> 110 401114164 401114164103897501         1524               0.909
#> 111 401114164 401114164103899301         1506               0.918
#> 112 401114164 401114164103899901         1500               0.905
#> 113 401114164 401114164103904601         1493               0.867
#> 114 401114164 401114164103905301         1486               0.883
#> 115 401114164 401114164103905901         1480               0.903
#> 116 401114164 401114164103908201         1457               0.890
#> 117 401114164 401114164103914401         1435               0.909
#> 118 401114164 401114164103915301         1426               0.925
#> 119 401114164 401114164103916201         1417               0.933
#> 120 401114164 401114164103919501         1384               0.921
#> 121 401114164 401114164103925801         1361               0.934
#> 122 401114164 401114164103934001         1319               0.890
#> 123 401114164 401114164103936901         1290               0.891
#> 124 401114164 401114164103939901         1260               0.880
#> 125 401114164 401114164103945401         1245               0.898
#> 126 401114164 401114164103946901         1230               0.906
#> 127 401114164 401114164103948401         1215               0.870
#> 128 401114164 401114164103954401         1195               0.876
#> 129 401114164 401114164103954901         1190               0.843
#> 130 401114164 401114164103955901         1180               0.856
#> 131 401114164 401114164103957601         1163               0.769
#> 132 401114164 401114164103957603         1163               0.771
#> 133 401114164 401114164103958401         1155               0.801
#> 134 401114164 401114164103965701         1122               0.777
#> 135 401114164 401114164103966201         1117               0.829
#> 136 401114164 401114164103966801         1111               0.798
#> 137 401114164 401114164103966901         1110               0.778
#> 138 401114164 401114164103967501         1104               0.766
#> 139 401114164 401114164103969201         1087               0.752
#> 140 401114164 401114164103976601         1053               0.694
#> 141 401114164 401114164103977401         1045               0.671
#> 142 401114164 401114164103979801         1021               0.720
#> 143 401114164 401114164103985401         1005               0.641
#> 144 401114164 401114164103988401          975               0.646
#> 145 401114164 401114164103994801          951               0.503
#> 146 401114164 401114164103994804          951               0.516
#> 147 401114164 401114164103999901          900               0.536
#> 148 401114164 401114164103999902          900               0.536
#> 149 401114164 401114164104855401          885               0.545
#> 150 401114164 401114164104858301          856               0.506
#> 151 401114164 401114164104858901          850               0.596
#> 152 401114164 401114164104867001          809               0.585
#> 153 401114164 401114164104875101          768               0.640
#> 154 401114164 401114164104879001          729               0.600
#> 155 401114164 401114164104884301          716               0.683
#> 156 401114164 401114164104888201          677               0.564
#> 157 401114164 401114164104895501          644               0.561
#> 158 401114164 401114164104896001          639               0.595
#> 159 401114164 401114164104896002          639               0.601
#> 160 401114164 401114164104898401          615               0.625
#> 161 401114164 401114164104904001          599               0.535
#> 162 401114164 401114164104906201          577               0.474
#> 163 401114164 401114164104908401          555               0.437
#> 164 401114164 401114164104915301          526               0.422
#> 165 401114164 401114164104918101          498               0.403
#> 166 401114164 401114164104926701          452               0.775
#> 167 401114164 401114164104927301          446               0.819
#> 168 401114164 401114164104929901          420               0.823
#> 169 401114164 401114164104939501          364               0.781
#> 170 401114164 401114164104946901          330               0.757
#> 171 401114164 401114164104949701          302               0.968
#> 172 401114164 401114164104949703          302               0.975
#> 173 401114164 401114164104954901          295               0.906
#> 174 401114164 401114164104957801          286               0.934
#> 175 401114164 401114164104959501          280               0.943
#> 176 401114164 401114164104964101          274               0.951
#> 177 401114164 401114164104964102          267               0.956
#> 178 401114164 401114164104964103          260               0.962
#> 179 401114164 401114164104964104          253               0.988
#> 180 401114164 401114164104964105          247               0.998
#> 181 401114164 401114164104964601          242               0.919
#> 182 401114164 401114164104964602          233               0.950
#> 183 401114164 401114164104965501          224               0.933
#> 184 401114164 401114164104969601          183               0.966
#> 185 401114164 401114164104975201          167               0.984
#> 186 401114164 401114164104976202          157               0.999
#> 187 401114164 401114164104976501          154               0.977
#> 188 401114164 401114164104976701          152               0.982
#> 189 401114164 401114164104978901          130               0.988
#> 190 401114164 401114164104985201          107               0.992
#> 191 401114164 401114164104989401           65               0.987
#> 192 401114164 401114164104994801           51               0.995
#> 193 401114164 401114164104995401           45               0.998
#> 194 401114164 401114164104996101           38               0.994
#> 195 401114164 401114164104996901           30               0.994
#> 196 401114164 401114164104997901           20               0.986
#> 197 401114164 401114164104998701           12               0.980
#> 198 401114164 401114164104998901           10               0.982
#> 199 401114164 401114164104999901            0               1.000
#> 200 401114164 401114164104999902            0               1.000
#>     away_win_percentage tie_percentage
#> 1                 0.595              0
#> 2                 0.591              0
#> 3                 0.586              0
#> 4                 0.628              0
#> 5                 0.630              0
#> 6                 0.582              0
#> 7                 0.625              0
#> 8                 0.605              0
#> 9                 0.595              0
#> 10                0.563              0
#> 11                0.584              0
#> 12                0.546              0
#> 13                0.545              0
#> 14                0.512              0
#> 15                0.515              0
#> 16                0.472              0
#> 17                0.474              0
#> 18                0.434              0
#> 19                0.446              0
#> 20                0.424              0
#> 21                0.419              0
#> 22                0.398              0
#> 23                0.394              0
#> 24                0.389              0
#> 25                0.424              0
#> 26                0.424              0
#> 27                0.432              0
#> 28                0.355              0
#> 29                0.252              0
#> 30                0.253              0
#> 31                0.244              0
#> 32                0.234              0
#> 33                0.217              0
#> 34                0.221              0
#> 35                0.214              0
#> 36                0.209              0
#> 37                0.213              0
#> 38                0.266              0
#> 39                0.258              0
#> 40                0.265              0
#> 41                0.290              0
#> 42                0.270              0
#> 43                0.321              0
#> 44                0.321              0
#> 45                0.289              0
#> 46                0.372              0
#> 47                0.371              0
#> 48                0.372              0
#> 49                0.383              0
#> 50                0.354              0
#> 51                0.357              0
#> 52                0.328              0
#> 53                0.348              0
#> 54                0.309              0
#> 55                0.288              0
#> 56                0.290              0
#> 57                0.200              0
#> 58                0.194              0
#> 59                0.191              0
#> 60                0.205              0
#> 61                0.223              0
#> 62                0.241              0
#> 63                0.317              0
#> 64                0.348              0
#> 65                0.315              0
#> 66                0.323              0
#> 67                0.296              0
#> 68                0.320              0
#> 69                0.266              0
#> 70                0.285              0
#> 71                0.273              0
#> 72                0.232              0
#> 73                0.284              0
#> 74                0.288              0
#> 75                0.337              0
#> 76                0.353              0
#> 77                0.369              0
#> 78                0.426              0
#> 79                0.425              0
#> 80                0.558              0
#> 81                0.561              0
#> 82                0.540              0
#> 83                0.520              0
#> 84                0.522              0
#> 85                0.497              0
#> 86                0.335              0
#> 87                0.313              0
#> 88                0.312              0
#> 89                0.318              0
#> 90                0.324              0
#> 91                0.288              0
#> 92                0.315              0
#> 93                0.332              0
#> 94                0.308              0
#> 95                0.294              0
#> 96                0.362              0
#> 97                0.339              0
#> 98                0.316              0
#> 99                0.290              0
#> 100               0.284              0
#> 101               0.293              0
#> 102               0.232              0
#> 103               0.205              0
#> 104               0.167              0
#> 105               0.166              0
#> 106               0.134              0
#> 107               0.126              0
#> 108               0.113              0
#> 109               0.105              0
#> 110               0.091              0
#> 111               0.082              0
#> 112               0.095              0
#> 113               0.133              0
#> 114               0.117              0
#> 115               0.097              0
#> 116               0.110              0
#> 117               0.091              0
#> 118               0.075              0
#> 119               0.067              0
#> 120               0.079              0
#> 121               0.066              0
#> 122               0.110              0
#> 123               0.109              0
#> 124               0.120              0
#> 125               0.102              0
#> 126               0.094              0
#> 127               0.130              0
#> 128               0.124              0
#> 129               0.157              0
#> 130               0.144              0
#> 131               0.231              0
#> 132               0.229              0
#> 133               0.199              0
#> 134               0.223              0
#> 135               0.171              0
#> 136               0.202              0
#> 137               0.222              0
#> 138               0.234              0
#> 139               0.248              0
#> 140               0.306              0
#> 141               0.329              0
#> 142               0.280              0
#> 143               0.359              0
#> 144               0.354              0
#> 145               0.497              0
#> 146               0.484              0
#> 147               0.464              0
#> 148               0.464              0
#> 149               0.455              0
#> 150               0.494              0
#> 151               0.404              0
#> 152               0.415              0
#> 153               0.360              0
#> 154               0.400              0
#> 155               0.317              0
#> 156               0.436              0
#> 157               0.439              0
#> 158               0.405              0
#> 159               0.399              0
#> 160               0.375              0
#> 161               0.465              0
#> 162               0.526              0
#> 163               0.563              0
#> 164               0.578              0
#> 165               0.597              0
#> 166               0.225              0
#> 167               0.181              0
#> 168               0.177              0
#> 169               0.219              0
#> 170               0.243              0
#> 171               0.032              0
#> 172               0.025              0
#> 173               0.094              0
#> 174               0.066              0
#> 175               0.057              0
#> 176               0.049              0
#> 177               0.044              0
#> 178               0.038              0
#> 179               0.012              0
#> 180               0.002              0
#> 181               0.081              0
#> 182               0.050              0
#> 183               0.067              0
#> 184               0.034              0
#> 185               0.016              0
#> 186               0.001              0
#> 187               0.023              0
#> 188               0.018              0
#> 189               0.012              0
#> 190               0.008              0
#> 191               0.013              0
#> 192               0.005              0
#> 193               0.002              0
#> 194               0.006              0
#> 195               0.006              0
#> 196               0.014              0
#> 197               0.020              0
#> 198               0.018              0
#> 199               0.000              0
#> 200               0.000              0

# }